--- resources/assets/assets/index-rlghdUvH.js
@@ -1,10 +1,0 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/vendor-utils-D7YtnGoz.js","assets/rolldown-runtime-BHe-jwch.js"])))=>i.map(i=>d[i]);
-import{r as e}from"./rolldown-runtime-BHe-jwch.js";import{_ as t,a as n,c as r,d as i,f as a,g as o,h as s,i as c,l,m as u,n as d,o as f,p,r as m,s as h,t as g,u as _,v}from"./vendor-react-CNktMmef.js";import{C as y,t as b}from"./vendor-CsjAK7B8.js";import{a as x,c as S,i as C,n as w,o as T,r as E,s as D,t as O}from"./vendor-d3-DThTr3c3.js";import{t as k}from"./vendor-utils-D7YtnGoz.js";(function(){let e=document.createElement(`link`).relList;if(e&&e.supports&&e.supports(`modulepreload`))return;for(let e of document.querySelectorAll(`link[rel="modulepreload"]`))n(e);new MutationObserver(e=>{for(let t of e)if(t.type===`childList`)for(let e of t.addedNodes)e.tagName===`LINK`&&e.rel===`modulepreload`&&n(e)}).observe(document,{childList:!0,subtree:!0});function t(e){let t={};return e.integrity&&(t.integrity=e.integrity),e.referrerPolicy&&(t.referrerPolicy=e.referrerPolicy),e.crossOrigin===`use-credentials`?t.credentials=`include`:e.crossOrigin===`anonymous`?t.credentials=`omit`:t.credentials=`same-origin`,t}function n(e){if(e.ep)return;e.ep=!0;let n=t(e);fetch(e.href,n)}})();var A=e(v(),1),j=t();function M(){let[e,t]=(0,A.useState)(null),[n,r]=(0,A.useState)(!0),[i,a]=(0,A.useState)(null);return(0,A.useEffect)(()=>{fetch(`/_laravel-brain/.graph-manifest.json`).then(e=>{if(!e.ok)throw Error(`HTTP ${e.status}`);return e.json()}).then(e=>{t(e),r(!1)}).catch(e=>{a(e.message),r(!1)})},[]),{manifest:e,loading:n,error:i}}function N(e){return[...e.nodes.map(e=>({data:{id:e.id,label:e.label,type:e.type,...e.data,metrics_cc:e.data?.metrics?.cyclomaticComplexity??0}})),...e.edges.map(e=>({data:{id:e.id,source:e.source,target:e.target,label:e.label,type:e.type}}))]}function P(){let[e,t]=(0,A.useState)({data:null,loading:!1,error:null}),n=(0,A.useRef)(new Map),r=(0,A.useRef)(null),i=(0,A.useCallback)(e=>{if(r.current===e)return;r.current=e;let i=n.current.get(e);if(i){t({data:i,loading:!1,error:null});return}t(e=>({...e,loading:!0,error:null})),fetch(`/_laravel-brain/`+e).then(e=>{if(!e.ok)throw Error(`HTTP ${e.status}`);return e.json()}).then(i=>{n.current.set(e,i),r.current===e&&t({data:i,loading:!1,error:null})}).catch(n=>{r.current===e&&t({data:null,loading:!1,error:n.message})})},[]);return{state:e,elements:(0,A.useMemo)(()=>e.data?N(e.data):[],[e.data]),load:i}}function F(e,t=300){let[n,r]=(0,A.useState)(t),[i,a]=(0,A.useState)(e);return e!==i&&(a(e),r(t)),(0,A.useEffect)(()=>{if(n>=e.length)return;let t=window,i=(t.requestIdleCallback?t.requestIdleCallback.bind(t):e=>setTimeout(()=>e({didTimeout:!1,timeRemaining:()=>0}),100))(()=>{r(t=>Math.min(t+200,e.length))});return()=>{t.cancelIdleCallback?t.cancelIdleCallback(i):clearTimeout(i)}},[n,e.length]),(0,A.useMemo)(()=>e.length<=t?e:e.slice(0,n),[e,n,t])}function ee(){let[e,t]=(0,A.useState)(()=>localStorage.getItem(`lb-theme`)??`dark`);return(0,A.useEffect)(()=>{document.documentElement.setAttribute(`data-theme`,e),localStorage.setItem(`lb-theme`,e)},[e]),{theme:e,toggle:()=>t(e=>e===`dark`?`light`:`dark`)}}var te=e(k(),1),I={transaction:`#d99a2b`,rollback:`#c2554a`,chain:`#5f8fa8`,batch:`#8a7fb5`},ne={transaction:`6 5`,rollback:`2 4`,chain:`10 4`,batch:`4 4`},L={macro:`#F59E0B`,macro_group:`#B45309`,route:`#4CAF50`,middleware:`#FF9800`,controller:`#2196F3`,livewire_component:`#FB7185`,action:`#03A9F4`,service:`#9C27B0`,action_class:`#84cc16`,validation_request:`#0d9488`,model:`#F44336`,event:`#FFD600`,listener:`#C9A227`,job:`#607D8B`,command:`#14b8a6`,channel:`#8b5cf6`,schedule:`#f97316`,view:`#ec4899`,mail:`#f472b6`,notification:`#db2777`,enum:`#0ea5e9`,interface:`#38bdf8`,trait:`#a78bfa`,abstract_class:`#94a3b8`,service_provider:`#ca8a04`,facade:`#00BCD4`,ai_agent:`#A3E635`,ai_tool:`#65A30D`,filament_panel:`#7C3AED`,filament_resource:`#A855F7`,filament_page:`#C084FC`,filament_page_method:`#E879F9`,filament_widget:`#06B6D4`,filament_relation_manager:`#0891B2`},re={macro:`#B45309`,macro_group:`#92400E`,route:`#2e7d32`,middleware:`#e65100`,controller:`#1565c0`,livewire_component:`#e11d48`,action:`#0277bd`,service:`#6a1b9a`,action_class:`#4d7c0f`,validation_request:`#0f766e`,model:`#c62828`,event:`#b45309`,listener:`#8a6d1f`,job:`#37474f`,command:`#0d7d6e`,channel:`#5b21b6`,schedule:`#c2410c`,view:`#be185d`,mail:`#c026d3`,notification:`#9d174d`,enum:`#0369a1`,interface:`#0c4a6e`,trait:`#5b21b6`,abstract_class:`#64748b`,service_provider:`#a16207`,facade:`#00838f`,ai_agent:`#4d7c0f`,ai_tool:`#3f6212`,filament_panel:`#5b21b6`,filament_resource:`#7e22ce`,filament_page:`#9333ea`,filament_page_method:`#a21caf`,filament_widget:`#0369a1`,filament_relation_manager:`#075985`},ie={route:`#0C1A0C`,middleware:`#1C1408`,controller:`#08141C`,livewire_component:`#1C0711`,action:`#07151D`,service:`#150C1C`,action_class:`#131C06`,validation_request:`#042f2e`,model:`#1C0C0C`,event:`#1C1A08`,listener:`#181405`,job:`#0D1113`,command:`#061514`,channel:`#110c1c`,schedule:`#1c1008`,view:`#1c0a14`,mail:`#1c0f18`,notification:`#1c0510`,enum:`#071318`,interface:`#081420`,trait:`#140822`,abstract_class:`#0f172a`,service_provider:`#422006`,facade:`#001F28`,ai_agent:`#131A08`,ai_tool:`#101705`,filament_panel:`#150C2A`,filament_resource:`#1A0C26`,filament_page:`#1E0F2E`,filament_page_method:`#240E30`,filament_widget:`#071A1E`,filament_relation_manager:`#06161A`},R={route:`#f0fdf4`,middleware:`#fff7ed`,controller:`#eff6ff`,livewire_component:`#fff1f2`,action:`#e0f7fa`,service:`#fdf4ff`,action_class:`#f7fee7`,validation_request:`#ccfbf1`,model:`#fff1f2`,event:`#fefce8`,listener:`#fdf6dd`,job:`#f1f5f9`,command:`#f0fdfa`,channel:`#f5f3ff`,schedule:`#fff7ed`,view:`#fdf2f8`,mail:`#fce7f3`,notification:`#fce7f3`,enum:`#f0f9ff`,interface:`#ecfeff`,trait:`#f5f3ff`,abstract_class:`#f1f5f9`,service_provider:`#fef9c3`,facade:`#e0f7fa`,ai_agent:`#f7fee7`,ai_tool:`#ecfccb`,filament_panel:`#f5f3ff`,filament_resource:`#faf5ff`,filament_page:`#fdf4ff`,filament_page_method:`#fef0ff`,filament_widget:`#ecfeff`,filament_relation_manager:`#e0f2fe`},ae=`#8B6FE8`,z={public:{bg:`#1c0808`,border:`#ef4444`,accent:`#f87171`,label:`Public`},guest:{bg:`#1c1408`,border:`#f59e0b`,accent:`#fbbf24`,label:`Guest`},authed:{bg:`#081c10`,border:`#10b981`,accent:`#34d399`,label:`Auth`},admin:{bg:`#110c1c`,border:`#8b5cf6`,accent:`#a78bfa`,label:`Admin`}},oe={public:{bg:`#fff1f2`,border:`#ef4444`,accent:`#dc2626`,label:`Public`},guest:{bg:`#fffbeb`,border:`#f59e0b`,accent:`#d97706`,label:`Guest`},authed:{bg:`#ecfdf5`,border:`#10b981`,accent:`#059669`,label:`Auth`},admin:{bg:`#f5f3ff`,border:`#8b5cf6`,accent:`#7c3aed`,label:`Admin`}},B={none:`#6b7280`,low:`#10b981`,medium:`#f59e0b`,high:`#f97316`,critical:`#ef4444`},se={none:`No Issues`,low:`Low`,medium:`Medium`,high:`High`,critical:`Critical`},ce={MISSING_FK_INDEX:{icon:`🔑`,name:`Unindexed foreign key`},MASS_ASSIGNMENT:{icon:`⚡`,name:`Mass Assignment`},UNVALIDATED_INPUT:{icon:`⚠️`,name:`Unvalidated Input`},MISSING_THROTTLE:{icon:`🔓`,name:`Missing Throttle`},PUBLIC_WRITE:{icon:`🌐`,name:`Public Write`},XSS_DIRECT_OUTPUT:{icon:`💉`,name:`XSS — Direct Output`},XSS_HTML_DECODE:{icon:`🔓`,name:`XSS — HTML Decode`},XSS_BLADE_UNESCAPED:{icon:`📄`,name:`XSS — Blade {!! !!}`},SQL_INJECTION:{icon:`🛢️`,name:`SQL Injection`},OPEN_REDIRECT:{icon:`↪️`,name:`Open Redirect`},SSRF:{icon:`🌐`,name:`SSRF`},DEBUG_CODE:{icon:`🐞`,name:`Debug Code Leak`},ENV_LEAK:{icon:`🔑`,name:`Env Leak`},CSRF_BYPASS:{icon:`🛡️`,name:`CSRF Bypass`},INSECURE_COOKIE:{icon:`🍪`,name:`Insecure Cookie`},UNSAFE_STORAGE_PATH:{icon:`📁`,name:`Unsafe Storage Path`},FILE_UPLOAD_VALIDATION:{icon:`📎`,name:`File Upload Validation`},UNSAFE_AUTH:{icon:`🚪`,name:`Unsafe Auth`},UNSAFE_CRYPT:{icon:`🔐`,name:`Unsafe Crypt`},ARTISAN_CALL:{icon:`⚙️`,name:`Tainted Artisan Call`},PROCESS_SHELL:{icon:`💻`,name:`Shell Injection`},CONFIG_INJECTION:{icon:`🧩`,name:`Config Injection`},TAINTED_VIEW_NAME:{icon:`🖼️`,name:`Tainted View Name`},SESSION_FIXATION:{icon:`🎫`,name:`Session Fixation`},MAIL_TAINTED_HEADER:{icon:`✉️`,name:`Mail Header Injection`}},le=[{label:`Low`,min:1,max:5,fill:`#0d2e1a`,border:`#4ade80`},{label:`Moderate`,min:6,max:10,fill:`#2e2200`,border:`#facc15`},{label:`High`,min:11,max:15,fill:`#2e1200`,border:`#fb923c`},{label:`Critical`,min:16,max:1/0,fill:`#2e0a0a`,border:`#f87171`}],ue=[{label:`Low`,min:1,max:5,fill:`#f0fdf4`,border:`#16a34a`},{label:`Moderate`,min:6,max:10,fill:`#fefce8`,border:`#ca8a04`},{label:`High`,min:11,max:15,fill:`#fff7ed`,border:`#ea580c`},{label:`Critical`,min:16,max:1/0,fill:`#fff1f2`,border:`#dc2626`}],de=[`chain`],fe={transaction:`transaction`,rollback:`rollback`,chain:`chain`,batch:`batch`},pe={transaction:`transactions`,rollback:`rollbacks`,chain:`chains`,batch:`batches`},V=[`transaction`,`rollback`,`chain`,`batch`];function H(e){let t=e.width/2,n=e.height/2;return[[e.x-t,e.y-n],[e.x+t,e.y-n],[e.x+t,e.y+n],[e.x-t,e.y+n]]}function U(e){if(e.length<3)return e;let t=[...e].sort((e,t)=>e[0]-t[0]||e[1]-t[1]),n=(e,t,n)=>(t[0]-e[0])*(n[1]-e[1])-(t[1]-e[1])*(n[0]-e[0]),r=e=>{let t=[];for(let r of e){for(;t.length>=2&&n(t[t.length-2],t[t.length-1],r)<=0;)t.pop();t.push(r)}return t.pop(),t};return[...r(t),...r([...t].reverse())]}function W(e,t){if(e.length===0)return e;let n=e.reduce((e,t)=>e+t[0],0)/e.length,r=e.reduce((e,t)=>e+t[1],0)/e.length;return e.map(([e,i])=>{let a=e-n,o=i-r,s=Math.hypot(a,o)||1;return[e+a/s*t,i+o/s*t]})}function G(e,t,n){let r=!1;for(let i=0,a=e.length-1;i<e.length;a=i++){let[o,s]=e[i],[c,l]=e[a];s>n!=l>n&&t<(c-o)*(n-s)/(l-s)+o&&(r=!r)}return r}var K=new Set([`transaction`,`rollback`,`chain`,`batch`]);function me(e,t,n=4){let r=t.x-e.x,i=t.y-e.y;if(r===0&&i===0)return null;let a=e=>{let t=r===0?1/0:(e.width/2+n)/Math.abs(r),a=i===0?1/0:(e.height/2+n)/Math.abs(i);return Math.min(t,a)},o=a(e),s=1-a(t);return o>=s?null:{x1:e.x+r*o,y1:e.y+i*o,x2:e.x+r*s,y2:e.y+i*s}}function he(e){let t=e.data?.regions;if(!Array.isArray(t))return[];let n=[];for(let e of t){let t=e?.id,r=e?.kind;typeof t!=`string`||t===``||typeof r!=`string`||!K.has(r)||n.push({id:t,kind:r,position:typeof e.position==`number`?e.position:null})}return n}function ge(e,t=22){let n=new Map;for(let t of e)for(let e of he(t)){let r=n.get(e.id)??{kind:e.kind,members:[]};r.members.push({node:t,position:e.position}),n.set(e.id,r)}let r=[],i=new Map,a=new Map;for(let e of[...n.keys()].sort()){let t=n.get(e).kind,r=(a.get(t)??0)+1;a.set(t,r),i.set(e,r)}for(let[a,o]of n){let n=de.includes(o.kind),s=(n?[...o.members].sort((e,t)=>(e.position??0)-(t.position??0)):o.members).map(e=>e.node),c=W(U(s.flatMap(H)),t);if(c.length<3)continue;let l=new Set(s.map(e=>e.id)),u=!e.some(e=>!l.has(e.id)&&H(e).some(([e,t])=>G(c,e,t)));r.push({id:a,kind:o.kind,index:i.get(a)??1,points:c,members:s,ordered:n,pure:u})}return r}var q=e(y(),1);function J(e,t){let n=e.indexOf(`@`),r=e.indexOf(`::`);return n===-1?r===-1?{className:e,method:t??``}:{className:e.slice(0,r),method:e.slice(r+2)}:{className:e.slice(0,n),method:t??e.slice(n+1)}}function Y(e,t=!1){let{className:n,method:r}=J(String(e.label??e.id),e.method),i=t||n.length>r.length?n:r,a=Math.max(t?120:185,Math.min(270,i.length*7.6+44)),o=t?40:90;return{id:e.id,x:0,y:0,width:a,height:o,lines:[n,r].filter(Boolean),data:e}}function _e(e){if(!e.length)return;let t=0,n=0;for(let r of e)t+=r.x,n+=r.y;let r=t/e.length,i=n/e.length;for(let t of e)t.x-=r,t.y-=i}function ve(e,t,n){let r=new q.default.graphlib.Graph({compound:!0});r.setGraph({rankdir:n,nodesep:n===`TB`?70:50,ranksep:n===`TB`?100:120,marginx:60,marginy:60}),r.setDefaultEdgeLabel(()=>({}));for(let t of e)r.setNode(t.id,{width:t.width,height:t.height});for(let[t,n]of be(e)){r.setNode(t,{});for(let e of n)r.setParent(e.id,t)}for(let e of t)r.hasNode(e.source)&&r.hasNode(e.target)&&r.setEdge(e.source,e.target);q.default.layout(r);for(let t of e){let e=r.node(t.id);e&&(t.x=e.x,t.y=e.y)}}function ye(e){let t=he(e);return t.length===0?null:(t.find(e=>de.includes(e.kind))??t[0]).id}function be(e){let t=new Map;for(let n of e){let e=ye(n);e!==null&&t.set(e,[...t.get(e)??[],n])}for(let[e,n]of t)n.length<2&&t.delete(e);return new Map([...t].map(([e,t])=>[`cluster::${e}`,t]))}function xe(e,t,n,r=60,i=110){let a=new Set(e.map(e=>e.id)),o=new Map,s=new Map;for(let t of e)o.set(t.id,[]),s.set(t.id,0);for(let e of t)!a.has(e.source)||!a.has(e.target)||(o.get(e.source).push(e.target),s.set(e.target,(s.get(e.target)??0)+1));let c=e.filter(e=>s.get(e.id)===0).map(e=>e.id),l=new Map,u=[...c];for(let e of c)l.set(e,0);let d=0,f=()=>{for(;d<u.length;){let e=u[d++],t=l.get(e);for(let n of o.get(e)??[])l.has(n)||(l.set(n,t+1),u.push(n))}};for(f();;){let t=null;for(let n of e)l.has(n.id)||(t===null||s.get(n.id)<s.get(t.id))&&(t=n);if(t===null)break;l.set(t.id,0),u.push(t.id),f()}let p=new Map;for(let t of e){let e=l.get(t.id);p.has(e)||p.set(e,[]),p.get(e).push(t.id)}let m=new Map;for(let t of e){let e=ye(t);e!==null&&m.set(t.id,e)}if(m.size>0)for(let[,e]of p){let t=new Map,n=0;for(let r of e){let e=m.get(r)??`\u0000${r}`;t.has(e)||t.set(e,n++)}e.sort((e,n)=>t.get(m.get(e)??`\u0000${e}`)-t.get(m.get(n)??`\u0000${n}`))}for(let e of p.values())e.sort();let h=new Map(e.map(e=>[e.id,e])),g=0;for(let e of[...p.keys()].sort((e,t)=>e-t)){let t=p.get(e).map(e=>h.get(e)),a=Se(t.length);if(n===`TB`){let e=Ce(t,a),n=g;for(let t of e){let e=t.reduce((e,t)=>e+t.width,0)+r*(t.length-1),i=Te(t,e=>e.height),a=-e/2;for(let e of t)e.x=a+e.width/2,e.y=n+i/2,a+=e.width+r;n+=i+r}g=n-r+i}else{let e=Ce(t,a),n=g;for(let t of e){let e=t.reduce((e,t)=>e+t.height,0)+r*(t.length-1),i=Te(t,e=>e.width),a=-e/2;for(let e of t)e.x=n+i/2,e.y=a+e.height/2,a+=e.height+r;n+=i+r}g=n-r+i}}}function Se(e,t=12){return e<=t?e:Math.ceil(Math.sqrt(e)*1.4)}function Ce(e,t){if(t>=e.length)return[e];let n=[];for(let r=0;r<e.length;r+=t)n.push(e.slice(r,r+t));return n}function we(e,t){let n=e.map(e=>Object.assign({},e)),r=new Map(n.map(e=>[e.id,e])),i=t.filter(e=>r.has(e.source)&&r.has(e.target)).map(e=>({source:e.source,target:e.target})),a=C(n).force(`link`,x(i).id(e=>e.id).distance(90)).force(`charge`,E().strength(-420)).force(`center`,D(0,0)).force(`collide`,T().radius(e=>Math.hypot(e.width,e.height)/2+14));a.stop();for(let e=0;e<450&&a.alpha()>.02;e++)a.tick();for(let t of e){let e=r.get(t.id);e&&(t.x=e.x??0,t.y=e.y??0)}}function Te(e,t){return e.reduce((e,n)=>Math.max(e,t(n)),-1/0)}function Ee(e,t=40){let n=e.length;if(!n)return;let r=Te(e,e=>Math.max(e.width,e.height))+t,i=Math.max(r,n*r/(2*Math.PI));e.forEach((e,t)=>{let r=t/n*Math.PI*2-Math.PI/2;e.x=i*Math.cos(r),e.y=i*Math.sin(r)})}function De(e,t=60,n=60){if(!e.length)return;let r=Te(e,e=>e.width)+t,i=Te(e,e=>e.height)+n,a=Math.ceil(Math.sqrt(e.length));e.forEach((e,t)=>{e.x=t%a*r,e.y=Math.floor(t/a)*i})}function Oe(e,t,n){return e===`dagre`&&t>n?`breadthfirst`:e===`dagre`?`dagre`:e===`cose-bilkent`?`force`:e===`breadthfirst`?`breadthfirst`:e===`circle`?`circle`:e===`grid`?`grid`:`dagre`}function ke(e,t=!1){let n=[],r=[];for(let i of e){let e=i.data;e.source!=null&&e.target!=null?r.push({id:e.id,source:String(e.source),target:String(e.target),data:e}):n.push(Y(e,t))}return{nodes:n,edges:r}}var X=o();function Ae(e){return Math.max(0,Math.min(255,Math.round(e))).toString(16).padStart(2,`0`)}function je(e,t,n){return{x:n.applyX(e),y:n.applyY(t)}}function Me(e,t){if(t.length===0)return{x:0,y:0};if(t.length===1||e<=0)return t[0];if(e>=1)return t[t.length-1];let n=0,r=[];for(let e=0;e<t.length-1;e++){let i=t[e+1].x-t[e].x,a=t[e+1].y-t[e].y,o=Math.sqrt(i*i+a*a);r.push(o),n+=o}if(n===0)return t[0];let i=e*n,a=0;for(let e=0;e<r.length;e++){let n=r[e];if(n!==0){if(a+n>=i||e===r.length-1){let r=(i-a)/n,o=t[e],s=t[e+1];return{x:o.x+(s.x-o.x)*r,y:o.y+(s.y-o.y)*r}}a+=n}}return t[t.length-1]}function Ne(e,t){let n=t.x-e.x,r=t.y-e.y,i=Math.abs(n)-(e.width+t.width)/2;return Math.abs(r)-(e.height+t.height)/2>=i?r>=0?{ex:e.x,ey:e.y+e.height/2,tx:t.x,ty:t.y-t.height/2,vertical:!0}:{ex:e.x,ey:e.y-e.height/2,tx:t.x,ty:t.y+t.height/2,vertical:!0}:n>=0?{ex:e.x+e.width/2,ey:e.y,tx:t.x-t.width/2,ty:t.y,vertical:!1}:{ex:e.x-e.width/2,ey:e.y,tx:t.x+t.width/2,ty:t.y,vertical:!1}}function Pe(e,t){let{ex:n,ey:r,tx:i,ty:a,vertical:o}=Ne(e,t);if(o){if(Math.abs(n-i)<3||Math.abs(a-r)<=8)return[{x:n,y:r},{x:i,y:a}];let e=(r+a)/2;return[{x:n,y:r},{x:n,y:e},{x:i,y:e},{x:i,y:a}]}else{if(Math.abs(r-a)<3||Math.abs(i-n)<=8)return[{x:n,y:r},{x:i,y:a}];let e=(n+i)/2;return[{x:n,y:r},{x:e,y:r},{x:e,y:a},{x:i,y:a}]}}var Fe=7;function Ie(...e){return Math.max(0,Math.min(Fe,...e.map(e=>e-1)))}function Le(e,t){let{ex:n,ey:r,tx:i,ty:a,vertical:o}=Ne(e,t);if(o){if(Math.abs(n-i)<3||Math.abs(a-r)<=8)return{d:`M${n},${r} L${i},${a}`,lx:n+6,ly:(r+a)/2,exitX:n,exitY:r,entryX:i,entryY:a};let e=(r+a)/2,t=a>r?1:-1,o=Ie(Math.abs(e-r),Math.abs(a-e),Math.abs(i-n)),s=i>n?o:-o;return{d:o>0?`M${n},${r} V${e-o*t} Q${n},${e} ${n+s},${e} H${i-s} Q${i},${e} ${i},${e+o*t} V${a}`:`M${n},${r} V${e} H${i} V${a}`,lx:(n+i)/2,ly:e-14*t,exitX:n,exitY:r,entryX:i,entryY:a}}else{if(Math.abs(r-a)<3||Math.abs(i-n)<=8)return{d:`M${n},${r} L${i},${a}`,lx:(n+i)/2,ly:r-10,exitX:n,exitY:r,entryX:i,entryY:a};let e=(n+i)/2,t=i>n?1:-1,o=Ie(Math.abs(e-n),Math.abs(i-e),Math.abs(a-r)),s=a>r?o:-o;return{d:o>0?`M${n},${r} H${e-o*t} Q${e},${r} ${e},${r+s} V${a-s} Q${e},${a} ${e+o*t},${a} H${i}`:`M${n},${r} H${e} V${a} H${i}`,lx:e+6*t,ly:(r+a)/2,exitX:n,exitY:r,entryX:i,entryY:a}}}function Re(e,t){let n=String(e.label??``);return n?{text:n,fill:t?`rgba(255,255,255,0.4)`:`rgba(0,0,0,0.5)`,bg:t?`#111218`:`#fff`}:null}function ze(e,t,n,r,i,a){let o=String(e.data.type??``),s=t?L[o]??`#c9d1d9`:re[o]??`#333`,c=t?ie[o]??`#0d1117`:R[o]??`#ffffff`,l=Number(e.data.metrics_cc??0)||0;if(n){let n=t?le:ue,r=n.find(e=>l>=e.min&&l<=e.max)??n[0],a=i?`#a855f7`:e.data.hasN1?`#F44336`:r.border;return{bg:r.fill,border:a,borderW:1.5,accent:r.border}}if(a&&o===`route`){let n=e.data.security;if(n){let e=t?z:oe,a=e[n.exposure]??e.public,o=B[n.riskLevel]??B.none,c=r?s:i?`#a855f7`:n.riskLevel===`none`?a.border:o;return{bg:a.bg,border:c,borderW:r||n.riskLevel!==`none`?2:1.5,accent:a.accent}}}let u=t?`rgba(255,255,255,0.1)`:`rgba(0,0,0,0.12)`,d=1;return e.data.hasN1&&(u=`#F44336`,d=2),r&&(u=s,d=2),i&&(u=`#a855f7`,d=2),{bg:c,border:u,borderW:d,accent:s}}function Be(e){if(e.ctrlKey)return!1;if(e.deltaX!==0)return!0;if(e.deltaMode!==0)return!1;let t=e.wheelDeltaY;return typeof t==`number`&&t!==0?Math.abs(t+3*e.deltaY)<=2?!0:!(Math.abs(t)%120==0&&Math.abs(e.deltaY)>=100):!(Number.isInteger(e.deltaY)&&Math.abs(e.deltaY)>=100)}function Ve({elements:e,layout:t,rankDir:n,searchQuery:r,visibleTypes:i,theme:a,onNodeSelect:o,graphRef:s,stressTestNodeId:c,stressRunKey:l,complexityOverlay:u,securityOverlay:d=!1,compact:f=!1,onLayoutChange:p,onRankDirChange:m,onToggleComplexityOverlay:h,onToggleSecurityOverlay:g,onToggleCompact:_}){let v=a===`dark`,y=v?`rgba(255,255,255,0.32)`:`rgba(0,0,0,0.38)`,b=v?`rgba(255,255,255,0.55)`:`rgba(0,0,0,0.55)`,{nodes:x,edges:C}=(0,A.useMemo)(()=>ke(e,f),[e,f]),T=(0,A.useMemo)(()=>x.filter(e=>i.has(String(e.data.type))).length,[x,i]),[E,D]=(0,A.useState)(0),k=(0,A.useRef)(null),j=(0,A.useRef)(!0);(0,A.useEffect)(()=>{if(j.current){j.current=!1;return}return k.current&&window.clearTimeout(k.current),k.current=window.setTimeout(()=>{D(e=>e+1)},200),()=>{k.current&&window.clearTimeout(k.current)}},[i,t,n,f]);let{nodes:M,edges:N}=(0,A.useMemo)(()=>{let e=x.map(e=>({...e,lines:[...e.lines]})),r=C.map(e=>({...e})),i=Oe(t,T,80);return i===`dagre`?ve(e,r,n):i===`breadthfirst`?xe(e,r,n):i===`force`?we(e,r):i===`circle`?Ee(e):De(e),_e(e),{nodes:e,edges:r}},[x,C,t,n,E,T]),P=(0,A.useMemo)(()=>new Map(M.map(e=>[e.id,e])),[M]),[F,ee]=(0,A.useState)(new Map),re=(0,A.useRef)(null),ie=(0,A.useRef)(!1),[R,oe]=(0,A.useState)(new Set),[se,ce]=(0,A.useState)(M);se!==M&&(ce(M),ee(new Map),oe(new Set));let ue=(0,A.useMemo)(()=>F.size===0?M:M.map(e=>{let t=F.get(e.id);return t?{...e,x:t.x,y:t.y}:e}),[M,F]),de=(0,A.useMemo)(()=>ge(ue),[ue]),H=(0,A.useCallback)(e=>i.has(e===`rollback`?`transaction`:e),[i]),U=(0,A.useMemo)(()=>de.filter(e=>H(e.kind)),[de,H]),W=(0,A.useMemo)(()=>{let e=new Map,t=[],n=(n,r,i)=>{let a=0;for(;t.some(e=>Math.abs(e.x-r)<140&&Math.abs(e.y-(i-a*11))<10);)a++;t.push({x:r,y:i-a*11}),e.set(n,a)};for(let e of U){if(e.pure){n(e.id,Math.min(...e.points.map(([e])=>e)),Math.min(...e.points.map(([,e])=>e)));continue}for(let t of e.members)n(`${e.id}|${t.id}`,t.x-t.width/2,t.y-t.height/2)}return e},[U]),G=(0,A.useMemo)(()=>new Map(ue.map(e=>[e.id,e])),[ue]),K=(0,A.useRef)(G);(0,A.useEffect)(()=>{K.current=G},[G]);let he=(0,A.useCallback)(e=>i.has(String(e)),[i]),q=(0,A.useCallback)(e=>he(P.get(e.source)?.data.type)&&he(P.get(e.target)?.data.type),[P,he]),Y=(0,A.useMemo)(()=>{let e=new Map;for(let t of M)e.set(t.id,[]);for(let t of N)q(t)&&e.get(t.source)?.push(t.target);let t=new Set;for(let n of R){let r=[n],i=new Set([n]);for(;r.length;){let n=r.shift();for(let a of e.get(n)??[])i.has(a)||(i.add(a),t.add(a),r.push(a))}}return t},[M,N,q,R]),ye=(0,A.useMemo)(()=>{let e=new Map;for(let t of N)q(t)&&(Y.has(t.target)||e.set(t.source,(e.get(t.source)??0)+1));return e},[N,q,Y]),be=(0,A.useCallback)((e,t)=>{e.stopPropagation(),oe(e=>{let n=new Set(e);return n.has(t)?n.delete(t):n.add(t),n})},[]),Se=(0,A.useMemo)(()=>{let e=new Map;for(let t of R){let n=0,r=new Set,i=[t];for(;i.length;){let e=i.shift();for(let t of N){if(t.source!==e||!q(t))continue;let a=t.target;r.has(a)||(r.add(a),Y.has(a)&&(n++,i.push(a)))}}e.set(t,n)}return e},[R,Y,N,q]),Ce=(0,A.useMemo)(()=>{if(!r.trim())return null;let e=r.toLowerCase(),t=new Set;for(let n of M)String(n.data.label??n.id).toLowerCase().includes(e)&&t.add(n.id);return t},[M,r]),Te=(0,A.useMemo)(()=>{let e=new Set,t=new Set;if(!c||!P.has(c))return{nodes:e,edges:t};let n=new Set,r=new Set,i=new Set,a=[c];for(;a.length;){let e=a.shift();if(!i.has(e)){i.add(e),n.add(e);for(let t of N){if(t.source!==e||!q(t))continue;r.add(t.id);let n=t.target;i.has(n)||a.push(n)}}}return{nodes:n,edges:r}},[c,l,N,q,P]),[Ne,Fe]=(0,A.useState)(new Set),[Ie,Ve]=(0,A.useState)(null),He=(0,A.useCallback)(e=>{let t=new Set;for(let n of N)(n.source===e||n.target===e)&&t.add(n.id);Fe(t),Ve(e),o(e)},[N,o]),Ue=(0,A.useCallback)(()=>{Fe(new Set),Ve(null),o(null)},[o]),We=(0,A.useCallback)((e,t,n,r)=>{e.stopPropagation(),e.currentTarget.setPointerCapture(e.pointerId),ie.current=!1,re.current={nodeId:t,startSX:e.clientX,startSY:e.clientY,origMX:n,origMY:r}},[]),Ge=(0,A.useCallback)((e,t)=>{let n=re.current;if(!n||n.nodeId!==t)return;let r=e.clientX-n.startSX,i=e.clientY-n.startSY;if(!ie.current&&Math.abs(r)<4&&Math.abs(i)<4)return;ie.current=!0;let a=tt.current.k;ee(e=>{let o=new Map(e);return o.set(t,{x:n.origMX+r/a,y:n.origMY+i/a}),o})},[]),Ke=(0,A.useCallback)((e,t)=>{re.current?.nodeId===t&&(re.current=null)},[]),qe=(0,A.useRef)(null),Je=(0,A.useRef)(null),Ye=(0,A.useRef)(null),Xe=(0,A.useRef)(null),Ze=(0,A.useRef)([]),Qe=(0,A.useRef)([]),$e=(0,A.useRef)(0),et=(0,A.useRef)(new Map),tt=(0,A.useRef)(w),nt=(0,A.useRef)(null),[rt,it]=(0,A.useState)(100),[at,ot]=(0,A.useState)(!0),st=(0,A.useCallback)((e,t,n=0,r=!1)=>{let i=N.find(t=>t.id===e);if(!i||!q(i))return;let a=K.current.get(i.source),o=K.current.get(i.target);if(!a||!o)return;let s=Pe(a,o),c=r&&Math.random()<.65?.15+Math.random()*.55:0,l=c>0?120+Math.random()*700:0,u=r&&Math.random()<.12,d=u?.25+Math.random()*.55:0;setTimeout(()=>{Ze.current.push({id:`${e}-${Date.now()}-${Math.random()}`,waypoints:s,progress:0,speed:9e-4+Math.random()*4e-4,color:t,pulse:0,sparkCooldown:0,tgtNodeId:i.target,chained:r,arrived:!1,stallAt:c,stallRemaining:l,timedOut:u,timeoutAt:d})},n)},[N,q]),ct=(0,A.useCallback)((e,t,n=0)=>{let r=Date.now();if(r-(et.current.get(e)??0)<1800)return;et.current.set(e,r);let i=0;for(let r of N)r.source===e&&q(r)&&(st(r.id,t,n+i*60,!0),i++)},[N,q,st]);(0,A.useEffect)(()=>{if(!c||!P.has(c))return;let e=()=>{let e=0;for(let t of N)t.source===c&&q(t)&&(st(t.id,`#a855f7`,e*80,!0),e++)};e();let t=window.setInterval(e,700);return()=>window.clearInterval(t)},[c,l,N,q,P,st]),(0,A.useEffect)(()=>{let e;function t(n){e=requestAnimationFrame(t);let r=Xe.current;if(!r)return;let i=Math.min(n-$e.current,50);$e.current=n;let a=r.getContext(`2d`);if(!a)return;a.clearRect(0,0,r.width,r.height);let o=tt.current,s=Math.max(.6,o.k);a.globalCompositeOperation=`lighter`;let l=[],u=M.length<=40||c,d=Ze.current.filter(e=>e.progress<1).length,f=Math.max(.12,1-Math.max(0,d-4)*.055);for(let e of Ze.current){if(!u)continue;if(e.timedOut&&e.timeoutAt>0&&e.progress>=e.timeoutAt){let t=e.waypoints.map(e=>je(e.x,e.y,o)),n=Me(e.timeoutAt,t);for(let e=0;e<18;e++){let t=e/18*Math.PI*2+Math.random()*.4,r=.06+Math.random()*.14;Qe.current.push({x:n.x,y:n.y,vx:Math.cos(t)*r,vy:Math.sin(t)*r,life:1,decay:.0014+Math.random()*.001,size:(1.4+Math.random()*2)*s,color:`#ef4444`})}continue}let t=e.stallAt>0&&e.progress>=e.stallAt&&e.stallRemaining>0;t?e.stallRemaining-=i:e.progress<1&&(e.progress=Math.min(1,e.progress+e.speed*f*i));let r=e.waypoints.map(e=>je(e.x,e.y,o)),c=r[r.length-1],d=Me(e.progress,r);if(!isFinite(d.x)||!isFinite(d.y)){l.push(e);continue}let p=e.stallAt>0&&e.stallRemaining>0?Math.min(1,e.stallRemaining/400):0,m=t?p>.5?`#f59e0b`:`#fb923c`:e.color;for(let t=18;t>=1;t--){let n=e.progress-t/18*.09;if(n<0)continue;let i=Me(n,r),o=1-t/18,c=o*o*.55,l=(.8+o*2.6)*s;a.beginPath(),a.arc(i.x,i.y,l,0,Math.PI*2),a.fillStyle=m+Ae(c*255),a.fill()}a.save(),a.shadowBlur=(t?34:24)*s,a.shadowColor=m,a.beginPath(),a.arc(d.x,d.y,5*s,0,Math.PI*2),a.fillStyle=m+`66`,a.fill(),a.restore();let h=a.createRadialGradient(d.x,d.y,0,d.x,d.y,8*s);if(h.addColorStop(0,`#ffffffee`),h.addColorStop(.35,m+`cc`),h.addColorStop(1,m+`00`),a.fillStyle=h,a.beginPath(),a.arc(d.x,d.y,8*s,0,Math.PI*2),a.fill(),t){let e=.5+.5*Math.sin(n*.012);a.beginPath(),a.arc(d.x,d.y,(10+e*8)*s,0,Math.PI*2),a.strokeStyle=`#f59e0b`+Ae(e*160),a.lineWidth=1.5*s,a.stroke()}let g=1+.18*Math.sin(n*.018+e.progress*12);if(a.beginPath(),a.arc(d.x,d.y,2.2*s*g,0,Math.PI*2),a.fillStyle=`#ffffff`,a.fill(),e.progress<1&&(e.sparkCooldown-=i,e.sparkCooldown<=0)){e.sparkCooldown=35+Math.random()*40;let t=Math.random()*Math.PI*2,n=.02+Math.random()*.04;Qe.current.push({x:d.x,y:d.y,vx:Math.cos(t)*n,vy:Math.sin(t)*n,life:1,decay:.0028+Math.random()*.0012,size:(.8+Math.random()*1.4)*s,color:m})}if(e.progress>=1){if(!e.arrived){e.arrived=!0;for(let t=0;t<14;t++){let n=t/14*Math.PI*2+Math.random()*.3,r=.08+Math.random()*.12;Qe.current.push({x:c.x,y:c.y,vx:Math.cos(n)*r,vy:Math.sin(n)*r,life:1,decay:.0018+Math.random()*8e-4,size:(1.2+Math.random()*1.6)*s,color:e.color})}if(e.chained){let t=P.get(e.tgtNodeId),n=t&&L[String(t.data.type)]||e.color;ct(e.tgtNodeId,n,120)}}if(e.pulse=Math.min(1,e.pulse+.025),e.pulse<1){for(let t=0;t<3;t++){let n=e.pulse-t*.18;if(n<=0||n>=1)continue;let r=(3+n*38)*s,i=(1-n)*(1-n)*220;a.beginPath(),a.arc(c.x,c.y,r,0,Math.PI*2),a.strokeStyle=e.color+Ae(i),a.lineWidth=1.5*s,a.stroke()}let t=(1-e.pulse)*(1-e.pulse)*255;a.save(),a.shadowBlur=18*s,a.shadowColor=e.color,a.beginPath(),a.arc(c.x,c.y,4*s,0,Math.PI*2),a.fillStyle=`#ffffff`+Ae(t),a.fill(),a.restore(),l.push(e)}}else l.push(e)}let p=[];for(let e of Qe.current){if(e.x+=e.vx*i,e.y+=e.vy*i,e.vx*=.985,e.vy*=.985,e.life-=e.decay*i,e.life<=0)continue;let t=Math.max(.3,e.size*e.life);a.beginPath(),a.arc(e.x,e.y,t,0,Math.PI*2),a.fillStyle=e.color+Ae(e.life*220),a.fill(),p.push(e)}Qe.current=p,a.globalCompositeOperation=`source-over`,Ze.current=l}return $e.current=performance.now(),e=requestAnimationFrame(t),()=>cancelAnimationFrame(e)},[P,ct,M.length,c]),(0,A.useEffect)(()=>{M.length>40&&!c&&(Ze.current=[],Qe.current=[])},[M.length,c]),(0,A.useEffect)(()=>{let e=qe.current,t=Xe.current;if(!e||!t)return;let n=new ResizeObserver(()=>{t.width=e.clientWidth,t.height=e.clientHeight});return n.observe(e),t.width=e.clientWidth,t.height=e.clientHeight,()=>n.disconnect()},[]),(0,A.useEffect)(()=>{let e=Je.current,t=Ye.current;if(!e||!t)return;let n=O().scaleExtent([.02,5]).filter(e=>!re.current&&(!e.ctrlKey||e.type===`wheel`)&&!(e.type===`wheel`&&Be(e))&&!e.button).on(`zoom`,e=>{tt.current=e.transform,S(t).attr(`transform`,e.transform.toString()),it(Math.round(e.transform.k*100))});S(e).call(n),nt.current=n;let r=t=>{if(!Be(t))return;t.preventDefault();let r=tt.current.k;S(e).call(n.translateBy,-t.deltaX/r,-t.deltaY/r)};return e.addEventListener(`wheel`,r,{passive:!1}),()=>{S(e).on(`.zoom`,null),e.removeEventListener(`wheel`,r)}},[]);let lt=(0,A.useCallback)(()=>{let e=Je.current,t=qe.current,n=nt.current;if(!e||!t||!n||!M.length)return;let r=1/0,i=1/0,a=-1/0,o=-1/0;for(let e of M)r=Math.min(r,e.x-e.width/2),a=Math.max(a,e.x+e.width/2),i=Math.min(i,e.y-e.height/2),o=Math.max(o,e.y+e.height/2);let s=a-r+96,c=o-i+96,l=t.clientWidth,u=t.clientHeight,d=Math.min(l/s,u/c,2)*.92,f=(r+a)/2,p=(i+o)/2,m=l/2-d*f,h=u/2-d*p,g=w.translate(m,h).scale(d);S(e).call(n.transform,g)},[M]),ut=(0,A.useCallback)(e=>{let t=Je.current,n=nt.current;!t||!n||S(t).transition().duration(150).call(n.scaleBy,e)},[]),dt=(0,A.useCallback)(async e=>{let t=qe.current;return t?(await(0,te.default)(t,{scale:e?.scale??2,useCORS:!0,backgroundColor:v?`#0a0c10`:`#f6f7f9`,ignoreElements:e=>e.classList?.contains(`g-rails`)||e.classList?.contains(`g-toolbar`)||e.classList?.contains(`g-breadcrumb`)||e.classList?.contains(`g-zoom`)})).toDataURL(`image/png`):null},[v]);(0,A.useEffect)(()=>(s.current={fit:lt,toPng:dt},()=>{s.current=null}),[s,lt,dt]);let ft=(0,A.useRef)(!1);return(0,A.useEffect)(()=>{ft.current=!1},[e]),(0,A.useEffect)(()=>{if(!M.length||ft.current)return;ft.current=!0;let e=requestAnimationFrame(()=>lt());return()=>cancelAnimationFrame(e)},[M.length,lt,e]),(0,X.jsxs)(`div`,{ref:qe,className:`g-canvas ${at?``:`g-no-edge-labels`}`,style:{position:`relative`,width:`100%`,height:`100%`},children:[(0,X.jsxs)(`svg`,{ref:Je,role:`img`,"aria-label":`Execution graph`,style:{width:`100%`,height:`100%`,display:`block`,cursor:`grab`,touchAction:`none`},children:[(0,X.jsxs)(`defs`,{children:[(0,X.jsx)(`marker`,{id:`arrow-def`,markerWidth:`9`,markerHeight:`9`,refX:`8`,refY:`4.5`,orient:`auto`,markerUnits:`strokeWidth`,children:(0,X.jsx)(`path`,{d:`M0,0 L0,9 L9,4.5 z`,fill:b})}),(0,X.jsx)(`marker`,{id:`arrow-hi`,markerWidth:`9`,markerHeight:`9`,refX:`8`,refY:`4.5`,orient:`auto`,markerUnits:`strokeWidth`,children:(0,X.jsx)(`path`,{d:`M0,0 L0,9 L9,4.5 z`,fill:ae})}),(0,X.jsx)(`marker`,{id:`arrow-st`,markerWidth:`9`,markerHeight:`9`,refX:`8`,refY:`4.5`,orient:`auto`,markerUnits:`strokeWidth`,children:(0,X.jsx)(`path`,{d:`M0,0 L0,9 L9,4.5 z`,fill:`#a855f7`})}),(0,X.jsx)(`marker`,{id:`arrow-region`,markerWidth:`8`,markerHeight:`8`,refX:`7`,refY:`4`,orient:`auto`,markerUnits:`strokeWidth`,children:(0,X.jsx)(`path`,{d:`M0,0.5 L0,7.5 L8,4 z`,fill:`context-stroke`})})]}),(0,X.jsxs)(`g`,{ref:Ye,children:[(0,X.jsx)(`rect`,{x:-1e5,y:-1e5,width:2e5,height:2e5,fill:`transparent`,onClick:Ue,style:{pointerEvents:`all`}}),U.map(e=>{let t=I[e.kind]??`#d99a2b`,n=ne[e.kind]??`6 5`,r=`${fe[e.kind]} ${e.index}`;return(0,X.jsxs)(`g`,{style:{pointerEvents:`none`},children:[e.pure&&(0,X.jsx)(`polygon`,{points:e.points.map(([e,t])=>`${e},${t}`).join(` `),fill:t,fillOpacity:.05,stroke:t,strokeWidth:1.5,strokeDasharray:n,opacity:.55}),!e.pure&&e.members.map(e=>(0,X.jsx)(`rect`,{x:e.x-e.width/2-5,y:e.y-e.height/2-5,width:e.width+10,height:e.height+10,rx:13,fill:`none`,stroke:t,strokeWidth:1.5,strokeDasharray:n,opacity:.85},e.id)),e.ordered&&e.members.slice(1).map((n,r)=>{let i=me(e.members[r],n);return i?(0,X.jsx)(`line`,{x1:i.x1,y1:i.y1,x2:i.x2,y2:i.y2,stroke:t,strokeWidth:1.4,opacity:.75,markerEnd:`url(#arrow-region)`},`${e.id}-${n.id}-step`):null}),e.pure?(0,X.jsx)(`text`,{x:Math.min(...e.points.map(([e])=>e))+10,y:Math.min(...e.points.map(([,e])=>e))-6-(W.get(e.id)??0)*11,fontSize:10,fontFamily:`ui-monospace, monospace`,fill:t,opacity:.9,children:r}):e.members.map((n,i)=>(0,X.jsx)(`text`,{x:n.x-n.width/2-4,y:n.y-n.height/2-10-(W.get(`${e.id}|${n.id}`)??0)*11,fontSize:9,fontFamily:`ui-monospace, monospace`,fill:t,opacity:.85,children:e.ordered?`${r} · ${i+1}`:r},`${n.id}-label`))]},e.id)}),N.map(e=>{if(!q(e)||R.has(e.source)||Y.has(e.source)||Y.has(e.target))return null;let t=G.get(e.source),n=G.get(e.target);if(!t||!n)return null;let{d:r,lx:i,ly:a}=Le(t,n),o={x:i,y:a},s=Re(e.data,v),c=Ne.has(e.id),l=Te.edges.has(e.id),u=y,d=1.75,f=`url(#arrow-def)`,p=1;return l&&(u=`#a855f7`,d=2,f=`url(#arrow-st)`,p=.7),c&&(u=ae,d=1.5,f=`url(#arrow-hi)`,p=1),Ce&&!(Ce.has(e.source)||Ce.has(e.target))&&(p*=.02),(0,X.jsxs)(`g`,{children:[(0,X.jsx)(`path`,{d:r,fill:`none`,stroke:u,strokeWidth:d,strokeLinecap:`round`,strokeLinejoin:`round`,opacity:p,markerEnd:f,style:{pointerEvents:`auto`}}),s&&p>.05&&(0,X.jsx)(`g`,{className:`g-edge-label`,transform:`translate(${o.x},${o.y})`,children:(0,X.jsx)(`text`,{textAnchor:`middle`,dominantBaseline:`middle`,fill:s.fill,fontSize:9,fontFamily:`ui-monospace, monospace`,style:{pointerEvents:`none`},children:(0,X.jsx)(`tspan`,{dx:0,dy:-8,paintOrder:`stroke fill`,stroke:s.bg,strokeWidth:6,strokeLinejoin:`round`,children:s.text})})})]},e.id)}),ue.map(e=>{if(Y.has(e.id))return null;let t=he(e.data.type),n=Ce&&!Ce.has(e.id),r=t?n?.07:1:0,i=Te.nodes.has(e.id),a=Ie===e.id,{bg:o,border:s,borderW:c,accent:l}=ze(e,v,u,a,i,d),{className:p,method:m}=J(String(e.data.label??e.id),e.data.method),h=m&&!m.includes(`(`)?m+`()`:m,g=String(e.data.type??``),_=e.width,y=e.height,b=_/2,x=y/2,S=v?`#e6edf3`:`#0d1117`,C=v?`rgba(255,255,255,0.5)`:`rgba(0,0,0,0.5)`,w=e.data.security,T=!!(e.data.hasN1||e.data.fatMethod||e.data.fatClass||w&&((w.issues?.length??0)>0||w.riskLevel&&w.riskLevel!==`none`)),E=e.data.httpCalls??[],D=Array.from(new Set(E.map(e=>e.host||e.configKey||`external`))).map(e=>e.length>14?e.slice(0,13)+`…`:e),O=p.length>24?p.slice(0,23)+`…`:p,k=h.length>26?h.slice(0,25)+`…`:h;return(0,X.jsxs)(`g`,{className:`g-node`,transform:`translate(${e.x},${e.y})`,opacity:r,style:{pointerEvents:t&&r>.05?`auto`:`none`,cursor:`grab`},onPointerDown:t=>We(t,e.id,e.x,e.y),onPointerMove:t=>Ge(t,e.id),onPointerUp:t=>Ke(t,e.id),onClick:t=>{t.stopPropagation(),ie.current||He(e.id)},children:[a&&(0,X.jsx)(`rect`,{x:-b-3,y:-x-3,width:_+6,height:y+6,rx:f?7:13,fill:`none`,stroke:l,strokeWidth:6,opacity:.15}),(0,X.jsx)(`rect`,{x:-b,y:-x,width:_,height:y,rx:f?6:10,fill:o,stroke:s,strokeWidth:c,filter:e.data.hasN1&&!u?`drop-shadow(0 0 8px rgba(244,67,54,0.4))`:void 0}),T&&(0,X.jsxs)(`g`,{style:{pointerEvents:`none`},children:[(0,X.jsx)(`circle`,{cx:b-3,cy:-x+3,r:10,fill:`#ef4444`,opacity:.22}),(0,X.jsx)(`circle`,{cx:b-3,cy:-x+3,r:5,fill:`#ef4444`,stroke:o,strokeWidth:1.5})]}),f?(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:-b+10,cy:0,r:3.5,fill:l}),(0,X.jsx)(`text`,{x:-b+20,y:0,fontSize:11,fontWeight:700,fontFamily:`ui-sans-serif, system-ui, -apple-system, sans-serif`,fill:S,dominantBaseline:`middle`,style:{pointerEvents:`none`},children:O}),!!e.data.hasN1&&(0,X.jsx)(`text`,{x:b-6,y:0,fontSize:9,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:`#F44336`,dominantBaseline:`middle`,style:{pointerEvents:`none`},children:`N+1`}),D.length>0&&!e.data.hasN1&&(0,X.jsx)(`text`,{x:b-6,y:0,fontSize:9,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:`#38bdf8`,dominantBaseline:`middle`,style:{pointerEvents:`none`},children:`🌐`}),d&&e.data.security&&(0,X.jsx)(`text`,{x:e.data.hasN1?b-28:b-6,y:0,fontSize:8,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:(z[e.data.security.exposure]??z.public).accent,dominantBaseline:`middle`,style:{pointerEvents:`none`},children:(z[e.data.security.exposure]??z.public).label.toUpperCase()})]}):(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:-b+14,cy:-x+18,r:4,fill:l}),(0,X.jsx)(`text`,{x:-b+24,y:-x+22,fontSize:10,fontFamily:`ui-monospace, monospace`,fill:l,opacity:.9,style:{pointerEvents:`none`},children:g}),!!e.data.hasN1&&(0,X.jsx)(`text`,{x:b-10,y:-x+22,fontSize:10,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:`#F44336`,style:{pointerEvents:`none`},children:`N+1`}),d&&e.data.security&&(()=>{let t=e.data.security,n=z[t.exposure]??z.public,r=B[t.riskLevel]??B.none;return(0,X.jsxs)(X.Fragment,{children:[(0,X.jsxs)(`text`,{x:e.data.hasN1?b-42:b-10,y:-x+22,fontSize:9,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:n.accent,style:{pointerEvents:`none`},children:[`🔒 `,n.label.toUpperCase()]}),t.riskLevel!==`none`&&(0,X.jsxs)(`text`,{x:b-10,y:-x+38,fontSize:8,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:r,style:{pointerEvents:`none`},children:[`⚠ `,t.issues.length,` issue`,t.issues.length===1?``:`s`]})]})})(),(0,X.jsx)(`text`,{x:-b+14,y:-x+46,fontSize:13,fontWeight:700,fontFamily:`ui-sans-serif, system-ui, -apple-system, sans-serif`,fill:S,style:{pointerEvents:`none`},children:O}),k&&(0,X.jsxs)(`text`,{x:-b+14,y:-x+64,fontSize:11,fontFamily:`ui-monospace, monospace`,fill:C,style:{pointerEvents:`none`},children:[`↻ `,k]}),D.length>0&&(0,X.jsxs)(`text`,{x:b-10,y:x-10,fontSize:9,textAnchor:`end`,fontFamily:`ui-monospace, monospace`,fill:`#38bdf8`,style:{pointerEvents:`none`},children:[`🌐 `,D[0],D.length>1?` +${D.length-1}`:``]})]}),(R.has(e.id)||(ye.get(e.id)??0)>4)&&(0,X.jsxs)(`g`,{transform:`translate(${b+2}, 0)`,onPointerDown:e=>e.stopPropagation(),onClick:t=>be(t,e.id),style:{cursor:`pointer`,pointerEvents:`all`},children:[(0,X.jsx)(`rect`,{x:0,y:-10,width:64,height:20,rx:10,fill:R.has(e.id)?l:v?`rgba(255,255,255,0.12)`:`rgba(0,0,0,0.10)`,stroke:l,strokeWidth:1}),(0,X.jsx)(`text`,{x:32,y:0,textAnchor:`middle`,dominantBaseline:`middle`,fill:R.has(e.id)?`#fff`:l,fontSize:10,fontWeight:700,fontFamily:`ui-monospace, monospace`,style:{pointerEvents:`none`},children:R.has(e.id)?`▶ ${Se.get(e.id)??ye.get(e.id)} hidden`:`▾ fold`})]})]},e.id)})]})]}),(0,X.jsx)(`canvas`,{ref:Xe,style:{position:`absolute`,top:0,left:0,pointerEvents:`none`,width:`100%`,height:`100%`}}),(u||d)&&(0,X.jsxs)(`div`,{className:`g-legends`,children:[u&&(0,X.jsxs)(`div`,{className:`cc-legend`,children:[(0,X.jsx)(`div`,{className:`cc-legend-title`,children:`Cyclomatic Complexity`}),le.map(e=>(0,X.jsxs)(`div`,{className:`cc-legend-row`,children:[(0,X.jsx)(`span`,{className:`cc-legend-swatch`,style:{background:e.border}}),(0,X.jsx)(`span`,{className:`cc-legend-label`,style:{color:e.border},children:e.label}),(0,X.jsx)(`span`,{className:`cc-legend-range`,children:e.max===1/0?`≥${e.min}`:`${e.min}–${e.max}`})]},e.label))]}),d&&(0,X.jsxs)(`div`,{className:`cc-legend`,children:[(0,X.jsx)(`div`,{className:`cc-legend-title`,children:`🔒 Security Surface`}),Object.entries(z).map(([e,t])=>(0,X.jsxs)(`div`,{className:`cc-legend-row`,children:[(0,X.jsx)(`span`,{className:`cc-legend-swatch`,style:{background:t.border}}),(0,X.jsx)(`span`,{className:`cc-legend-label`,style:{color:t.accent},children:t.label})]},e)),(0,X.jsx)(`div`,{className:`cc-legend-title`,style:{marginTop:`8px`},children:`Risk Level`}),[{key:`critical`,label:`Critical`,color:B.critical},{key:`high`,label:`High`,color:B.high},{key:`medium`,label:`Medium`,color:B.medium},{key:`none`,label:`Clean`,color:B.none}].map(({key:e,label:t,color:n})=>(0,X.jsxs)(`div`,{className:`cc-legend-row`,children:[(0,X.jsx)(`span`,{className:`cc-legend-swatch`,style:{background:n}}),(0,X.jsx)(`span`,{className:`cc-legend-label`,style:{color:n},children:t})]},e))]})]}),(0,X.jsx)(`div`,{className:`g-rails`,"aria-hidden":!0,children:[{n:1,label:`Route`,c:`var(--nc-route)`},{n:2,label:`Controller`,c:`var(--nc-controller)`},{n:3,label:`Action`,c:`var(--nc-action)`},{n:4,label:`Service · View`,c:`var(--nc-service)`},{n:5,label:`Interface`,c:`var(--nc-interface)`},{n:6,label:`Implementation`,c:`var(--nc-provider)`}].map(e=>(0,X.jsxs)(`div`,{className:`g-rail`,children:[(0,X.jsx)(`span`,{className:`g-rail-pill`,style:{"--rc":e.c},children:e.n}),(0,X.jsx)(`span`,{className:`g-rail-label`,children:e.label})]},e.n))}),(0,X.jsxs)(`div`,{className:`g-toolbar`,children:[(0,X.jsxs)(`select`,{className:`g-tool-select`,value:t,onChange:e=>p(e.target.value),title:`Layout algorithm`,children:[(0,X.jsx)(`option`,{value:`dagre`,children:`Hierarchical`}),(0,X.jsx)(`option`,{value:`breadthfirst`,children:`Breadth-first`}),(0,X.jsx)(`option`,{value:`cose-bilkent`,children:`Force`}),(0,X.jsx)(`option`,{value:`circle`,children:`Circle`}),(0,X.jsx)(`option`,{value:`grid`,children:`Grid`})]}),(0,X.jsx)(`button`,{type:`button`,className:`g-tool ${n===`TB`?`g-tool--on`:``}`,onClick:()=>m(n===`TB`?`LR`:`TB`),title:`Toggle orientation`,children:n===`TB`?`Top-down`:`Left-right`}),(0,X.jsx)(`span`,{className:`g-tool-sep`}),(0,X.jsx)(`button`,{type:`button`,className:`g-tool ${at?`g-tool--on`:``}`,onClick:()=>ot(e=>!e),children:`Edge labels`}),(0,X.jsx)(`button`,{type:`button`,className:`g-tool ${u?`g-tool--on`:``}`,onClick:h,children:`Complexity`}),(0,X.jsx)(`button`,{type:`button`,className:`g-tool ${d?`g-tool--on`:``}`,onClick:g,children:`Security`}),(0,X.jsx)(`button`,{type:`button`,className:`g-tool ${f?`g-tool--on`:``}`,onClick:_,children:`Compact`})]}),(0,X.jsxs)(`div`,{className:`g-breadcrumb`,children:[[{label:`Route`,c:`var(--nc-route)`},{label:`Controller`,c:`var(--nc-controller)`},{label:`Action`,c:`var(--nc-action)`},{label:`Service`,c:`var(--nc-service)`},{label:`Interface`,c:`var(--nc-interface)`},{label:`Impl`,c:`var(--nc-provider)`}].map((e,t,n)=>(0,X.jsxs)(`span`,{className:`g-crumb`,children:[(0,X.jsx)(`span`,{className:`g-crumb-dot`,style:{background:e.c}}),e.label,t<n.length-1&&(0,X.jsx)(`span`,{className:`g-crumb-arrow`,children:`→`})]},e.label)),V.map(e=>{let t=U.filter(t=>t.kind===e).length;return t===0?null:(0,X.jsxs)(`span`,{className:`g-crumb g-crumb--aside`,children:[(0,X.jsx)(`span`,{className:`g-crumb-sep`,children:`·`}),(0,X.jsx)(`span`,{className:`g-crumb-dot g-crumb-dot--dashed`,style:{borderColor:I[e]}}),t===1?fe[e]:`${t} ${pe[e]}`]},e)})]}),(0,X.jsxs)(`div`,{className:`g-zoom`,children:[(0,X.jsx)(`button`,{type:`button`,className:`g-zoom-btn`,onClick:()=>ut(.8),"aria-label":`Zoom out`,children:`−`}),(0,X.jsxs)(`span`,{className:`g-zoom-pct`,children:[rt,`%`]}),(0,X.jsx)(`button`,{type:`button`,className:`g-zoom-btn`,onClick:()=>ut(1.25),"aria-label":`Zoom in`,children:`+`}),(0,X.jsx)(`button`,{type:`button`,className:`g-zoom-btn g-zoom-fit`,onClick:()=>lt(),"aria-label":`Fit to view`,children:`⊡`})]})]})}var He=`modulepreload`,Ue=function(e){return`/_laravel-brain/`+e},We={},Ge=function(e,t,n){let r=Promise.resolve();if(t&&t.length>0){let e=document.getElementsByTagName(`link`),i=document.querySelector(`meta[property=csp-nonce]`),a=i?.nonce||i?.getAttribute(`nonce`);function o(e){return Promise.all(e.map(e=>Promise.resolve(e).then(e=>({status:`fulfilled`,value:e}),e=>({status:`rejected`,reason:e}))))}r=o(t.map(t=>{if(t=Ue(t,n),t in We)return;We[t]=!0;let r=t.endsWith(`.css`),i=r?`[rel="stylesheet"]`:``;if(n)for(let n=e.length-1;n>=0;n--){let i=e[n];if(i.href===t&&(!r||i.rel===`stylesheet`))return}else if(document.querySelector(`link[href="${t}"]${i}`))return;let o=document.createElement(`link`);if(o.rel=r?`stylesheet`:He,r||(o.as=`script`),o.crossOrigin=``,o.href=t,a&&o.setAttribute(`nonce`,a),document.head.appendChild(o),r)return new Promise((e,n)=>{o.addEventListener(`load`,e),o.addEventListener(`error`,()=>n(Error(`Unable to preload CSS for ${t}`)))})}))}function i(e){let t=new Event(`vite:preloadError`,{cancelable:!0});if(t.payload=e,window.dispatchEvent(t),!t.defaultPrevented)throw e}return r.then(t=>{for(let e of t||[])e.status===`rejected`&&i(e.reason);return e().catch(i)})},Ke=`route.middleware.controller.action.action_class.service.validation_request.repository.model.job.event.view.mail.notification.enum.interface.trait.abstract_class.service_provider.ai_agent.ai_tool.filament_panel.filament_resource.filament_page.filament_page_method.filament_widget.filament_relation_manager`.split(`.`);function qe(e,t){let n=[];n.push(`%%{init: {'theme': 'dark', 'themeVariables': {`),n.push(`  'background': '#0a0c10',`),n.push(`  'mainBkg': '#0d1117',`),n.push(`  'lineColor': 'rgba(255,255,255,0.35)',`),n.push(`  'edgeLabelBackground': '#111218',`),n.push(`  'edgeLabelColor': 'rgba(255,255,255,0.5)'`),n.push(`}}}%%`),n.push(`%% Laravel Brain — ${t}`),n.push(`flowchart TD`),n.push(``);let r=new Map,i=new Set,a=e=>{if(r.has(e))return r.get(e);let t=e.replace(/[^a-zA-Z0-9_]/g,`_`).replace(/^_+/,``).replace(/_+$/,``).substring(0,40);t||=`node`;let n=t,a=0;for(;i.has(n);)n=`${t}_${++a}`;return i.add(n),r.set(e,n),n},o=new Map;for(let t of e.nodes)o.has(t.type)||o.set(t.type,[]),o.get(t.type).push(t);let s=[...new Set([...Ke,...o.keys()])].filter(e=>(o.get(e)?.length??0)>0);for(let e of s){let t=o.get(e);n.push(`  %% ${e}`);for(let e of t){let t=a(e.id),r=Je(e);n.push(`  ${t}["${rt(r)}"]`)}n.push(``)}n.push(`  %% Edges`);for(let t of e.edges){let e=a(t.source),r=a(t.target),i=t.label?`|"${rt(t.label)}"| `:``;n.push(`  ${e} -->${i}${r}`)}n.push(``),n.push(`  %% Styles`);for(let e of s){let t=L[e]??`#c9d1d9`,r=ie[e]??`#0d1117`;n.push(`  classDef cls_${e} fill:${r},stroke:${t},stroke-width:2px,color:#e6edf3`)}n.push(``);for(let e of s){let t=o.get(e).map(e=>a(e.id)).join(`,`);n.push(`  class ${t} cls_${e}`)}return n.join(`
-`)}function Je(e){let t=String(e.label??``),n=e.data?.method,{className:r,method:i}=J(t,n),a=i&&!i.includes(`(`)?i+`()`:i,o=[`● ${e.type}`,r];return a&&o.push(`↻ ${a}`),o.join(`
-`)}function Ye(e,t){let n=[`%% Method Flow — ${t}`,`flowchart TD`],r=0,i=()=>`s${r++}`,a=i();n.push(`  ${a}([" 🚀 ${rt(t)} "])`);let o=(e,t)=>{let r=t;for(let t of e){let e=i();if(t.type===`if`){let[a,s]=[`{`,`}`];if(n.push(`  ${e}${a}"${rt(t.label)}"${s}`),n.push(`  ${r} --> ${e}`),n.push(`  class ${e} cls_if`),t.then&&t.then.length>0){let r=i(),a=t.then[0];n.push(`  ${r}${et(a.type)}"${rt(a.label)}"${tt(a.type)}`),n.push(`  ${e} -->|"yes"| ${r}`),n.push(`  class ${r} cls_${a.type}`),o(t.then.slice(1),r)}if(t.else&&t.else.length>0){let r=i(),a=t.else[0];n.push(`  ${r}${et(a.type)}"${rt(a.label)}"${tt(a.type)}`),n.push(`  ${e} -->|"no"| ${r}`),n.push(`  class ${r} cls_${a.type}`),o(t.else.slice(1),r)}r=e}else if(t.type===`loop`){let i=t.n1?` ⚠️ N+1 `:``,a=t.cache?` [cache ${t.cache.kind}] `:``;n.push(`  ${e}[/"${i}${a}${rt(t.label)}"/]`),n.push(`  ${r} --> ${e}`),n.push(`  class ${e} ${t.n1?`cls_n1`:t.cache?`cls_cache`:`cls_loop`}`),t.body&&t.body.length>0&&o(t.body,e),r=e}else{let[i,a]=[et(t.type),tt(t.type)],o=nt(t.type),s=t.n1?` ⚠️ N+1 `:``,c=t.cache?` [cache ${t.cache.kind}] `:``;n.push(`  ${e}${i}"${s}${c}${o}${rt(t.label)}"${a}`),n.push(`  ${r} --> ${e}`),n.push(`  class ${e} ${t.n1?`cls_n1`:`cls_${t.type}`}`),r=e}}return r};return o(e,a),n.push(``),n.push(`  %% STYLES`),n.push(`  classDef cls_call     fill:#0d47a1,stroke:#2196F3,color:#fff`),n.push(`  classDef cls_assign   fill:#212121,stroke:#616161,color:#ccc`),n.push(`  classDef cls_return   fill:#1b5e20,stroke:#4CAF50,color:#fff`),n.push(`  classDef cls_throw    fill:#b71c1c,stroke:#F44336,color:#fff`),n.push(`  classDef cls_if       fill:#f9a825,stroke:#fbc02d,color:#000`),n.push(`  classDef cls_loop     fill:#6a1b9a,stroke:#9c27b0,color:#fff`),n.push(`  classDef cls_n1       fill:#b71c1c,stroke:#ff5252,color:#fff`),n.push(`  classDef cls_dispatch fill:#bf360c,stroke:#FF5722,color:#fff`),n.push(`  classDef cls_event    fill:#0e47a1,stroke:#00BCD4,color:#fff`),n.push(`  classDef cls_cache    fill:#004d40,stroke:#009688,color:#fff`),n.join(`
-`)}function Xe(e,t){Qe(new Blob([e],{type:`text/plain`}),t)}function Ze(e,t){let n=document.createElement(`a`);n.href=e,n.download=t,n.click()}function Qe(e,t){let n=URL.createObjectURL(e),r=document.createElement(`a`);r.href=n,r.download=t,r.click(),setTimeout(()=>URL.revokeObjectURL(n),2e3)}async function $e(t,n=`#0d0f14`){let{default:r}=await Ge(async()=>{let{default:t}=await import(`./vendor-utils-D7YtnGoz.js`).then(t=>e(t.t(),1));return{default:t}},__vite__mapDeps([0,1]));return(await r(t,{backgroundColor:n,scale:2,useCORS:!0,logging:!1})).toDataURL(`image/png`)}function et(e){switch(e){case`return`:return`([`;case`throw`:return`([`;case`dispatch`:return`[[`;case`event`:return`((`;default:return`[`}}function tt(e){switch(e){case`return`:return`])`;case`throw`:return`])`;case`dispatch`:return`]]`;case`event`:return`))`;default:return`]`}}function nt(e){switch(e){case`call`:return`→ `;case`assign`:return`= `;case`return`:return`◀ `;case`throw`:return`⚠ `;case`dispatch`:return`⚡ `;case`event`:return`📡 `;case`cache`:return`⛃ `;default:return``}}function rt(e){return e.replace(/"/g,`'`).replace(/\n/g,`\\n`).replace(/[<>]/g,e=>e===`<`?`&lt;`:`&gt;`)}function it({mermaidCode:e,filename:t,title:n,onClose:r}){let[i,a]=(0,A.useState)(!1),o=(0,A.useRef)(null);return(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]),(0,X.jsx)(`div`,{className:`export-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,X.jsxs)(`div`,{className:`export-modal`,children:[(0,X.jsxs)(`div`,{className:`export-modal-header`,children:[(0,X.jsxs)(`div`,{className:`export-modal-title`,children:[(0,X.jsx)(`span`,{className:`export-modal-icon`,children:`🗺`}),(0,X.jsxs)(`div`,{children:[(0,X.jsx)(`h2`,{children:n}),(0,X.jsx)(`span`,{className:`export-modal-sub`,children:`Mermaid Flowchart`})]})]}),(0,X.jsx)(`button`,{className:`export-modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,X.jsxs)(`div`,{className:`export-modal-actions`,children:[(0,X.jsx)(`button`,{className:`export-btn export-btn--primary`,onClick:async()=>{try{await navigator.clipboard.writeText(e),a(!0),setTimeout(()=>a(!1),2e3)}catch{o.current?.select(),document.execCommand(`copy`),a(!0),setTimeout(()=>a(!1),2e3)}},children:i?`✓ Copied!`:`⎘ Copy Code`}),(0,X.jsx)(`button`,{className:`export-btn export-btn--secondary`,onClick:()=>Xe(e,t),children:`↓ Download .mmd`}),(0,X.jsx)(`button`,{className:`export-btn export-btn--accent`,onClick:()=>{let t=JSON.stringify({code:e,mermaid:`{}`,autoSync:!0}),n=btoa(unescape(encodeURIComponent(t)));window.open(`https://mermaid.live/edit#base64:${n}`,`_blank`)},children:`↗ Open in Mermaid Live`})]}),(0,X.jsxs)(`div`,{className:`export-modal-hint`,children:[`Paste this code at`,` `,(0,X.jsx)(`a`,{href:`https://mermaid.live`,target:`_blank`,rel:`noreferrer`,children:`mermaid.live`}),` `,`to render the diagram, or use any Mermaid-compatible tool.`]}),(0,X.jsxs)(`div`,{className:`export-code-wrapper`,children:[(0,X.jsx)(`div`,{className:`export-code-lang`,children:`mermaid`}),(0,X.jsx)(`textarea`,{ref:o,className:`export-code`,value:e,readOnly:!0,spellCheck:!1,onClick:e=>e.target.select()})]}),(0,X.jsxs)(`div`,{className:`export-modal-stats`,children:[(0,X.jsxs)(`span`,{children:[e.split(`
-`).length,` lines`]}),(0,X.jsxs)(`span`,{children:[(e.length/1024).toFixed(1),` KB`]})]})]})})}function at({steps:e,title:t,isFatMethod:n}){let[r,i]=(0,A.useState)(!1),[a,o]=(0,A.useState)(!1),s=(0,A.useRef)(null);if(!e||e.length===0)return(0,X.jsx)(`div`,{className:`flowchart-empty`,children:(0,X.jsx)(`span`,{children:`No flow data available`})});let c=t??`method`;return(0,X.jsxs)(X.Fragment,{children:[n&&(0,X.jsx)(`div`,{className:`flowchart-fat-banner`,title:`Fat Method: this method exceeds complexity or line-count thresholds`,children:`🧱 Fat Method — consider breaking this into smaller methods`}),(0,X.jsxs)(`div`,{className:`flowchart-export-bar`,children:[(0,X.jsxs)(`button`,{className:`flowchart-export-btn`,onClick:async()=>{if(s.current){o(!0);try{Ze(await $e(s.current),`${c.replace(/[^a-z0-9]/gi,`_`)}_flow.png`)}finally{o(!1)}}},disabled:a,title:`Export as PNG`,children:[a?`⏳`:`🖼`,` PNG`]}),(0,X.jsx)(`button`,{className:`flowchart-export-btn`,onClick:()=>i(!0),title:`Export as Mermaid`,children:`🗺 Mermaid`})]}),(0,X.jsxs)(`div`,{className:`flowchart-root`,ref:s,children:[t&&(0,X.jsx)(`div`,{className:`flowchart-title`,children:t}),(0,X.jsx)(ot,{steps:e})]}),r&&(0,X.jsx)(it,{mermaidCode:Ye(e,c),filename:`${c.replace(/[^a-z0-9]/gi,`_`)}_flow.mmd`,title:c,onClose:()=>i(!1)})]})}function ot({steps:e}){return(0,X.jsx)(`div`,{className:`flowchart-list`,children:e.map((t,n)=>(0,X.jsx)(st,{step:t,isLast:n===e.length-1},n))})}function st({step:e,isLast:t}){return e.type===`if`?(0,X.jsxs)(`div`,{className:`flowchart-branch-wrapper`,children:[(0,X.jsx)(ct,{step:e}),(0,X.jsxs)(`div`,{className:`flowchart-branches`,children:[e.then&&e.then.length>0&&(0,X.jsxs)(`div`,{className:`flowchart-branch flowchart-branch--then`,children:[(0,X.jsx)(`div`,{className:`flowchart-branch-label`,children:`then`}),(0,X.jsx)(ot,{steps:e.then})]}),e.else&&e.else.length>0&&(0,X.jsxs)(`div`,{className:`flowchart-branch flowchart-branch--else`,children:[(0,X.jsx)(`div`,{className:`flowchart-branch-label`,children:`else`}),(0,X.jsx)(ot,{steps:e.else})]})]}),!t&&(0,X.jsx)(ut,{})]}):e.type===`loop`?(0,X.jsxs)(`div`,{className:`flowchart-branch-wrapper`,children:[(0,X.jsx)(ct,{step:e}),e.body&&e.body.length>0&&(0,X.jsx)(`div`,{className:`flowchart-loop-body`,children:(0,X.jsx)(ot,{steps:e.body})}),!t&&(0,X.jsx)(ut,{})]}):(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(ct,{step:e}),!t&&(0,X.jsx)(ut,{})]})}function ct({step:e}){let t=`flowchart-box flowchart-box--${e.type} ${e.n1?`flowchart-box--n1`:``}`,n=dt[e.type]??``;return(0,X.jsxs)(`div`,{className:`${t} flowchart-shape--${e.type===`if`?`diamond`:e.type===`return`||e.type===`throw`?`terminal`:`rect`}`,title:e.label,children:[n&&(0,X.jsx)(`span`,{className:`flowchart-icon`,children:n}),(0,X.jsx)(`span`,{className:`flowchart-label`,children:e.label}),e.cache&&(0,X.jsx)(`span`,{className:`flowchart-cache-badge flowchart-cache-badge--${e.cache.kind}`,title:lt(e),children:e.cache.kind}),e.n1&&(0,X.jsx)(`span`,{className:`flowchart-n1-warn`,title:`N+1 Query Detected: This database operation is inside a loop!`,children:`⚠️ N+1`}),e.http&&e.http.length>0&&(0,X.jsxs)(`span`,{className:`flowchart-http`,title:e.http.map(e=>`${e.method||`REQUEST`} ${e.host||e.configKey&&`config('${e.configKey}')`||`address computed at runtime`}`+(e.timeout===null?` · no timeout`:` · timeout ${e.timeout}s`)).join(`
-`),children:[`🌐 `,e.http.map(e=>e.host).find(Boolean)??`external`]})]})}function lt(e){let t=e.cache;if(!t)return``;let n=t.keyKind===`computed`?`computed key`:t.keyKind===`none`?`whole store`:`"${t.key}"`,r=[t.ttl===null?``:`ttl ${t.ttl}s`,t.store===``?``:`store ${t.store}`,t.tags.length>0?`tags ${t.tags.join(`, `)}`:``].filter(Boolean);return`${t.kind} · ${t.method} ${n}${r.length>0?` · ${r.join(` · `)}`:``}`}function ut(){return(0,X.jsxs)(`div`,{className:`flowchart-arrow`,children:[(0,X.jsx)(`div`,{className:`flowchart-arrow-line`}),(0,X.jsx)(`div`,{className:`flowchart-arrow-head`})]})}var dt={call:`→`,assign:`=`,return:`◀`,throw:`⚠`,if:`◆`,loop:`↻`,dispatch:`⚡`,event:`📡`,cache:`⛃`};function ft({steps:e,title:t,isFatMethod:n,onClose:r}){return(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]),(0,X.jsx)(`div`,{className:`modal-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,X.jsxs)(`div`,{className:`modal-container modal-container--large`,children:[(0,X.jsxs)(`div`,{className:`modal-header`,children:[(0,X.jsxs)(`div`,{className:`modal-title`,children:[(0,X.jsx)(`span`,{className:`modal-icon`,children:`⛓`}),(0,X.jsxs)(`div`,{children:[(0,X.jsx)(`h2`,{children:t}),(0,X.jsx)(`span`,{className:`modal-sub`,children:`Method Flow Visualization`})]})]}),(0,X.jsx)(`button`,{className:`modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,X.jsx)(`div`,{className:`modal-body flowchart-modal-body`,children:(0,X.jsx)(at,{steps:e,isFatMethod:n})})]})})}function pt(e){let[t,n]=(0,A.useState)(null),[r,i]=(0,A.useState)(!1),[a,o]=(0,A.useState)(null),[s,c]=(0,A.useState)(e);return e!==s&&(c(e),e?(i(!0),o(null),n(null)):(n(null),o(null))),(0,A.useEffect)(()=>{e&&fetch(`/_laravel-brain/api/source?path=${encodeURIComponent(e)}`).then(e=>e.json()).then(e=>{if(e.error)throw Error(e.error);n(e.content)}).catch(e=>o(e.message)).finally(()=>i(!1))},[e]),{content:t,loading:r,error:a}}s.registerLanguage(`php`,u);function mt({filePath:e,highlightLine:t,theme:n}){let{content:r,loading:i,error:o}=pt(e),c=(0,A.useRef)(null);(0,A.useEffect)(()=>{c.current&&c.current.scrollIntoView({block:`center`,behavior:`smooth`})},[r]);let l=e.replace(/.*\/(app|src)\//,`$1/`);return i?(0,X.jsxs)(`div`,{className:`source-state`,children:[(0,X.jsx)(`div`,{className:`loading-spinner`,style:{width:20,height:20,borderWidth:2}}),(0,X.jsx)(`span`,{children:`Loading source…`})]}):o?(0,X.jsxs)(`div`,{className:`source-state source-state--error`,children:[`Could not load file`,(0,X.jsx)(`small`,{style:{display:`block`,opacity:.6,marginTop:4},children:o})]}):r?(0,X.jsxs)(`div`,{className:`source-view`,children:[(0,X.jsx)(`div`,{className:`source-path`,title:e,children:l}),(0,X.jsx)(s,{language:`php`,style:n===`dark`?p:a,showLineNumbers:!0,wrapLines:!0,lineNumberStyle:{minWidth:`2.5em`,paddingRight:`1em`,userSelect:`none`,opacity:.4,fontSize:11},lineProps:e=>e===t?{ref:c,style:{display:`block`,backgroundColor:n===`dark`?`rgba(139,111,232,0.2)`:`rgba(139,111,232,0.12)`,borderLeft:`3px solid #8B6FE8`}}:{style:{display:`block`}},customStyle:{margin:0,padding:`12px 0`,background:`transparent`,fontSize:12,lineHeight:`1.6`,fontFamily:`ui-monospace, "Cascadia Code", monospace`},children:r})]}):null}function ht({filePath:e,highlightLine:t,theme:n,onClose:r}){(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]);let i=e.split(`/`).pop()||`Source Code`;return(0,X.jsx)(`div`,{className:`modal-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,X.jsxs)(`div`,{className:`modal-container modal-container--large`,children:[(0,X.jsxs)(`div`,{className:`modal-header`,children:[(0,X.jsxs)(`div`,{className:`modal-title`,children:[(0,X.jsx)(`span`,{className:`modal-icon`,children:`📄`}),(0,X.jsxs)(`div`,{children:[(0,X.jsx)(`h2`,{children:i}),(0,X.jsx)(`span`,{className:`modal-sub`,children:e})]})]}),(0,X.jsx)(`button`,{className:`modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,X.jsx)(`div`,{className:`modal-body source-modal-body`,children:(0,X.jsx)(mt,{filePath:e,highlightLine:t,theme:n})})]})})}function gt(e){let[t,n]=(0,A.useState)(null),[r,i]=(0,A.useState)(!1),[a,o]=(0,A.useState)(null),[s,c]=(0,A.useState)(e);return e!==s&&(c(e),e?(i(!0),o(null),n(null)):(n(null),o(null))),(0,A.useEffect)(()=>{e&&fetch(`/_laravel-brain/api/usages?nodeId=${encodeURIComponent(e)}`).then(e=>e.json()).then(e=>{if(e.error)throw Error(e.error);n(e)}).catch(e=>o(e.message)).finally(()=>i(!1))},[e]),{data:t,loading:r,error:a}}function _t({nodeId:e}){let{data:t,loading:n,error:r}=gt(e);return n?(0,X.jsxs)(`div`,{className:`source-state`,children:[(0,X.jsx)(`div`,{className:`loading-spinner`,style:{width:20,height:20,borderWidth:2}}),(0,X.jsx)(`span`,{children:`Finding usages…`})]}):r?(0,X.jsxs)(`div`,{className:`source-state source-state--error`,children:[`Could not load usages`,(0,X.jsx)(`small`,{style:{display:`block`,opacity:.6,marginTop:4},children:r})]}):t?t.usageCount===0?(0,X.jsx)(`div`,{className:`sidebar-section`,children:(0,X.jsxs)(`div`,{className:`security-clean`,children:[(0,X.jsx)(`span`,{children:`✓`}),` Not used anywhere else in the project.`]})}):(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsxs)(`h3`,{children:[`Used in `,t.fileCount,` file`,t.fileCount===1?``:`s`,` · `,t.usageCount,` reference`,t.usageCount===1?``:`s`]}),t.files.map(e=>(0,X.jsxs)(`div`,{style:{marginBottom:12},children:[(0,X.jsxs)(`span`,{className:`ins-chip ins-chip--neutral`,title:e.file??`Location could not be resolved`,style:{display:`inline-block`,marginBottom:6},children:[e.file?e.file.split(`/`).slice(-2).join(`/`):`Unresolved location`,` · `,e.count]}),e.usages.map(e=>(0,X.jsxs)(`div`,{className:`edge-row`,children:[(0,X.jsx)(`span`,{className:`edge-target`,children:e.label}),(0,X.jsx)(`span`,{className:`edge-label`,children:e.edgeLabel})]},e.nodeId))]},e.file??`#${e.usages[0]?.nodeId??``}`))]}):null}var vt=new Set([`POST`,`PUT`,`PATCH`,`QUERY`]),yt=new Set([`POST`,`PUT`,`PATCH`,`DELETE`,`QUERY`]);function bt(e){let t=parseInt(e,10);return t>=200&&t<300?`#22c55e`:t>=400&&t<500?`#f97316`:t>=500?`#ef4444`:`#6b7280`}function xt(e){let t=Math.floor((Date.now()-e)/1e3);return t<60?`${t}s ago`:t<3600?`${Math.floor(t/60)}m ago`:`${Math.floor(t/3600)}h ago`}var St=new Map;function Z(e){let t=St.get(e);if(t)return t;try{let t=localStorage.getItem(`lb_st_${e}`);if(t){let n=JSON.parse(t);return St.set(e,n),n}}catch{}}function Ct(e,t){let n={...t,savedAt:Date.now()};St.set(e,n);try{localStorage.setItem(`lb_st_${e}`,JSON.stringify(n))}catch{}}function wt(e){let t=new Set,n=[];for(let r of e.matchAll(/\{([^}?]+)(\?)?\}/g))t.has(r[1])||(n.push({name:r[1],optional:!!r[2]}),t.add(r[1]));return n}function Tt(e,t){let n=e;return n=n.replace(/\/\{([^}?]+)\?\}/g,(e,n)=>{let r=t[n]?.trim();return r?`/`+encodeURIComponent(r):``}),n=n.replace(/\{([^}?]+)\}/g,(e,n)=>encodeURIComponent(t[n]?.trim()??``)),n||`/`}function Et(e){try{let t=JSON.parse(e);return typeof t!=`object`||!t||Array.isArray(t)?null:Object.entries(t).map(([e,t])=>`${encodeURIComponent(e)}=${encodeURIComponent(String(t))}`).join(`&`)}catch{return null}}function Dt({method:e,uri:t,selectedId:n,onStressChange:r}){let i=`${e}::${t}`,a=wt(t),[o,s]=(0,A.useState)(()=>{let e=window.location.href,t=e.indexOf(`/_laravel-brain`);return t===-1?window.location.origin:e.slice(0,t)}),[c,l]=(0,A.useState)(()=>Z(i)?.count??10),[u,d]=(0,A.useState)(()=>Z(i)?.concurrency??2),[f,p]=(0,A.useState)(()=>Z(i)?.headersRaw??``),[m,h]=(0,A.useState)(()=>Z(i)?.body??(vt.has(e.toUpperCase())?`{}`:``)),[g,_]=(0,A.useState)(()=>Z(i)?.timeout??10),[v,y]=(0,A.useState)(()=>{let e=Z(i);return!!(e?.jobId&&!e?.result)}),[b,x]=(0,A.useState)(()=>Z(i)?.jobId??null),[S,C]=(0,A.useState)(()=>Z(i)?.result??null),[w,T]=(0,A.useState)(()=>Z(i)?.error??null),[E,D]=(0,A.useState)(()=>Z(i)?.routeParams??{}),[O,k]=(0,A.useState)(()=>Z(i)?.includeCsrf??yt.has(e.toUpperCase())),[j,M]=(0,A.useState)(()=>Z(i)?.sendAsFormData??yt.has(e.toUpperCase())),[N,P]=(0,A.useState)(0),F=(0,A.useRef)(null),ee=(0,A.useRef)({result:S,error:w,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:b,routeParams:E,includeCsrf:O,sendAsFormData:j,key:i});async function te(e,t){let n=0;for(;n<180;){if(t.aborted||(await new Promise(e=>setTimeout(e,1e3)),n++,P(n),t.aborted))return;try{let n=await(await fetch(`/_laravel-brain/api/stress-test/${e}`,{signal:t})).json();if(n.status===`done`){let e=n.result;C(e),x(null),Ct(i,{result:e,error:null,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j}),P(0);return}if(n.status===`error`){T(n.error??`Unknown error`),x(null),Ct(i,{result:null,error:n.error??`Unknown error`,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j}),P(0);return}}catch(e){if(e.name===`AbortError`)return}}P(0),x(null),Ct(i,{result:null,error:`Stress test timed out after 3 minutes`,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j}),T(`Stress test timed out after 3 minutes`)}(0,A.useEffect)(()=>{ee.current={result:S,error:w,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:b,routeParams:E,includeCsrf:O,sendAsFormData:j,key:i}}),(0,A.useEffect)(()=>{let e=Z(i);e?.jobId&&!e?.result&&(r(n),F.current=new AbortController,te(e.jobId,F.current.signal).finally(()=>{y(!1),P(0),r(null)}))},[]),(0,A.useEffect)(()=>()=>{F.current?.abort();let e=ee.current;Ct(e.key,{result:e.result,error:e.error,count:e.count,concurrency:e.concurrency,headersRaw:e.headersRaw,body:e.body,timeout:e.timeout,jobId:e.jobId,routeParams:e.routeParams,includeCsrf:e.includeCsrf,sendAsFormData:e.sendAsFormData})},[]);let I=Z(i),ne=I?.savedAt&&I.result?xt(I.savedAt):null;function L(e){let t={};for(let n of e.split(`
-`)){let e=n.indexOf(`:`);if(e>0){let r=n.slice(0,e).trim(),i=n.slice(e+1).trim();r&&(t[r]=i)}}return t}async function re(){let s=a.filter(e=>!e.optional&&!E[e.name]?.trim());if(s.length>0){T(`Required route param${s.length>1?`s`:``} missing: ${s.map(e=>e.name).join(`, `)}`);return}y(!0),C(null),T(null),r(n);let l=Tt(t,E),d=o.replace(/\/$/,``)+`/`+l.replace(/^\//,``);F.current=new AbortController;let p=F.current.signal,h={},_=m||null;if(vt.has(e.toUpperCase())&&j&&m){let e=Et(m);e!==null&&(_=e,h[`Content-Type`]=`application/x-www-form-urlencoded`)}let v={...h,...L(f)};try{let t=await fetch(`/_laravel-brain/api/stress-test`,{method:`POST`,signal:p,headers:{"Content-Type":`application/json`,Accept:`application/json`},body:JSON.stringify({method:e.toUpperCase(),url:d,count:c,concurrency:u,headers:v,body:_,timeout:g,includeCsrf:yt.has(e.toUpperCase())?O:!1})}),n=await t.json();if(!t.ok){T(n.error??`Request failed (${t.status})`);return}if(n.jobId){x(n.jobId),Ct(i,{result:null,error:null,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:n.jobId,routeParams:E,includeCsrf:O,sendAsFormData:j}),await te(n.jobId,p);return}let r=n;C(r),x(null),Ct(i,{result:r,error:null,count:c,concurrency:u,headersRaw:f,body:m,timeout:g,jobId:null,routeParams:E,includeCsrf:O,sendAsFormData:j})}catch(e){e.name!==`AbortError`&&T(e instanceof Error?e.message:`Network error`)}finally{y(!1),P(0),r(null)}}let ie=S?[{label:`Min`,value:`${S.timing.min}ms`},{label:`Avg`,value:`${S.timing.avg}ms`},{label:`P50`,value:`${S.timing.p50}ms`},{label:`P95`,value:`${S.timing.p95}ms`},{label:`P99`,value:`${S.timing.p99}ms`},{label:`Max`,value:`${S.timing.max}ms`},{label:`Req/s`,value:String(S.throughput)},{label:`Success`,value:`${S.successRate}%`},{label:`Wall`,value:`${S.wallTimeMs}ms`}]:[];return(0,X.jsxs)(`div`,{className:`st-section sidebar-section`,children:[(0,X.jsx)(`div`,{className:`st-toggle`,children:(0,X.jsx)(`h3`,{children:`Stress Test`})}),(0,X.jsx)(`div`,{className:`st-body`,children:(0,X.jsxs)(`div`,{className:`st-form`,children:[(0,X.jsxs)(`div`,{className:`st-form-row`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`Base URL`}),(0,X.jsx)(`input`,{className:`st-input`,type:`text`,placeholder:`http://localhost:8000`,value:o,onChange:e=>s(e.target.value)})]}),(0,X.jsxs)(`div`,{className:`st-docker-hint`,children:[(0,X.jsx)(`strong`,{children:`Docker?`}),` The stress test runs `,(0,X.jsx)(`em`,{children:`inside`}),` the container — `,(0,X.jsx)(`code`,{children:`localhost:8080`}),` is the host-side port and won't be reachable there. Change Base URL to the internal service address, e.g. `,(0,X.jsx)(`code`,{children:`http://nginx`}),` or `,(0,X.jsx)(`code`,{children:`http://localhost:80`}),`.`]}),(0,X.jsxs)(`div`,{className:`st-form-row`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`Target`}),(0,X.jsxs)(`span`,{className:`st-uri-preview`,children:[(0,X.jsx)(`span`,{className:`st-method-badge`,children:e.toUpperCase()}),a.length>0?Tt(t,E):t]})]}),a.length>0&&(0,X.jsxs)(`div`,{className:`st-form-col`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`Route Params`}),a.map(({name:e,optional:t})=>(0,X.jsxs)(`div`,{className:`st-form-row`,style:{marginTop:4},children:[(0,X.jsxs)(`span`,{className:`st-label`,style:{minWidth:80},children:[e,t?` (opt)`:``]}),(0,X.jsx)(`input`,{className:`st-input`,type:`text`,placeholder:t?`optional`:`required`,value:E[e]??``,onChange:t=>D(n=>({...n,[e]:t.target.value}))})]},e))]}),(0,X.jsxs)(`div`,{className:`st-form-row`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`Requests`}),(0,X.jsx)(`input`,{className:`st-input st-input--short`,type:`number`,min:1,max:200,value:c,onChange:e=>l(Math.max(1,Math.min(200,parseInt(e.target.value)||1)))}),(0,X.jsx)(`span`,{className:`st-label`,style:{minWidth:`auto`,marginLeft:8},children:`Concurrency`}),(0,X.jsx)(`input`,{className:`st-input st-input--short`,type:`number`,min:1,max:20,value:u,onChange:e=>d(Math.max(1,Math.min(20,parseInt(e.target.value)||1)))})]}),(0,X.jsxs)(`div`,{className:`st-form-row`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`Timeout (s)`}),(0,X.jsx)(`input`,{className:`st-input st-input--short`,type:`number`,min:1,max:30,value:g,onChange:e=>_(Math.max(1,Math.min(30,parseInt(e.target.value)||10)))})]}),yt.has(e.toUpperCase())&&(0,X.jsxs)(`div`,{className:`st-form-row`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`CSRF Token`}),(0,X.jsxs)(`label`,{style:{display:`flex`,alignItems:`center`,gap:6,cursor:`pointer`,fontSize:13},children:[(0,X.jsx)(`input`,{type:`checkbox`,checked:O,onChange:e=>k(e.target.checked)}),`Auto-inject from session`]})]}),vt.has(e.toUpperCase())&&(0,X.jsxs)(`div`,{className:`st-form-row`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`Body Format`}),(0,X.jsxs)(`label`,{style:{display:`flex`,alignItems:`center`,gap:6,cursor:`pointer`,fontSize:13},children:[(0,X.jsx)(`input`,{type:`checkbox`,checked:j,onChange:e=>M(e.target.checked)}),`Form data (application/x-www-form-urlencoded)`]})]}),(0,X.jsxs)(`div`,{className:`st-form-col`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:`Headers`}),(0,X.jsx)(`textarea`,{className:`st-textarea`,rows:3,placeholder:`Authorization: Bearer token
-Accept: application/json`,value:f,onChange:e=>p(e.target.value)})]}),vt.has(e.toUpperCase())&&(0,X.jsxs)(`div`,{className:`st-form-col`,children:[(0,X.jsx)(`span`,{className:`st-label`,children:j?`Body (JSON → form)`:`Body (JSON)`}),(0,X.jsx)(`textarea`,{className:`st-textarea`,rows:4,placeholder:`{}`,value:m,onChange:e=>h(e.target.value)})]}),(0,X.jsx)(`button`,{className:`st-run-btn`,onClick:re,disabled:v,children:v?`⏳ Running… ${N>0?`(${N}s)`:``}`:`▶ Run Stress Test`}),ne&&(0,X.jsxs)(`span`,{className:`st-last-run st-last-run--form`,children:[`Last run: `,ne]}),w&&(0,X.jsx)(`div`,{className:`st-error-box`,children:w}),S&&(0,X.jsxs)(`div`,{className:`st-results`,children:[(0,X.jsx)(`div`,{className:`st-metrics-grid`,children:ie.map(e=>(0,X.jsxs)(`div`,{className:`st-metric`,children:[(0,X.jsx)(`div`,{className:`st-metric-value`,children:e.value}),(0,X.jsx)(`div`,{className:`st-metric-label`,children:e.label})]},e.label))}),Object.keys(S.statusDistribution).length>0&&(0,X.jsxs)(`div`,{className:`st-dist`,children:[(0,X.jsx)(`div`,{className:`st-dist-title`,children:`Status Distribution`}),Object.entries(S.statusDistribution).sort(([e],[t])=>parseInt(e)-parseInt(t)).map(([e,t])=>(0,X.jsxs)(`div`,{className:`st-dist-row`,children:[(0,X.jsx)(`span`,{className:`st-dist-label`,children:e===`0`?`err`:e}),(0,X.jsx)(`div`,{className:`st-dist-bar-wrap`,children:(0,X.jsx)(`div`,{className:`st-dist-bar`,style:{width:`${t/S.total*100}%`,background:bt(e)}})}),(0,X.jsx)(`span`,{className:`st-dist-count`,children:t})]},e))]}),S.errors.length>0&&(0,X.jsx)(`div`,{className:`st-error-box`,style:{marginTop:8},children:S.errors.map((e,t)=>(0,X.jsx)(`div`,{children:e},t))})]})]})})]})}var Ot=[`route`,`middleware`,`controller`,`action`,`validation_request`,`action_class`,`service`,`model`,`event`,`job`,`command`,`channel`,`schedule`,`view`,`mail`,`notification`,`enum`,`interface`,`trait`,`abstract_class`,`service_provider`,`ai_agent`,`ai_tool`];function kt(e){return e===`action`?`controller`:e}function At(e){if(!e)return 99;let t=kt(e.type),n=Ot.indexOf(t);return n===-1?99:n}function jt(e){let t=e.split(`\\`),n=t[t.length-1];return n.length<=20?n:n.substring(0,18)+`…`}function Mt(e){let t=new Map;for(let n of e)t.has(n.source)||t.set(n.source,[]),t.get(n.source).push(n);return t}function Nt(e){return e.includes(`-to-job`)||e.includes(`-to-event`)||e===`model-to-event`}function Pt(e,t){let n=new Map(t.nodes.map(e=>[e.id,e])),r=Mt(t.edges),i=new Set,a=[],o=[],s=[e];for(i.add(e);s.length>0;){let e=s.shift();a.push(e);for(let t of r.get(e)??[])o.push(t),i.has(t.target)||(i.add(t.target),s.push(t.target))}let c=[],l=new Map,u=[...a].sort((e,t)=>{let r=At(n.get(e)),i=At(n.get(t));return r===i?e.localeCompare(t):r-i});for(let e of u){let t=n.get(e);if(!t)continue;let r=c.length;l.set(e,r);let i=kt(t.type);c.push({id:t.id,label:jt(t.label),type:i,color:L[t.type]??L[i]??`#888`})}c.unshift({id:`__client__`,label:`Client`,type:`client`,color:`#78909C`});for(let e of[...l.keys()])l.set(e,l.get(e)+1);let d=u.filter(e=>n.get(e)?.type===`model`),f=null;d.length>0&&(f=c.length,c.push({id:`__db__`,label:`Database`,type:`db`,color:`#78909C`}));let p=[],m=l.get(e);m!==void 0&&p.push({fromIndex:0,toIndex:m,label:`request`,isReturn:!1});for(let e of o){let t=l.get(e.source),n=l.get(e.target);if(t===void 0||n===void 0||t===n)continue;let r=Nt(e.type);p.push({fromIndex:t,toIndex:n,label:e.label||``,isAsync:r})}if(f!==null)for(let e of d){let t=l.get(e);t!==void 0&&(p.push({fromIndex:t,toIndex:f,label:`query`,isReturn:!1}),p.push({fromIndex:f,toIndex:t,label:`result`,isReturn:!0}))}m!==void 0&&p.push({fromIndex:m,toIndex:0,label:`response`,isReturn:!0});let h=new Map,g=[];for(let e of p){let t=`${e.fromIndex}|${e.toIndex}|${e.label}|${e.isReturn?`r`:``}|${e.isAsync?`a`:``}`,n=h.get(t);if(n){n.count++;let t=e.label;g[n.idx]={...g[n.idx],label:`${t} ×${n.count}`}}else h.set(t,{idx:g.length,count:1}),g.push(e)}return{actors:c,messages:g}}function Ft(e,t){let n=[`%% Sequence Diagram — ${t}`,`sequenceDiagram`,`  autonumber`];for(let t=0;t<e.actors.length;t++){let r=e.actors[t],i=`P${t}`,a=r.label.replace(/"/g,`'`);n.push(`  participant ${i} as "${a}"`)}n.push(``);for(let t of e.messages){let e=`P${t.fromIndex}`,r=`P${t.toIndex}`,i=t.label.replace(/"/g,`'`),a;a=t.isAsync?`->>`:t.isReturn?`-->>`:`->>`,n.push(`  ${e}${a}${r}: ${i}`)}return n.join(`
-`)}var It=110,Q=52,Lt=38,Rt=16;function zt({diagram:e,title:t,theme:n=`dark`}){let[r,i]=(0,A.useState)(!1),[a,o]=(0,A.useState)(!1),s=(0,A.useRef)(null),c=n===`dark`,l=Rt*2+e.actors.length*It,u=Q+e.messages.length*Lt+Lt+Q,d=e=>Rt+e*It+It/2,f=e=>Q+e*Lt+Lt/2,p=c?`#e0e0e0`:`#1a1a1a`,m=c?`#888`:`#999`,h=c?`rgba(255,255,255,0.10)`:`rgba(0,0,0,0.12)`,g=c?`#0d0f14`:`#ffffff`,_=c?`rgba(255,255,255,0.35)`:`rgba(0,0,0,0.30)`,v=c?`seq-arrow-dark`:`seq-arrow-light`,y=c?`seq-arrow-return-dark`:`seq-arrow-return-light`,b=c?`seq-arrow-async-dark`:`seq-arrow-async-light`,x=c?`#a0aec0`:`#555`,S=c?`#b39ddb`:`#7c4dff`;return e.actors.length===0?(0,X.jsx)(`div`,{className:`flowchart-empty`,children:(0,X.jsx)(`span`,{children:`No sequence data available`})}):(0,X.jsxs)(X.Fragment,{children:[(0,X.jsxs)(`div`,{className:`flowchart-export-bar`,children:[(0,X.jsxs)(`button`,{className:`flowchart-export-btn`,onClick:async()=>{if(s.current){o(!0);try{Ze(await $e(s.current,g),`${(t??`sequence`).replace(/[^a-z0-9]/gi,`_`)}_sequence.png`)}finally{o(!1)}}},disabled:a,title:`Export as PNG`,children:[a?`⏳`:`🖼`,` PNG`]}),(0,X.jsx)(`button`,{className:`flowchart-export-btn`,onClick:()=>i(!0),title:`Export as Mermaid`,children:`🧜 Mermaid`})]}),(0,X.jsx)(`div`,{className:`seq-diagram-root`,ref:s,children:(0,X.jsxs)(`svg`,{className:`seq-diagram-svg`,viewBox:`0 0 ${l} ${u}`,width:`100%`,style:{background:g,display:`block`},xmlns:`http://www.w3.org/2000/svg`,children:[(0,X.jsxs)(`defs`,{children:[(0,X.jsx)(`marker`,{id:v,markerWidth:`8`,markerHeight:`6`,refX:`7`,refY:`3`,orient:`auto`,children:(0,X.jsx)(`polygon`,{points:`0 0, 8 3, 0 6`,fill:x})}),(0,X.jsx)(`marker`,{id:y,markerWidth:`8`,markerHeight:`6`,refX:`7`,refY:`3`,orient:`auto`,children:(0,X.jsx)(`polyline`,{points:`0 0, 8 3, 0 6`,fill:`none`,stroke:_,strokeWidth:`1.5`})}),(0,X.jsx)(`marker`,{id:b,markerWidth:`8`,markerHeight:`6`,refX:`7`,refY:`3`,orient:`auto`,children:(0,X.jsx)(`polygon`,{points:`0 0, 8 3, 0 6`,fill:S})})]}),e.actors.map((e,t)=>{let n=d(t),r=It-8,i=n-r/2,a=Math.floor(r/6.5),o=e.label.length>a?e.label.substring(0,a-1)+`…`:e.label;return(0,X.jsxs)(`g`,{children:[(0,X.jsx)(`rect`,{x:i,y:4,width:r,height:Q-10,rx:5,fill:c?`#1a1d24`:`#f5f5f5`,stroke:e.color,strokeWidth:1.5}),(0,X.jsx)(`text`,{x:n,y:Q/2-4,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:10.5,fontFamily:`system-ui, sans-serif`,fill:e.color,fontWeight:`600`,children:o}),(0,X.jsx)(`text`,{x:n,y:Q-12,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:8,fontFamily:`system-ui, sans-serif`,fill:m,children:e.type})]},e.id)}),e.actors.map((e,t)=>(0,X.jsx)(`line`,{x1:d(t),y1:Q,x2:d(t),y2:u-Q,stroke:h,strokeWidth:1,strokeDasharray:`4 4`},`life-${e.id}`)),e.actors.map((e,t)=>{let n=d(t),r=It-8,i=n-r/2,a=u-Q+4,o=Math.floor(r/6.5),s=e.label.length>o?e.label.substring(0,o-1)+`…`:e.label;return(0,X.jsxs)(`g`,{children:[(0,X.jsx)(`rect`,{x:i,y:a,width:r,height:Q-10,rx:5,fill:c?`#1a1d24`:`#f5f5f5`,stroke:e.color,strokeWidth:1.5}),(0,X.jsx)(`text`,{x:n,y:a+Q/2-8,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:10.5,fontFamily:`system-ui, sans-serif`,fill:e.color,fontWeight:`600`,children:s}),(0,X.jsx)(`text`,{x:n,y:a+Q-18,textAnchor:`middle`,dominantBaseline:`middle`,fontSize:8,fontFamily:`system-ui, sans-serif`,fill:m,children:e.type})]},`bottom-${e.id}`)}),e.messages.map((e,t)=>{let n=f(t),r=d(e.fromIndex),i=d(e.toIndex),a=i>r,o=a?r+6:r-6,s=a?i-6:i+6,c=e.isReturn===!0,l=e.isAsync===!0,u=c?_:l?S:x,h=c?`5 3`:l?`6 3`:void 0,g=c?y:l?b:v,C=(r+i)/2,w=Math.abs(i-r)-12,T=Math.max(10,Math.floor(w/6)),E=e.label.length>T?e.label.substring(0,T-1)+`…`:e.label;return(0,X.jsxs)(`g`,{children:[(0,X.jsx)(`line`,{x1:o,y1:n,x2:s,y2:n,stroke:u,strokeWidth:c?1:1.5,strokeDasharray:h,markerEnd:`url(#${g})`}),e.label&&(0,X.jsx)(`text`,{x:C,y:n-6,textAnchor:`middle`,fontSize:9,fontFamily:`system-ui, sans-serif`,fill:c?m:p,opacity:c?.75:1,children:E})]},t)})]})}),r&&(0,X.jsx)(it,{mermaidCode:Ft(e,t??`sequence`),filename:`${(t??`sequence`).replace(/[^a-z0-9]/gi,`_`)}_sequence.mmd`,title:t??`Sequence Diagram`,onClose:()=>i(!1)})]})}function Bt({diagram:e,title:t,theme:n,onClose:r}){return(0,A.useEffect)(()=>{let e=e=>{e.key===`Escape`&&r()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[r]),(0,X.jsx)(`div`,{className:`modal-overlay`,onClick:e=>{e.target===e.currentTarget&&r()},children:(0,X.jsxs)(`div`,{className:`modal-container modal-container--large`,children:[(0,X.jsxs)(`div`,{className:`modal-header`,children:[(0,X.jsxs)(`div`,{className:`modal-title`,children:[(0,X.jsx)(`span`,{className:`modal-icon`,children:`⇄`}),(0,X.jsxs)(`div`,{children:[(0,X.jsx)(`h2`,{children:t}),(0,X.jsx)(`span`,{className:`modal-sub`,children:`Sequence Diagram`})]})]}),(0,X.jsx)(`button`,{className:`modal-close`,onClick:r,title:`Close (Esc)`,children:`×`})]}),(0,X.jsx)(`div`,{className:`modal-body sequence-modal-body`,children:(0,X.jsx)(zt,{diagram:e,title:t,theme:n,compact:!1})})]})})}function $({content:e,children:t,placement:a=`top`,disabled:o=!1,className:s}){let[u,p]=(0,A.useState)(!1),{refs:v,floatingStyles:y,context:x}=m({open:o?!1:u,onOpenChange:p,placement:a,middleware:[_(8),l(),i({padding:8})],whileElementsMounted:b}),{getReferenceProps:S,getFloatingProps:C}=f([n(x,{move:!1,enabled:!o,delay:{open:280,close:80}}),c(x,{enabled:!o}),d(x),r(x,{role:`tooltip`})]),w=h([v.setReference]);return(0,A.isValidElement)(t)?(0,X.jsxs)(X.Fragment,{children:[(0,A.cloneElement)(t,{ref:w,...S()}),u&&!o&&(0,X.jsx)(g,{children:(0,X.jsx)(`div`,{ref:v.setFloating,style:y,className:[`floating-tooltip`,s].filter(Boolean).join(` `),...C(),children:e})})]}):(0,X.jsx)(X.Fragment,{children:t})}var Vt=360,Ht=640,Ut=380,Wt={macro:`#F59E0B`,macro_group:`#B45309`,route:`#4CAF50`,middleware:`#FF9800`,controller:`#2196F3`,action:`#03A9F4`,action_class:`#84cc16`,service:`#9C27B0`,validation_request:`#0d9488`,model:`#F44336`,event:`#FFD600`,job:`#607D8B`,command:`#14b8a6`,channel:`#8b5cf6`,schedule:`#f97316`,view:`#ec4899`,mail:`#f472b6`,notification:`#db2777`,enum:`#0ea5e9`,interface:`#38bdf8`,trait:`#a78bfa`,abstract_class:`#94a3b8`,service_provider:`#ca8a04`,facade:`#00BCD4`,ai_agent:`#A3E635`,ai_tool:`#65A30D`,filament_panel:`#7C3AED`,filament_resource:`#A855F7`,filament_page:`#C084FC`,filament_page_method:`#E879F9`,filament_widget:`#06B6D4`,filament_relation_manager:`#0891B2`};function Gt(e){if(e===null)return`—`;if(e<1024)return`${e} B`;let t=[`KB`,`MB`,`GB`,`TB`],n=e/1024,r=0;for(;n>=1024&&r<t.length-1;)n/=1024,r++;return`${n<10?n.toFixed(1):Math.round(n)} ${t[r]}`}function Kt(e,t){return e===null?`—`:`${t?`~`:``}${e.toLocaleString(`en-US`)}`}var qt={read:"Reads a cached value — this data can be stale. `remember` counts as a read: it writes only on a miss.",write:`Writes a value into the cache.`,invalidate:"Clears cached data. `pull` is here too: it reads the key and removes it.",lock:`Takes an atomic lock through the cache store.`},Jt=`#A3E635`;function Yt(e){let t=typeof e.model==`string`?e.model:null;if(t)return t;let n=typeof e.modelSource==`string`?e.modelSource:``,r=typeof e.modelTier==`string`?e.modelTier:``;return n===`tier`?`${r} tier — provider decides`:n===`method`?`decided at runtime by model()`:`provider's default model`}var Xt=[{key:`provider`,label:`provider`},{key:`maxSteps`,label:`max steps`},{key:`maxTokens`,label:`max tokens`},{key:`temperature`,label:`temperature`},{key:`topP`,label:`top p`},{key:`timeout`,label:`timeout`}],Zt=[{key:`strict`,label:`strict`},{key:`repairToolCalls`,label:`repairs tool calls`},{key:`withoutBroadcasting`,label:`no broadcasting`}];function Qt({selectedId:e,graphData:t,theme:n,onClose:r,onStressChange:i}){let[a,o]=(0,A.useState)(Ut),s=(0,A.useRef)(!1),c=(0,A.useRef)(0),l=(0,A.useRef)(Ut),u=(0,A.useCallback)(e=>{e.preventDefault(),s.current=!0,c.current=e.clientX,l.current=a;let t=e=>{if(!s.current)return;let t=c.current-e.clientX;o(Math.min(Ht,Math.max(Vt,l.current+t)))},n=()=>{s.current=!1,window.removeEventListener(`mousemove`,t),window.removeEventListener(`mouseup`,n),document.body.style.cursor=``,document.body.style.userSelect=``};document.body.style.cursor=`col-resize`,document.body.style.userSelect=`none`,window.addEventListener(`mousemove`,t),window.addEventListener(`mouseup`,n)},[a]),[d,f]=(0,A.useState)(`info`),[p,m]=(0,A.useState)(!1),[h,g]=(0,A.useState)(!1),[_,v]=(0,A.useState)(!1),[y,b]=(0,A.useState)(!1),[x,S]=(0,A.useState)(!1),[C,w]=(0,A.useState)(e);e!==C&&(w(e),f(`info`),m(!1),g(!1),v(!1),b(!1),S(!1));let T=(0,A.useMemo)(()=>{let e=new Map;return t&&t.nodes.forEach(t=>e.set(t.id,t)),e},[t]),E=(0,A.useMemo)(()=>{let e=new Map;return t&&t.edges.forEach(t=>{let n=e.get(t.target)??[];n.push(t),e.set(t.target,n)}),e},[t]),D=(0,A.useMemo)(()=>{let e=new Map;return t&&t.edges.forEach(t=>{let n=e.get(t.source)??[];n.push(t),e.set(t.source,n)}),e},[t]),O=(0,A.useMemo)(()=>!t||!e||t.nodes.find(t=>t.id===e)?.type!==`route`?null:Pt(e,t),[e,t]),k=(0,A.useCallback)(async()=>{if(e){S(!0);try{let t=await fetch(`/_laravel-brain/api/context?nodeId=${encodeURIComponent(e)}&budget=6000`);if(!t.ok)throw Error(`Failed to fetch context`);let n=await t.text();await navigator.clipboard.writeText(n),b(!0),setTimeout(()=>b(!1),2500)}catch{alert(`Could not copy AI context.`)}finally{S(!1)}}},[e]);if(!t)return null;if(!e)return(0,X.jsxs)(`div`,{className:`sidebar-resizable`,style:{width:a},children:[(0,X.jsx)($,{content:`Drag to resize`,children:(0,X.jsx)(`div`,{className:`sidebar-drag-handle`,onMouseDown:u})}),(0,X.jsxs)(`div`,{className:`sidebar`,children:[(0,X.jsxs)(`div`,{className:`sidebar-header`,children:[(0,X.jsx)(`h2`,{children:t.meta.project}),(0,X.jsx)(`span`,{className:`sidebar-subtitle`,children:`Laravel Lifecycle Graph`})]}),(0,X.jsxs)(`div`,{className:`sidebar-stats`,children:[(0,X.jsx)($,{content:`Total symbols in this tab's JSON graph (routes, classes, views, …).`,children:(0,X.jsxs)(`div`,{className:`stat`,children:[(0,X.jsx)(`span`,{className:`stat-value`,children:t.meta.nodeCount}),(0,X.jsx)(`span`,{className:`stat-label`,children:`Nodes`})]})}),(0,X.jsx)($,{content:`Directed links between nodes: calls, type-hints, events, views, Eloquent relations, etc.`,children:(0,X.jsxs)(`div`,{className:`stat`,children:[(0,X.jsx)(`span`,{className:`stat-value`,children:t.meta.edgeCount}),(0,X.jsx)(`span`,{className:`stat-label`,children:`Edges`})]})}),(0,X.jsx)($,{content:`HTTP route entry nodes only (subset of all node types).`,children:(0,X.jsxs)(`div`,{className:`stat`,children:[(0,X.jsx)(`span`,{className:`stat-value`,children:t.nodes.filter(e=>e.type===`route`).length}),(0,X.jsx)(`span`,{className:`stat-label`,children:`Routes`})]})})]}),(0,X.jsx)($,{content:`The inspector shows details for the selected node: metrics, flow, source, and incoming/outgoing edges.`,children:(0,X.jsx)(`p`,{className:`sidebar-hint`,children:`Click any node to inspect it`})})]})]});let j=T.get(e);if(!j)return null;let M=E.get(e)??[],N=D.get(e)??[],P=j.data?.flowSteps??[],F=j.data?.file||null,ee=j.data?.line||void 0,te=Wt[j.type]??`#999`,I=j.data?.metrics,ne=!!j.data?.fatMethod,L=!!j.data?.fatClass,re=!!j.data?.hasN1,ie=j.data?.dbQueries??[],R=j.data?.cacheOps??[],ae=j.data?.httpCalls??[],le=j.data?.relationships??[],ue=j.type===`middleware`&&typeof j.data?.params==`string`&&j.data.params?j.data.params.split(`,`).map(e=>e.trim()).filter(Boolean):[],de=j.data?.members??[],fe=j.data?.validationRules??[],pe=Object.entries(j.data??{}).filter(([e,t])=>e!==`flowSteps`&&e!==`metrics`&&e!==`fatMethod`&&e!==`fatClass`&&e!==`hasN1`&&e!==`classMetrics`&&e!==`dbQueries`&&e!==`cacheOps`&&e!==`httpCalls`&&e!==`relationships`&&e!==`params`&&e!==`members`&&e!==`validationRules`&&e!==`security`&&e!==`erd`&&e!==`tableStats`&&e!==`schema`&&e!==`event`&&e!==`listener`&&e!==`job`&&e!==`origin`&&e!==`mixin`&&e!==`registrar`&&!(Array.isArray(t)&&t.length===0)),V=j.data?.erd,H=j.data?.tableStats,U=j.data?.schema,W=j.data?.event,G=j.data?.listener,K=j.data?.job,me=P.length>0||!!O,he=!!F,ge=M.length>0||N.length>0,q=j.type===`route`,J=j.data?.security?j.data.security:null,Y=d===`flow`&&!me||d===`source`&&!he||d===`edges`&&!ge||d===`stress`&&!q||d===`schema`&&!U||d===`risks`&&!q&&!J?`info`:d,_e=J?J.issues.length:0,ve=n===`light`?oe:z,ye=[{id:`info`,label:`Info`,title:`Identity, type, smells, and code metrics (lines, cyclomatic complexity, …).`},...q||_e>0?[{id:`risks`,label:`Risks`,count:_e||void 0,alert:_e>0,title:`Findings that need attention: a route’s exposure and rate-limiting, or a table’s missing indexes.`}]:[],...U?[{id:`schema`,label:`Schema`,count:U.columns.length||void 0,title:`Columns, indexes and foreign keys as the database itself reports them.`}]:[],...me?[{id:`flow`,label:`Flow`,title:`Control-flow steps through this method or request (and sequence diagram for routes).`}]:[],...ge?[{id:`edges`,label:`Edges`,count:M.length+N.length,title:`What calls or references this node (incoming) and what it calls (outgoing).`}]:[],{id:`usages`,label:`Usages`,title:`Where this symbol is referenced across the whole project, grouped by file.`},...he?[{id:`source`,label:`Source`,title:`Syntax-highlighted PHP source around this symbol.`}]:[],...q?[{id:`stress`,label:`Stress`,title:`Send HTTP requests against this route and inspect responses (dev only).`}]:[]];return(0,X.jsxs)(`div`,{className:`sidebar-resizable`,style:{width:a},children:[(0,X.jsx)($,{content:`Drag to resize`,children:(0,X.jsx)(`div`,{className:`sidebar-drag-handle`,onMouseDown:u})}),(0,X.jsxs)(`div`,{className:`sidebar`,children:[(0,X.jsxs)(`div`,{className:`sidebar-header`,children:[(0,X.jsxs)(`div`,{className:`sidebar-header-actions`,children:[(0,X.jsx)($,{content:`Copy AI context to clipboard`,children:(0,X.jsx)(`span`,{className:`tooltip-trigger-wrap`,children:(0,X.jsx)(`button`,{type:`button`,className:`flow-popup-btn sidebar-ai-btn`,onClick:k,disabled:x,children:x?`…`:y?`✓`:`🤖`})})}),(0,X.jsx)($,{content:`Clear selection (close inspector header)`,children:(0,X.jsx)(`button`,{className:`sidebar-close`,type:`button`,onClick:r,children:`×`})})]}),(0,X.jsxs)(`div`,{className:`sidebar-eyebrow`,children:[(0,X.jsx)(`span`,{className:`sidebar-eyebrow-dot`,style:{backgroundColor:te}}),(0,X.jsx)(`span`,{className:`sidebar-eyebrow-type`,children:j.type.replace(/_/g,` `)})]}),(0,X.jsx)(`h2`,{className:`sidebar-node-title`,children:j.label}),(0,X.jsxs)(`div`,{className:`sidebar-chips`,children:[J&&ve[J.exposure]&&(()=>{let e=ve[J.exposure];return(0,X.jsxs)(`span`,{className:`ins-chip`,style:{"--cc":e.accent},children:[`● `,e.label]})})(),J&&J.riskLevel!==`none`&&(0,X.jsxs)(`span`,{className:`ins-chip`,style:{"--cc":B[J.riskLevel]},children:[`⚠ `,se[J.riskLevel],` risk · `,_e]}),ae.length>0&&(0,X.jsx)($,{content:`Leaves the application: ${ae.map(e=>e.host||e.configKey||`computed address`).join(`, `)}`,children:(0,X.jsxs)(`span`,{className:`ins-chip ins-chip--http`,children:[`🌐 `,ae.length,` outgoing`]})}),M.length+N.length>0&&(0,X.jsxs)(`span`,{className:`ins-chip ins-chip--neutral`,children:[`Edges `,M.length+N.length]}),F&&(0,X.jsxs)(`span`,{className:`ins-chip ins-chip--neutral`,title:F,children:[F.split(`/`).slice(-2).join(`/`),ee?` : ${ee}`:``]})]})]}),(ne||L||re)&&(0,X.jsxs)(`div`,{className:`sidebar-smells`,children:[re&&(0,X.jsx)($,{content:`N+1 Query: database query inside a loop`,children:(0,X.jsx)(`span`,{className:`smell-badge smell-badge--n1`,children:`⚠️ N+1 Query`})}),ne&&(0,X.jsx)($,{content:`Fat Method: more than 30 lines or cyclomatic complexity > 10`,children:(0,X.jsx)(`span`,{className:`smell-badge smell-badge--fat-method`,children:`🧱 Fat Method`})}),L&&(0,X.jsx)($,{content:`Fat Class: more than 10 methods or 300+ total lines`,children:(0,X.jsx)(`span`,{className:`smell-badge smell-badge--fat-class`,children:`🏗️ Fat Class`})})]}),(0,X.jsx)(`div`,{className:`sidebar-tab-bar`,children:ye.map(e=>(0,X.jsx)($,{content:e.title,children:(0,X.jsxs)(`button`,{type:`button`,className:`sidebar-tab${Y===e.id?` sidebar-tab--active`:``}`,onClick:()=>f(e.id),children:[e.label,e.count!==void 0&&(0,X.jsx)(`span`,{className:`sidebar-tab-badge${e.alert?` sidebar-tab-badge--alert`:``}`,children:e.count})]})},e.id))}),(0,X.jsxs)(`div`,{className:`sidebar-tab-content`,children:[Y===`info`&&(0,X.jsxs)(X.Fragment,{children:[(0,X.jsxs)(`div`,{className:`ins-actions`,children:[(0,X.jsxs)(`button`,{type:`button`,className:`ins-action-btn`,disabled:!he,onClick:()=>f(`source`),children:[(0,X.jsxs)(`svg`,{className:`ins-action-icon`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,X.jsx)(`path`,{d:`M18 13v6a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2V8a2 2 0 0 1 2-2h6`}),(0,X.jsx)(`polyline`,{points:`15 3 21 3 21 9`}),(0,X.jsx)(`line`,{x1:`10`,y1:`14`,x2:`21`,y2:`3`})]}),`Open file`]}),(0,X.jsxs)(`button`,{type:`button`,className:`ins-action-btn`,onClick:()=>navigator.clipboard.writeText(String(j.data?.uri??j.label)),children:[(0,X.jsxs)(`svg`,{className:`ins-action-icon`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,X.jsx)(`rect`,{x:`9`,y:`9`,width:`13`,height:`13`,rx:`2`,ry:`2`}),(0,X.jsx)(`path`,{d:`M5 15H4a2 2 0 0 1-2-2V4a2 2 0 0 1 2-2h9a2 2 0 0 1 2 2v1`})]}),`Copy URI`]})]}),(()=>{let e=I?.cyclomaticComplexity??0,t=N.length,n={none:0,low:25,medium:55,high:80,critical:100}[J?.riskLevel??`none`]??0;return(0,X.jsx)(`div`,{className:`ins-meters`,children:[{label:`Complexity`,value:e,pct:Math.min(100,e*6),tone:e>15?`var(--danger)`:e>10?`var(--warn)`:`var(--ok)`},{label:`Fan-out`,value:t,pct:Math.min(100,t*10),tone:t>8?`var(--danger)`:t>4?`var(--warn)`:`var(--ok)`},{label:`Risk`,value:_e,pct:n,tone:n>=80?`var(--danger)`:n>=55?`var(--warn)`:`var(--ok)`}].map(e=>(0,X.jsxs)(`div`,{className:`ins-meter`,children:[(0,X.jsx)(`span`,{className:`ins-meter-label`,children:e.label}),(0,X.jsx)(`span`,{className:`ins-meter-track`,children:(0,X.jsx)(`span`,{className:`ins-meter-fill`,style:{width:`${e.pct}%`,background:e.tone}})}),(0,X.jsx)(`span`,{className:`ins-meter-value`,children:e.value})]},e.label))})})(),I&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--metrics`,children:[(0,X.jsx)(`h3`,{children:`Code Metrics`}),(0,X.jsxs)(`div`,{className:`metrics-grid`,children:[(0,X.jsx)($,{content:`Physical lines of code in this method (approximate, from static analysis).`,children:(0,X.jsxs)(`div`,{className:`metric-item`,children:[(0,X.jsx)(`span`,{className:`metric-value`,children:I.lineCount}),(0,X.jsx)(`span`,{className:`metric-label`,children:`Lines`})]})}),(0,X.jsx)($,{content:`Cyclomatic complexity: decision paths (branches, loops, boolean operators). Rough guide: above 10 is harder to test; above 15 is very complex.`,children:(0,X.jsxs)(`div`,{className:`metric-item`,children:[(0,X.jsx)(`span`,{className:`metric-value`,style:{color:I.cyclomaticComplexity>10?`#FF6D00`:`inherit`},children:I.cyclomaticComplexity}),(0,X.jsx)(`span`,{className:`metric-label`,children:`Complexity`})]})}),(0,X.jsx)($,{content:`Executable statements counted in this method body.`,children:(0,X.jsxs)(`div`,{className:`metric-item`,children:[(0,X.jsx)(`span`,{className:`metric-value`,children:I.statementCount}),(0,X.jsx)(`span`,{className:`metric-label`,children:`Statements`})]})}),(0,X.jsx)($,{content:`Parameters on this function or method signature.`,children:(0,X.jsxs)(`div`,{className:`metric-item`,children:[(0,X.jsx)(`span`,{className:`metric-value`,children:I.paramCount}),(0,X.jsx)(`span`,{className:`metric-label`,children:`Params`})]})})]})]}),j.type===`filament_resource`&&!!j.data?.route&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Filament URL`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`route`}),(0,X.jsx)(`span`,{className:`prop-value`,style:{fontFamily:`monospace`,color:`#A855F7`},children:String(j.data.route)})]})]}),j.type===`ai_agent`&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Model & limits`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`model`}),(0,X.jsx)(`span`,{className:`prop-value`,style:{fontFamily:`monospace`,color:Jt},children:Yt(j.data)})]}),Xt.map(({key:e,label:t})=>j.data?.[e]===void 0?null:(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:t}),(0,X.jsx)(`span`,{className:`prop-value`,style:{fontFamily:`monospace`},children:String(j.data[e])})]},e)),Zt.map(({key:e,label:t})=>j.data?.[e]?(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:t}),(0,X.jsx)(`span`,{className:`prop-value`,children:`yes`})]},e):null),Array.isArray(j.data?.methodOverrides)&&j.data.methodOverrides.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`overridable`}),(0,X.jsx)(`span`,{className:`prop-value`,children:j.data.methodOverrides.join(`, `)})]}),typeof j.data?.shadowedModelAttribute==`string`&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,style:{color:`#FF6D00`},children:`dead #[Model]`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[j.data.shadowedModelAttribute,` — a model() method is read instead`]})]}),typeof j.data?.shadowedProviderAttribute==`string`&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,style:{color:`#FF6D00`},children:`dead #[Provider]`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[j.data.shadowedProviderAttribute,` — a provider() method is read instead`]})]}),Array.isArray(j.data?.contracts)&&j.data.contracts.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`contracts`}),(0,X.jsx)(`span`,{className:`prop-value`,children:j.data.contracts.join(`, `)})]}),j.data?.toolsDecidedAtRuntime===!0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`tools()`}),(0,X.jsx)(`span`,{className:`prop-value`,children:`decided at runtime — this agent has tools Brain cannot name from tools()`})]}),Array.isArray(j.data?.injectedTools)&&j.data.injectedTools.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`supplied tools`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[j.data.injectedTools.length,` handed to the constructor where the agent is built`]})]}),Array.isArray(j.data?.unwiredTools)&&j.data.unwiredTools.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,style:{color:`#FF6D00`},children:`unwired tools`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[j.data.unwiredTools.map(e=>e.split(`\\`).pop()).join(`, `),` — tools() is never called without the HasTools contract`]})]}),Array.isArray(j.data?.unresolvedTools)&&j.data.unresolvedTools.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`unresolved tools`}),(0,X.jsx)(`span`,{className:`prop-value`,children:j.data.unresolvedTools.join(`, `)})]})]}),j.type===`ai_tool`&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Tool`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`kind`}),(0,X.jsx)(`span`,{className:`prop-value`,children:j.data?.toolKind===`mcp`?`MCP server tool`:`laravel/ai tool`})]}),typeof j.data?.description==`string`&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`description`}),(0,X.jsx)(`span`,{className:`prop-value`,children:j.data.description})]})]}),le.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Relationships`}),le.map((e,t)=>(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,style:{color:`#9C27B0`},children:e.type}),(0,X.jsx)(`span`,{className:`prop-value`,children:e.related.split(`\\`).pop()??e.related})]},t))]}),ue.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`ATTRIBUTES`}),ue.map((e,t)=>(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,style:{color:`#FF9800`},children:t+1}),(0,X.jsx)(`span`,{className:`prop-value`,children:e})]},t))]}),fe.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--validation-rules`,children:[(0,X.jsx)(`h3`,{children:`Validation rules`}),(0,X.jsx)(`ul`,{className:`sidebar-structure-list`,children:fe.map((e,t)=>(0,X.jsxs)(`li`,{className:`sidebar-structure-item`,children:[(0,X.jsx)(`span`,{className:`structure-kind`,children:`field`}),(0,X.jsx)(`span`,{className:`structure-name`,children:e.field}),(0,X.jsx)(`span`,{className:`structure-value`,children:e.rules})]},t))})]}),ie.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--queries`,children:[(0,X.jsx)(`h3`,{children:`DB Queries`}),(0,X.jsx)(`div`,{className:`query-list`,children:ie.map((e,t)=>{let n=e.table||(e.model?e.model.split(`\\`).pop():`?`);return(0,X.jsxs)(`div`,{className:`query-item`,children:[(0,X.jsx)(`span`,{className:`query-op query-op--${[`insert`,`update`,`delete`,`statement`].includes(e.operation)?`write`:`read`}`,children:e.operation}),(0,X.jsx)(`span`,{className:`query-table`,title:e.model||void 0,children:n}),e.type===`raw`&&(0,X.jsx)(`span`,{className:`query-badge query-badge--raw`,children:`SQL`})]},t)})})]}),R.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--cache`,children:[(0,X.jsx)(`h3`,{children:`Cache`}),(0,X.jsx)(`div`,{className:`cache-list`,children:R.map((e,t)=>(0,X.jsxs)(`div`,{className:`cache-item`,children:[(0,X.jsxs)(`div`,{className:`cache-item-head`,children:[(0,X.jsx)($,{content:qt[e.kind]??e.kind,children:(0,X.jsx)(`span`,{className:`cache-kind cache-kind--${e.kind}`,children:e.kind})}),(0,X.jsx)(`span`,{className:`cache-method`,children:e.method}),e.keyKind===`computed`?(0,X.jsx)($,{content:`The key is built at runtime, so it cannot be read from the source.`,children:(0,X.jsx)(`span`,{className:`cache-key cache-key--computed`,children:`computed key`})}):e.keyKind===`none`?(0,X.jsx)(`span`,{className:`cache-key cache-key--computed`,children:`whole store`}):(0,X.jsx)(`span`,{className:`cache-key cache-key--${e.keyKind}`,title:e.key,children:e.key})]}),(e.tags.length>0||e.store!==``||e.ttl!==null)&&(0,X.jsxs)(`div`,{className:`cache-item-meta`,children:[e.ttl!==null&&(0,X.jsxs)(`span`,{className:`cache-meta`,children:[`ttl `,e.ttl,`s`]}),e.store!==``&&(0,X.jsxs)(`span`,{className:`cache-meta`,children:[`store `,e.store]}),e.tags.map((e,t)=>(0,X.jsx)(`span`,{className:`cache-meta cache-meta--tag`,children:e},t))]})]},t))})]}),ae.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--http`,children:[(0,X.jsx)(`h3`,{children:`Outgoing HTTP`}),(0,X.jsx)(`div`,{className:`http-list`,children:ae.map((e,t)=>{let n=e.configKey?`config('${e.configKey}')${e.url}`:e.url||`address computed at runtime`;return(0,X.jsxs)(`div`,{className:`http-item`,children:[(0,X.jsxs)(`div`,{className:`http-item-head`,children:[(0,X.jsx)(`span`,{className:`http-method http-method--${(e.method||`unknown`).toLowerCase()}`,children:e.method||`REQUEST`}),(0,X.jsx)(`span`,{className:`http-target`,title:n,children:n})]}),(0,X.jsxs)(`div`,{className:`http-item-meta`,children:[(0,X.jsx)(`span`,{className:`http-badge http-badge--client`,children:e.client}),e.urlSource===`constructed`&&(0,X.jsx)($,{content:`The address starts with this literal and continues with something computed at runtime`,children:(0,X.jsx)(`span`,{className:`http-badge`,children:`partly computed`})}),e.async&&(0,X.jsx)(`span`,{className:`http-badge`,children:`async`}),e.timeout===null?(0,X.jsx)($,{content:`No timeout declared: this request waits as long as the third party takes`,children:(0,X.jsx)(`span`,{className:`http-badge http-badge--absent`,children:`no timeout`})}):(0,X.jsxs)(`span`,{className:`http-badge`,children:[`timeout `,e.timeout,`s`]}),e.retryTimes===null?(0,X.jsx)(`span`,{className:`http-badge http-badge--muted`,children:`no retry`}):(0,X.jsxs)(`span`,{className:`http-badge`,children:[`retry `,e.retryTimes,`×`,e.retrySleep===null?``:` / ${e.retrySleep}ms`]})]})]},t)})})]}),de.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Structure`}),(0,X.jsx)(`ul`,{className:`sidebar-structure-list`,children:de.map((e,t)=>(0,X.jsxs)(`li`,{className:`sidebar-structure-item`,children:[(0,X.jsx)(`span`,{className:`structure-kind`,children:String(e.kind??`item`)}),(0,X.jsx)(`span`,{className:`structure-name`,children:String(e.name??``)}),typeof e.declaringClass==`string`&&e.declaringClass!==``&&(0,X.jsx)(`span`,{className:`structure-decl`,title:`Declared on parent class`,children:e.declaringClass}),e.value!==void 0&&e.value!==null&&(0,X.jsx)(`span`,{className:`structure-value`,children:String(e.value)}),e.static===!0&&(0,X.jsx)(`span`,{className:`structure-flag`,children:`static`}),typeof e.visibility==`string`&&(0,X.jsx)(`span`,{className:`structure-vis`,children:e.visibility})]},t))})]}),H&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Table Data`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`rows`}),(0,X.jsx)(`span`,{className:`prop-value`,children:Kt(H.rows,H.rowsEstimated)})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`table`}),(0,X.jsx)(`span`,{className:`prop-value`,children:Gt(H.tableBytes)})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`indexes`}),(0,X.jsx)(`span`,{className:`prop-value`,children:Gt(H.indexBytes)})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`total`}),(0,X.jsx)(`span`,{className:`prop-value`,children:Gt(H.totalBytes)})]})]}),W&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Event`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`listeners`}),(0,X.jsx)(`span`,{className:`prop-value`,children:W.orphan?`none — firing this does nothing`:`${W.listenerCount}`})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`dispatch`}),(0,X.jsx)(`span`,{className:`prop-value`,children:W.deferred?`after commit (ShouldDispatchAfterCommit)`:`immediate`})]}),W.broadcast&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`broadcast`}),(0,X.jsx)(`span`,{className:`prop-value`,children:`yes`})]}),!W.orphan&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`before commit`}),(0,X.jsx)(`span`,{className:`prop-value`,children:W.observableBeforeCommit?`a listener can act before a surrounding transaction commits`:`no listener runs before the commit`})]}),W.properties?.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`payload`}),(0,X.jsx)(`span`,{className:`prop-value`,children:W.properties.join(`, `)})]})]}),G&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Listener`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`runs`}),(0,X.jsx)(`span`,{className:`prop-value`,children:G.queued?`on a queue`:`in the dispatching request`})]}),G.queued&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`waits for commit`}),(0,X.jsx)(`span`,{className:`prop-value`,children:G.deferred?`yes (queue after_commit)`:`no`})]})]}),K&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Queue behaviour`}),K.tries!==null&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`attempts`}),(0,X.jsx)(`span`,{className:`prop-value`,children:K.tries})]}),K.timeout!==null&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`timeout`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[K.timeout,`s`]})]}),K.backoff!==null&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`backoff`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[K.backoff,`s`]})]}),K.maxExceptions!==null&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`max exceptions`}),(0,X.jsx)(`span`,{className:`prop-value`,children:K.maxExceptions})]}),K.unique&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`unique`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[K.uniqueUntilProcessing?`until it starts processing`:`while it is queued or running`,K.uniqueFor===null?``:` \u00b7 ${K.uniqueFor}s`]})]}),K.batchable&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`batch`}),(0,X.jsx)(`span`,{className:`prop-value`,children:`runs as part of one`})]}),K.afterCommit&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`dispatch`}),(0,X.jsx)(`span`,{className:`prop-value`,children:`after the transaction commits`})]}),K.encrypted&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`payload`}),(0,X.jsx)(`span`,{className:`prop-value`,children:`encrypted`})]}),K.middleware.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`middleware`}),(0,X.jsx)(`span`,{className:`prop-value`,children:K.middleware.join(`, `)})]}),K.dynamic.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`decided at runtime`}),(0,X.jsx)(`span`,{className:`prop-value`,children:K.dynamic.map(e=>`${e}()`).join(`, `)})]})]}),j.type===`macro`&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Where this method comes from`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`adds to`}),(0,X.jsx)(`span`,{className:`prop-value`,children:String(j.data?.fqcn??`—`)})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`registered by`}),(0,X.jsx)(`span`,{className:`prop-value`,children:String(j.data?.registrar??`a file with no class`)})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`how`}),(0,X.jsx)(`span`,{className:`prop-value`,children:j.data?.origin===`mixin`?`mixin of ${String(j.data?.mixin??`—`)}`:`macro(), one method at a time`})]})]}),V&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Model Schema`}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`table`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.table||`—`})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`primary key`}),(0,X.jsxs)(`span`,{className:`prop-value`,children:[V.primaryKey,` (`,V.keyType,`)`]})]}),V.morphAlias&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`morph alias`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.morphAlias})]}),!V.morphAlias&&V.morphAliasMissing&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`morph alias`}),(0,X.jsx)(`span`,{className:`prop-value prop-value--warn`,children:`none — this app enforces a morph map`})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`timestamps`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.timestamps?`yes`:`no`})]}),(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`soft deletes`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.softDeletes?`yes`:`no`})]}),V.fillable?.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`fillable`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.fillable.join(`, `)})]}),V.guarded?.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`guarded`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.guarded.join(`, `)})]}),Object.keys(V.casts??{}).length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`casts`}),(0,X.jsx)(`span`,{className:`prop-value`,children:Object.entries(V.casts).map(([e,t])=>`${e}: ${t}`).join(`, `)})]}),V.dates?.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`dates`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.dates.join(`, `)})]}),V.appends?.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`appends`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.appends.join(`, `)})]}),V.accessors?.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`accessors`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.accessors.join(`, `)})]}),V.relationships?.length>0&&(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`relationships`}),(0,X.jsx)(`span`,{className:`prop-value`,children:V.relationships.map(e=>`${e.type}(${e.related})`).join(`, `)})]})]}),(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsx)(`h3`,{children:`Properties`}),pe.map(([e,t])=>(0,X.jsxs)(`div`,{className:`prop-row`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:e}),(0,X.jsx)(`span`,{className:`prop-value`,children:Array.isArray(t)?t.map(e=>typeof e==`object`&&e?Object.values(e).join(` `):String(e)).join(`, `)||`—`:String(t)||`—`})]},e))]})]}),Y===`flow`&&(0,X.jsxs)(X.Fragment,{children:[P.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--flowchart`,children:[(0,X.jsxs)(`div`,{className:`sidebar-section-header`,children:[(0,X.jsx)(`h3`,{children:`Method Flow`}),(0,X.jsx)(`button`,{className:`flow-popup-btn`,title:`Open in large view`,onClick:()=>m(!0),children:`⤢`})]}),(0,X.jsx)(at,{steps:P,isFatMethod:ne}),p&&(0,X.jsx)(ft,{steps:P,title:j.label,isFatMethod:ne,onClose:()=>m(!1)})]}),O&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--sequence`,children:[(0,X.jsxs)(`div`,{className:`sidebar-section-header`,children:[(0,X.jsx)(`h3`,{children:`Sequence Diagram`}),(0,X.jsx)(`button`,{className:`flow-popup-btn`,title:`Open in large view`,onClick:()=>v(!0),children:`⤢`})]}),(0,X.jsx)(zt,{diagram:O,title:j.label,theme:n}),_&&(0,X.jsx)(Bt,{diagram:O,title:j.label,theme:n,onClose:()=>v(!1)})]})]}),Y===`source`&&F&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--source`,children:[(0,X.jsxs)(`div`,{className:`sidebar-section-header`,children:[(0,X.jsx)(`h3`,{children:`Source Code`}),(0,X.jsx)(`button`,{className:`flow-popup-btn`,title:`Open in large view`,onClick:()=>g(!0),children:`⤢`})]}),(0,X.jsx)(mt,{filePath:F,highlightLine:ee,theme:n}),h&&(0,X.jsx)(ht,{filePath:F,highlightLine:ee,theme:n,onClose:()=>g(!1)})]}),Y===`edges`&&(0,X.jsxs)(X.Fragment,{children:[N.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsxs)(`h3`,{children:[`Outgoing (`,N.length,`)`]}),N.map(e=>{let t=T.get(e.target);return(0,X.jsxs)(`div`,{className:`edge-row`,children:[(0,X.jsx)(`span`,{className:`edge-label`,children:e.label}),(0,X.jsx)(`span`,{className:`edge-target`,children:t?.label??e.target})]},e.id)})]}),M.length>0&&(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsxs)(`h3`,{children:[`Incoming (`,M.length,`)`]}),M.map(e=>(0,X.jsxs)(`div`,{className:`edge-row`,children:[(0,X.jsx)(`span`,{className:`edge-target`,children:T.get(e.source)?.label??e.source}),(0,X.jsx)(`span`,{className:`edge-label`,children:e.label})]},e.id))]})]}),Y===`schema`&&U&&(0,X.jsxs)(`div`,{children:[(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsxs)(`h3`,{children:[`Columns `,(0,X.jsx)(`span`,{className:`section-count`,children:U.columns.length})]}),(0,X.jsx)(`div`,{className:`schema-table`,children:U.columns.map(e=>(0,X.jsxs)(`div`,{className:`schema-row`,children:[(0,X.jsx)(`span`,{className:`schema-name`,children:e.name}),(0,X.jsx)(`span`,{className:`schema-type`,children:e.type}),(0,X.jsxs)(`span`,{className:`schema-flags`,children:[e.autoIncrement&&(0,X.jsx)(`span`,{className:`schema-flag`,children:`auto`}),e.nullable?(0,X.jsx)(`span`,{className:`schema-flag schema-flag--muted`,children:`null`}):(0,X.jsx)(`span`,{className:`schema-flag`,children:`not null`}),e.default!==null&&(0,X.jsxs)(`span`,{className:`schema-flag schema-flag--muted`,children:[`= `,e.default]})]})]},e.name))})]}),(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsxs)(`h3`,{children:[`Indexes `,(0,X.jsx)(`span`,{className:`section-count`,children:U.indexes.length})]}),U.indexes.length===0&&(0,X.jsx)(`div`,{className:`sidebar-empty`,children:`No indexes.`}),(0,X.jsx)(`div`,{className:`schema-table`,children:U.indexes.map(e=>(0,X.jsxs)(`div`,{className:`schema-row`,children:[(0,X.jsx)(`span`,{className:`schema-name`,children:e.columns.join(`, `)}),(0,X.jsx)(`span`,{className:`schema-type`,children:e.name}),(0,X.jsxs)(`span`,{className:`schema-flags`,children:[e.primary&&(0,X.jsx)(`span`,{className:`schema-flag`,children:`primary`}),e.unique&&!e.primary&&(0,X.jsx)(`span`,{className:`schema-flag`,children:`unique`})]})]},e.name))})]}),(0,X.jsxs)(`div`,{className:`sidebar-section`,children:[(0,X.jsxs)(`h3`,{children:[`Foreign keys `,(0,X.jsx)(`span`,{className:`section-count`,children:U.foreignKeys.length})]}),U.foreignKeys.length===0&&(0,X.jsx)(`div`,{className:`sidebar-empty`,children:`No foreign keys.`}),(0,X.jsx)(`div`,{className:`schema-table`,children:U.foreignKeys.map(e=>{let t=U.indexes.some(t=>t.columns.slice(0,e.columns.length).join(`\0`)===e.columns.join(`\0`));return(0,X.jsxs)(`div`,{className:`schema-row${t?``:` schema-row--flagged`}`,children:[(0,X.jsx)(`span`,{className:`schema-name`,children:e.columns.join(`, `)}),(0,X.jsxs)(`span`,{className:`schema-type`,children:[`→ `,e.foreignTable,`.`,e.foreignColumns.join(`, `)]}),(0,X.jsxs)(`span`,{className:`schema-flags`,children:[e.onDelete&&e.onDelete!==`no action`&&(0,X.jsxs)(`span`,{className:`schema-flag schema-flag--muted`,children:[`on delete `,e.onDelete]}),!t&&(0,X.jsx)(`span`,{className:`schema-flag schema-flag--warn`,children:`no index`})]})]},e.name)})})]})]}),Y===`usages`&&e&&(0,X.jsx)(_t,{nodeId:e}),Y===`risks`&&J&&(0,X.jsxs)(`div`,{className:`sidebar-section sidebar-section--security`,children:[ve[J.exposure]&&(()=>{let e=ve[J.exposure],t={public:`This route is publicly accessible — no authentication middleware detected.`,guest:`This route is for unauthenticated users and redirects authenticated ones away.`,authed:`This route requires authentication (auth / sanctum / jwt / passport).`,admin:`This route requires elevated permissions (can:, role:, permission:, ability:, gate:).`};return(0,X.jsxs)(`div`,{className:`security-exposure-card`,style:{borderColor:e.border,background:e.bg+`88`},children:[(0,X.jsx)(`div`,{className:`security-exposure-header`,children:(0,X.jsxs)(`span`,{className:`security-exposure-badge`,style:{color:e.accent},children:[`🔒 `,e.label,` Route`]})}),(0,X.jsx)(`p`,{className:`security-exposure-desc`,children:t[J.exposure]??t.public})]})})(),J.issues.length===0?(0,X.jsxs)(`div`,{className:`security-clean`,children:[(0,X.jsx)(`span`,{style:{color:B.none},children:`✓`}),` Nothing flagged here.`]}):(0,X.jsxs)(X.Fragment,{children:[(0,X.jsxs)(`div`,{className:`security-issues-title`,children:[J.issues.length,` Issue`,J.issues.length===1?``:`s`,` Detected`]}),J.issues.map((e,t)=>{let n=ce[e.type]??{icon:`•`,name:e.type},r=B[e.severity]??B.medium;return(0,X.jsxs)(`div`,{className:`security-issue-card`,style:{borderLeftColor:r},children:[(0,X.jsxs)(`div`,{className:`security-issue-header`,children:[(0,X.jsx)(`span`,{className:`security-issue-icon`,children:n.icon}),(0,X.jsx)(`span`,{className:`security-issue-name`,style:{color:r},children:n.name}),(0,X.jsx)(`span`,{className:`security-issue-severity`,style:{color:r},children:e.severity.toUpperCase()})]}),(0,X.jsx)(`p`,{className:`security-issue-message`,children:e.message}),e.file&&(0,X.jsxs)(`div`,{className:`security-issue-location`,children:[(0,X.jsx)(`span`,{className:`prop-key`,children:`file`}),(0,X.jsxs)(`span`,{className:`prop-val`,title:e.file,children:[`…`,e.file.split(`/`).slice(-2).join(`/`),e.line?`:${e.line}`:``]})]})]},t)})]})]}),Y===`risks`&&q&&!J&&(0,X.jsx)(`div`,{className:`sidebar-section`,children:(0,X.jsxs)(`p`,{style:{opacity:.6,fontSize:13},children:[`Security data not available. Re-run `,(0,X.jsx)(`code`,{children:`brain:scan`}),` to generate it.`]})}),Y===`stress`&&q&&e&&(0,X.jsx)(Dt,{method:String(j.data?.method??`GET`),uri:String(j.data?.uri??`/`),theme:n,selectedId:e,onStressChange:i},e)]})]})]})}var $t=[{id:`claude`,label:`Claude Code`,path:`CLAUDE.md`,icon:`🟠`,description:`Anthropic Claude Code CLI & IDE`},{id:`cursor`,label:`Cursor`,path:`.cursor/rules/laravel-brain.mdc`,icon:`⬛`,description:`Cursor AI editor (MDC format with frontmatter)`},{id:`windsurf`,label:`Windsurf`,path:`.windsurf/rules/laravel-brain.md`,icon:`🌊`,description:`Windsurf by Codeium`},{id:`copilot`,label:`GitHub Copilot`,path:`.github/copilot-instructions.md`,icon:`🐙`,description:`Applied repo-wide automatically`},{id:`junie`,label:`JetBrains Junie`,path:`.junie/guidelines.md`,icon:`🧠`,description:`JetBrains AI assistant`},{id:`aider`,label:`Aider`,path:`CONVENTIONS.md`,icon:`⌨️`,description:`Load with: aider --read CONVENTIONS.md`},{id:`agents`,label:`AGENTS.md`,path:`AGENTS.md`,icon:`🌐`,description:`Universal open standard — 60+ tools`},{id:`codex`,label:`OpenAI Codex`,path:`CODEX.md`,icon:`🟢`,description:`Load with: codex --context CODEX.md`}];function en({onClose:e}){let[t,n]=(0,A.useState)(new Set($t.map(e=>e.id))),[r,i]=(0,A.useState)({}),[a,o]=(0,A.useState)(!1),[s,c]=(0,A.useState)(null),l=(0,A.useCallback)(e=>{n(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n})},[]),u=(0,A.useCallback)(()=>n(new Set($t.map(e=>e.id))),[]),d=(0,A.useCallback)(()=>n(new Set),[]),f=(0,A.useCallback)(async e=>{o(!0),c(null);let n={};t.forEach(e=>{n[e]={status:`generating`}}),i(n);try{let n=await fetch(`/_laravel-brain/api/generate-rules`,{method:`POST`,headers:{"Content-Type":`application/json`},body:JSON.stringify({targets:[...t],force:e})}),r=await n.json();if(n.status===409&&r.existing){c(r.existing);let e={};t.forEach(t=>{e[t]={status:`idle`}}),i(e);return}if(!n.ok){let e=r.error??`Generation failed`,n={};t.forEach(t=>{n[t]={status:`error`,error:e}}),i(n);return}let a={};for(let e of r.results??[])a[e.target]=e.success?{status:`success`,path:e.path}:{status:`error`,error:e.error??`Unknown error`};i(a)}catch{let e={};t.forEach(t=>{e[t]={status:`error`,error:`Network error`}}),i(e)}finally{o(!1)}},[t]),p=(0,A.useCallback)(()=>f(!1),[f]),m=(0,A.useCallback)(()=>f(!0),[f]),h=(0,A.useCallback)(()=>c(null),[]),g=Object.values(r).filter(e=>e.status===`success`).length,_=Object.values(r).filter(e=>e.status===`error`).length,v=g+_>0;return(0,X.jsx)(`div`,{className:`export-overlay`,onClick:t=>{t.target===t.currentTarget&&e()},children:(0,X.jsxs)(`div`,{className:`export-modal ai-rules-modal`,children:[(0,X.jsxs)(`div`,{className:`export-modal-header`,children:[(0,X.jsxs)(`div`,{className:`export-modal-title`,children:[(0,X.jsx)(`span`,{className:`export-modal-icon`,children:`🤖`}),(0,X.jsxs)(`div`,{children:[(0,X.jsx)(`h2`,{children:`Generate AI Rules Files`}),(0,X.jsx)(`div`,{className:`export-modal-sub`,children:`Write context files for AI coding assistants into your project`})]})]}),(0,X.jsx)(`button`,{className:`export-modal-close`,onClick:e,children:`×`})]}),s&&(0,X.jsxs)(`div`,{className:`ai-rules-overwrite-banner`,children:[(0,X.jsx)(`div`,{className:`ai-rules-overwrite-icon`,children:`⚠️`}),(0,X.jsxs)(`div`,{className:`ai-rules-overwrite-body`,children:[(0,X.jsxs)(`strong`,{children:[`The following file`,s.length===1?``:`s`,` already exist`,s.length===1?`s`:``,`:`]}),(0,X.jsx)(`ul`,{className:`ai-rules-overwrite-list`,children:s.map(e=>(0,X.jsx)(`li`,{children:(0,X.jsx)(`code`,{children:e.path})},e.target))}),(0,X.jsxs)(`span`,{children:[`Do you want to overwrite `,s.length===1?`it`:`them`,`?`]})]}),(0,X.jsxs)(`div`,{className:`ai-rules-overwrite-actions`,children:[(0,X.jsx)(`button`,{className:`export-btn export-btn--secondary`,onClick:h,children:`Cancel`}),(0,X.jsx)(`button`,{className:`export-btn export-btn--danger`,onClick:m,children:`Overwrite`})]})]}),(0,X.jsxs)(`div`,{className:`ai-rules-select-bar`,children:[(0,X.jsxs)(`span`,{className:`ai-rules-select-label`,children:[t.size,` of `,$t.length,` selected`]}),(0,X.jsx)(`button`,{className:`ai-rules-select-link`,onClick:u,children:`All`}),(0,X.jsx)(`span`,{className:`ai-rules-select-sep`,children:`·`}),(0,X.jsx)(`button`,{className:`ai-rules-select-link`,onClick:d,children:`None`})]}),(0,X.jsx)(`div`,{className:`ai-rules-grid`,children:$t.map(e=>{let n=t.has(e.id),i=r[e.id];return(0,X.jsxs)(`label`,{className:`ai-rules-card ${n?`ai-rules-card--selected`:``} ${a?`ai-rules-card--disabled`:``}`,children:[(0,X.jsx)(`input`,{type:`checkbox`,className:`ai-rules-checkbox`,checked:n,disabled:a,onChange:()=>l(e.id)}),(0,X.jsx)(`span`,{className:`ai-rules-card-icon`,children:e.icon}),(0,X.jsxs)(`div`,{className:`ai-rules-card-body`,children:[(0,X.jsx)(`span`,{className:`ai-rules-card-label`,children:e.label}),(0,X.jsx)(`code`,{className:`ai-rules-card-path`,children:e.path}),(0,X.jsx)(`span`,{className:`ai-rules-card-desc`,children:e.description})]}),(0,X.jsxs)(`div`,{className:`ai-rules-card-status`,children:[i?.status===`generating`&&(0,X.jsx)(`span`,{className:`ai-rules-status ai-rules-status--spinning`,children:`⏳`}),i?.status===`success`&&(0,X.jsx)(`span`,{className:`ai-rules-status ai-rules-status--ok`,title:i.path,children:`✓`}),i?.status===`error`&&(0,X.jsx)(`span`,{className:`ai-rules-status ai-rules-status--err`,title:i.error,children:`✗`})]})]},e.id)})}),v&&(0,X.jsxs)(`div`,{className:`ai-rules-summary`,children:[g>0&&(0,X.jsxs)(`span`,{className:`ai-rules-summary--ok`,children:[`✓ `,g,` file`,g===1?``:`s`,` written`]}),_>0&&(0,X.jsxs)(`span`,{className:`ai-rules-summary--err`,children:[`✗ `,_,` error`,_===1?``:`s`]})]}),(0,X.jsxs)(`div`,{className:`ai-rules-footer`,children:[(0,X.jsx)(`button`,{className:`export-btn export-btn--secondary`,onClick:e,disabled:a,children:v?`Close`:`Cancel`}),(0,X.jsx)(`button`,{className:`export-btn export-btn--primary ${a?`export-btn--loading`:``}`,onClick:p,disabled:a||t.size===0,children:a?(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`span`,{className:`btn-spinner btn-spinner--small`}),` Generating…`]}):`Generate ${t.size>0?t.size:``} File${t.size===1?``:`s`}`})]})]})})}function tn(e){let t=Math.floor(e/1e3);if(t<60)return`${t}s`;let n=Math.floor(t/60);if(n<60)return`${n}m`;let r=Math.floor(n/60);return r<24?`${r}h`:`${Math.floor(r/24)}d`}function nn({label:e,active:t,children:n}){let[r,i]=(0,A.useState)(!1),a=(0,A.useRef)(null);return(0,A.useEffect)(()=>{let e=e=>{a.current&&!a.current.contains(e.target)&&i(!1)};return document.addEventListener(`mousedown`,e,!0),()=>document.removeEventListener(`mousedown`,e,!0)},[]),(0,X.jsxs)(`div`,{className:`seg-dropdown`,ref:a,children:[(0,X.jsx)(`button`,{type:`button`,className:`seg-btn ${t||r?`seg-btn--active`:``}`,onClick:()=>i(!r),children:e}),r&&(0,X.jsx)(`div`,{className:`seg-dropdown-menu`,children:n})]})}function rn({nodeCount:e,edgeCount:t,visibleCount:n,activeTabLabel:r,graphData:i,analyzedAt:a,highRiskCount:o,onOpenRisks:s,theme:c,onSearch:l,onToggleTheme:u,graphRef:d}){let[f,p]=(0,A.useState)(``),[m,h]=(0,A.useState)(!1),[g,_]=(0,A.useState)(!1),[v,y]=(0,A.useState)(!1),b=(0,A.useRef)(null),x=(0,A.useRef)(null);(0,A.useEffect)(()=>(b.current&&clearTimeout(b.current),b.current=setTimeout(()=>l(f),250),()=>{b.current&&clearTimeout(b.current)}),[f,l]),(0,A.useEffect)(()=>{let e=e=>{(e.metaKey||e.ctrlKey)&&e.key.toLowerCase()===`k`?(e.preventDefault(),x.current?.focus(),x.current?.select()):e.key===`Escape`&&document.activeElement===x.current&&x.current?.blur()};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[]);let S=()=>{d.current?.toPng({scale:2}).then(e=>{e&&Ze(e,`${r.replace(/[^a-z0-9]/gi,`_`)}_graph.png`)})},C=()=>{i&&h(!0)},w=async()=>{if(window.confirm(`This will re-scan the entire project. Proceed?`)){y(!0);try{(await fetch(`/_laravel-brain/api/scan`,{method:`POST`})).ok?window.location.reload():alert(`Scan failed.`)}catch{alert(`Scan failed.`)}finally{y(!1)}}},[T,E]=(0,A.useState)(()=>Date.now());(0,A.useEffect)(()=>{let e=setInterval(()=>E(Date.now()),6e4);return()=>clearInterval(e)},[]);let D=(0,A.useMemo)(()=>a?`scanned ${tn(T-new Date(a).getTime())} ago`:null,[a,T]);return(0,X.jsxs)(X.Fragment,{children:[(0,X.jsxs)(`div`,{className:`toolbar`,children:[(0,X.jsxs)(`div`,{className:`toolbar-brand`,children:[(0,X.jsx)(`img`,{src:`/_laravel-brain/logo.png`,alt:`Laravel Brain`,className:`toolbar-logo-img`,width:28,height:28,decoding:`async`}),(0,X.jsxs)(`div`,{className:`toolbar-brand-text`,children:[(0,X.jsx)(`span`,{className:`toolbar-brand-name`,children:`Laravel Brain`}),D&&(0,X.jsx)(`span`,{className:`toolbar-brand-sub`,children:D})]})]}),(0,X.jsxs)(`div`,{className:`toolbar-center`,children:[(0,X.jsxs)(`div`,{className:`toolbar-search-wrapper`,children:[(0,X.jsxs)(`svg`,{className:`toolbar-search-icon`,width:`14`,height:`14`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,X.jsx)(`circle`,{cx:`11`,cy:`11`,r:`8`}),(0,X.jsx)(`line`,{x1:`21`,y1:`21`,x2:`16.65`,y2:`16.65`})]}),(0,X.jsx)(`input`,{ref:x,type:`search`,placeholder:`Search routes, nodes, files…`,className:`toolbar-search`,value:f,onChange:e=>p(e.target.value)}),(0,X.jsx)(`kbd`,{className:`toolbar-kbd`,children:`⌘K`})]}),(0,X.jsx)($,{content:`Routes flagged high or critical risk. Click to open the Risks list.`,children:(0,X.jsxs)(`button`,{type:`button`,className:`risk-pill ${o>0?`risk-pill--alert`:``}`,onClick:s,children:[(0,X.jsx)(`span`,{className:`risk-pill-dot`}),`High-risk`,(0,X.jsx)(`span`,{className:`risk-pill-count`,children:o})]})}),e>80&&(0,X.jsx)($,{content:`Large graph: dagre auto-switched to breadthfirst`,children:(0,X.jsx)(`span`,{className:`stat-chip stat-chip--warn`,children:`⚠ large`})}),(0,X.jsx)($,{content:`Nodes / edges in this graph (visible respects type filters).`,children:(0,X.jsxs)(`span`,{className:`stat-chip`,children:[n,`/`,e,` · `,t,`e`]})})]}),(0,X.jsxs)(`div`,{className:`toolbar-right`,children:[(0,X.jsx)($,{content:c===`dark`?`Switch to light mode`:`Switch to dark mode`,children:(0,X.jsx)(`button`,{type:`button`,onClick:u,className:`icon-btn`,children:c===`dark`?`☀`:`☾`})}),(0,X.jsxs)(nn,{label:`↧`,children:[(0,X.jsx)(`button`,{type:`button`,onClick:S,className:`seg-menu-btn`,children:`Download PNG`}),(0,X.jsx)(`button`,{type:`button`,onClick:C,className:`seg-menu-btn`,disabled:!i,children:`Copy Mermaid`}),(0,X.jsx)(`button`,{type:`button`,onClick:()=>_(!0),className:`seg-menu-btn`,children:`Generate AI Rules`})]}),(0,X.jsx)(`button`,{type:`button`,onClick:w,className:`rescan-btn ${v?`rescan-btn--loading`:``}`,disabled:v,"aria-busy":v,children:v?(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`span`,{className:`btn-spinner btn-spinner--small`,"aria-hidden":!0}),(0,X.jsx)(`span`,{children:`Scanning…`})]}):(0,X.jsxs)(X.Fragment,{children:[(0,X.jsxs)(`svg`,{width:`14`,height:`14`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,X.jsx)(`path`,{d:`M21 12a9 9 0 0 0-9-9 9.75 9.75 0 0 0-6.74 2.74L3 8`}),(0,X.jsx)(`path`,{d:`M3 3v5h5`}),(0,X.jsx)(`path`,{d:`M3 12a9 9 0 0 0 9 9 9.75 9.75 0 0 0 6.74-2.74L21 16`}),(0,X.jsx)(`path`,{d:`M16 16h5v5`})]}),(0,X.jsx)(`span`,{children:`Re-scan`})]})})]})]}),g&&(0,X.jsx)(en,{onClose:()=>_(!1)}),m&&i&&(0,X.jsx)(it,{mermaidCode:qe(i,r),filename:`${r.replace(/[^a-z0-9]/gi,`_`)}_graph.mmd`,title:`${r} — Full Lifecycle Graph`,onClose:()=>h(!1)})]})}var an={route:`Routes`,middleware:`Middleware`,controller:`Controllers`,livewire_component:`Livewire`,action:`Controller actions`,action_class:`Actions`,service:`Services`,validation_request:`Validation`,model:`Models`,event:`Events`,listener:`Listeners`,job:`Jobs`,command:`Commands`,channel:`Channels`,schedule:`Schedules`,view:`Views`,mail:`Mail`,notification:`Notifications`,enum:`Enums`,interface:`Interfaces`,trait:`Traits`,abstract_class:`Abstract`,service_provider:`Providers`,facade:`Facades`,ai_agent:`AI Agents`,ai_tool:`AI Tools`,filament_panel:`F. Panels`,filament_resource:`F. Resources`,filament_page:`F. Pages`,filament_page_method:`F. Methods`,filament_widget:`F. Widgets`,filament_relation_manager:`F. Relations`,macro:`Macros`,macro_group:`Macro receivers`},on=`route.middleware.controller.livewire_component.action.action_class.service.validation_request.model.event.listener.job.command.channel.schedule.view.mail.notification.enum.interface.trait.abstract_class.service_provider.facade.ai_agent.ai_tool.filament_panel.filament_resource.filament_page.filament_page_method.filament_widget.filament_relation_manager.macro.macro_group`.split(`.`),sn=[{type:`transaction`,label:`Transactions`,description:`the boundary drawn around work that runs in one transaction`},{type:`chain`,label:`Chains`,description:`the boundary and the arrows drawn around jobs that run one after another`},{type:`batch`,label:`Batches`,description:`the boundary drawn around jobs dispatched together, in no particular order`}];function cn({visibleTypes:e,counts:t,onToggle:n,onShowAll:r,onHideAll:i}){let a=on.filter(e=>(t[e]??0)>0),o=new Map(sn.map(e=>[e.type,e]));for(let e of sn)(t[e.type]??0)>0&&a.push(e.type);return(0,X.jsxs)(`div`,{className:`show-graph`,children:[(0,X.jsxs)(`div`,{className:`show-graph-header`,children:[(0,X.jsx)(`span`,{className:`show-graph-title`,children:`Show on graph`}),(0,X.jsxs)(`div`,{className:`show-graph-actions`,children:[(0,X.jsx)(`button`,{type:`button`,onClick:r,className:`show-graph-link`,children:`All`}),(0,X.jsx)(`span`,{className:`show-graph-sep`,children:`/`}),(0,X.jsx)(`button`,{type:`button`,onClick:i,className:`show-graph-link`,children:`None`})]})]}),(0,X.jsx)(`div`,{className:`show-graph-grid`,children:a.map(r=>{let i=t[r]??0,a=e.has(r),s=o.get(r),c=s?I[r]??`#94a3b8`:L[r]??`#94a3b8`,l=s?.label??an[r]??r;return(0,X.jsx)($,{content:s?`${a?`Hide`:`Show`} ${s.description}`:`${a?`Hide`:`Show`} ${l} nodes`,children:(0,X.jsxs)(`button`,{type:`button`,className:`show-graph-item ${a?``:`show-graph-item--off`}`,onClick:()=>n(r),children:[(0,X.jsx)(`span`,{className:`show-graph-dot`,style:{backgroundColor:c}}),(0,X.jsx)(`span`,{className:`show-graph-label`,children:l}),(0,X.jsx)(`span`,{className:`show-graph-count`,children:i})]})},r)})})]})}var ln={none:0,low:1,medium:2,high:3,critical:4},un=280,dn=480,fn=300,pn={GET:`#4ade80`,POST:`#60a5fa`,PUT:`#f59e0b`,PATCH:`#a78bfa`,DELETE:`#f87171`,OPTIONS:`#22d3ee`,QUERY:`#f472b6`},mn=[`GET`,`POST`,`PUT`,`PATCH`,`DELETE`,`OPTIONS`,`QUERY`];function hn(e){let[t,...n]=e.split(` `);return t in pn?{method:t,uri:n.join(` `)}:{method:null,uri:e}}function gn(e){let t=e.schedule,n=t?` ${t.cadence} ${t.timezone} ${t.modifiers.join(` `)}`:``;return`${e.label}${n}`.toLowerCase()}function _n(e){return e.riskLevel??`none`}function vn(e){let t=[];e.securityCount&&t.push(`${e.securityCount} security`),e.n1Count&&t.push(`${e.n1Count} N+1`);let n=(e.fatMethodCount??0)+(e.fatClassCount??0);return n&&t.push(`${n} fat`),t.length?t.join(` · `):`flagged for review`}function yn(e){if(!e)return`new`;let t=Date.now()-new Date(e).getTime(),n=Math.floor(t/6e4);if(n<60)return`${n}m ago`;let r=Math.floor(n/60);return r<24?`${r}h ago`:`${Math.floor(r/24)}d ago`}function bn({tab:e,isActive:t,isLoading:n,onSelect:r}){let{method:i,uri:a}=hn(e.label),o=i?pn[i]:`var(--faint)`,s=_n(e),c=s===`high`||s===`critical`?`var(--danger)`:e.issueCount?`var(--warn)`:null;return(0,X.jsx)($,{content:`Open lifecycle graph · ${e.nodeCount} nodes · ${e.edgeCount} edges`,children:(0,X.jsxs)(`button`,{className:`route-row ${t?`route-row--active`:``}`,type:`button`,onClick:()=>r(e),children:[(0,X.jsx)(`span`,{className:`route-row-method`,style:{color:o},children:i??`›`}),(0,X.jsx)(`span`,{className:`route-row-uri`,children:a}),c&&(0,X.jsx)(`span`,{className:`route-row-risk`,style:{"--rc":c},children:e.issueCount}),n&&(0,X.jsx)(`span`,{className:`route-row-loading`,children:`…`})]})})}var xn={command:`CMD`,job:`JOB`,call:`FN`},Sn={withoutOverlapping:`no overlap`,onOneServer:`one server`,runInBackground:`background`,evenInMaintenanceMode:`in maintenance`};function Cn(e){let t=e.split(`\\`).pop();return t&&t.length>0?t:e}function wn({tab:e,schedule:t,isActive:n,isLoading:r,onSelect:i}){let a=t.cadence||`no cadence stated`,o=t.modifiers.map(e=>Sn[e]??e);return(0,X.jsx)($,{content:`${t.target} · ${a}${t.timezone?` · ${t.timezone}`:``} · ${e.nodeCount} nodes`,children:(0,X.jsxs)(`button`,{className:`route-row route-row--stacked ${n?`route-row--active`:``}`,type:`button`,onClick:()=>i(e),children:[(0,X.jsx)(`span`,{className:`route-row-method`,style:{color:L[t.type===`job`?`job`:`command`]},children:xn[t.type]??`›`}),(0,X.jsxs)(`span`,{className:`schedule-row-body`,children:[(0,X.jsx)(`span`,{className:`schedule-row-scroll`,children:(0,X.jsx)(`span`,{className:`route-row-uri`,children:Cn(t.target)})}),(0,X.jsx)(`span`,{className:`schedule-cadence ${t.cadence?``:`schedule-cadence--unknown`}`,children:a}),(t.timezone||o.length>0)&&(0,X.jsxs)(`span`,{className:`schedule-row-scroll schedule-row-badges`,children:[t.timezone&&(0,X.jsx)(`span`,{className:`schedule-chip`,children:t.timezone}),o.map(e=>(0,X.jsx)(`span`,{className:`schedule-chip`,children:e},e))]})]}),r&&(0,X.jsx)(`span`,{className:`route-row-loading`,children:`…`})]})})}function Tn({tab:e,isActive:t,isLoading:n,onSelect:r}){return e.schedule?(0,X.jsx)(wn,{tab:e,schedule:e.schedule,isActive:t,isLoading:n,onSelect:r}):(0,X.jsx)(bn,{tab:e,isActive:t,isLoading:n,onSelect:r})}var En={shield:(0,X.jsx)(`path`,{d:`M12 22s8-4 8-10V5l-8-3-8 3v7c0 6 8 10 8 10z`}),lock:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`rect`,{x:`3`,y:`11`,width:`18`,height:`11`,rx:`2`}),(0,X.jsx)(`path`,{d:`M7 11V7a5 5 0 0 1 10 0v4`})]}),key:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:`7.5`,cy:`15.5`,r:`4.5`}),(0,X.jsx)(`path`,{d:`m10.7 12.3 8.3-8.3`}),(0,X.jsx)(`path`,{d:`m17 5 3 3`}),(0,X.jsx)(`path`,{d:`m15 7 3 3`})]}),user:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M20 21v-2a4 4 0 0 0-4-4H8a4 4 0 0 0-4 4v2`}),(0,X.jsx)(`circle`,{cx:`12`,cy:`7`,r:`4`})]}),users:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M16 21v-2a4 4 0 0 0-4-4H6a4 4 0 0 0-4 4v2`}),(0,X.jsx)(`circle`,{cx:`9`,cy:`7`,r:`4`}),(0,X.jsx)(`path`,{d:`M22 21v-2a4 4 0 0 0-3-3.87`}),(0,X.jsx)(`path`,{d:`M16 3.13a4 4 0 0 1 0 7.75`})]}),building:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`rect`,{x:`4`,y:`2`,width:`16`,height:`20`,rx:`2`}),(0,X.jsx)(`path`,{d:`M9 22v-4h6v4`}),(0,X.jsx)(`path`,{d:`M8 6h.01M16 6h.01M8 10h.01M16 10h.01M8 14h.01M16 14h.01`})]}),dashboard:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`rect`,{x:`3`,y:`3`,width:`7`,height:`9`}),(0,X.jsx)(`rect`,{x:`14`,y:`3`,width:`7`,height:`5`}),(0,X.jsx)(`rect`,{x:`14`,y:`12`,width:`7`,height:`9`}),(0,X.jsx)(`rect`,{x:`3`,y:`16`,width:`7`,height:`5`})]}),settings:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M12.22 2h-.44a2 2 0 0 0-2 2v.18a2 2 0 0 1-1 1.73l-.43.25a2 2 0 0 1-2 0l-.15-.08a2 2 0 0 0-2.73.73l-.22.38a2 2 0 0 0 .73 2.73l.15.1a2 2 0 0 1 1 1.72v.51a2 2 0 0 1-1 1.74l-.15.09a2 2 0 0 0-.73 2.73l.22.38a2 2 0 0 0 2.73.73l.15-.08a2 2 0 0 1 2 0l.43.25a2 2 0 0 1 1 1.73V20a2 2 0 0 0 2 2h.44a2 2 0 0 0 2-2v-.18a2 2 0 0 1 1-1.73l.43-.25a2 2 0 0 1 2 0l.15.08a2 2 0 0 0 2.73-.73l.22-.39a2 2 0 0 0-.73-2.73l-.15-.08a2 2 0 0 1-1-1.74v-.5a2 2 0 0 1 1-1.74l.15-.09a2 2 0 0 0 .73-2.73l-.22-.38a2 2 0 0 0-2.73-.73l-.15.08a2 2 0 0 1-2 0l-.43-.25a2 2 0 0 1-1-1.73V4a2 2 0 0 0-2-2z`}),(0,X.jsx)(`circle`,{cx:`12`,cy:`12`,r:`3`})]}),card:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`rect`,{x:`2`,y:`5`,width:`20`,height:`14`,rx:`2`}),(0,X.jsx)(`line`,{x1:`2`,y1:`10`,x2:`22`,y2:`10`})]}),cart:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:`9`,cy:`21`,r:`1`}),(0,X.jsx)(`circle`,{cx:`20`,cy:`21`,r:`1`}),(0,X.jsx)(`path`,{d:`M1 1h4l2.68 13.39a2 2 0 0 0 2 1.61h9.72a2 2 0 0 0 2-1.61L23 6H6`})]}),package:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M21 16V8a2 2 0 0 0-1-1.73l-7-4a2 2 0 0 0-2 0l-7 4A2 2 0 0 0 3 8v8a2 2 0 0 0 1 1.73l7 4a2 2 0 0 0 2 0l7-4A2 2 0 0 0 21 16z`}),(0,X.jsx)(`path`,{d:`M3.27 6.96 12 12.01l8.73-5.05`}),(0,X.jsx)(`path`,{d:`M12 22.08V12`})]}),file:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M14 2H6a2 2 0 0 0-2 2v16a2 2 0 0 0 2 2h12a2 2 0 0 0 2-2V8z`}),(0,X.jsx)(`polyline`,{points:`14 2 14 8 20 8`}),(0,X.jsx)(`line`,{x1:`16`,y1:`13`,x2:`8`,y2:`13`}),(0,X.jsx)(`line`,{x1:`16`,y1:`17`,x2:`8`,y2:`17`})]}),message:(0,X.jsx)(`path`,{d:`M21 15a2 2 0 0 1-2 2H7l-4 4V5a2 2 0 0 1 2-2h14a2 2 0 0 1 2 2z`}),bell:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M18 8A6 6 0 0 0 6 8c0 7-3 9-3 9h18s-3-2-3-9`}),(0,X.jsx)(`path`,{d:`M13.73 21a2 2 0 0 1-3.46 0`})]}),mail:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`rect`,{x:`2`,y:`4`,width:`20`,height:`16`,rx:`2`}),(0,X.jsx)(`path`,{d:`m22 7-10 5L2 7`})]}),search:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:`11`,cy:`11`,r:`8`}),(0,X.jsx)(`line`,{x1:`21`,y1:`21`,x2:`16.65`,y2:`16.65`})]}),folder:(0,X.jsx)(`path`,{d:`M22 19a2 2 0 0 1-2 2H4a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h5l2 3h9a2 2 0 0 1 2 2z`}),download:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4`}),(0,X.jsx)(`polyline`,{points:`7 10 12 15 17 10`}),(0,X.jsx)(`line`,{x1:`12`,y1:`15`,x2:`12`,y2:`3`})]}),upload:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M21 15v4a2 2 0 0 1-2 2H5a2 2 0 0 1-2-2v-4`}),(0,X.jsx)(`polyline`,{points:`17 8 12 3 7 8`}),(0,X.jsx)(`line`,{x1:`12`,y1:`3`,x2:`12`,y2:`15`})]}),chart:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`line`,{x1:`12`,y1:`20`,x2:`12`,y2:`10`}),(0,X.jsx)(`line`,{x1:`18`,y1:`20`,x2:`18`,y2:`4`}),(0,X.jsx)(`line`,{x1:`6`,y1:`20`,x2:`6`,y2:`16`})]}),list:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`line`,{x1:`8`,y1:`6`,x2:`21`,y2:`6`}),(0,X.jsx)(`line`,{x1:`8`,y1:`12`,x2:`21`,y2:`12`}),(0,X.jsx)(`line`,{x1:`8`,y1:`18`,x2:`21`,y2:`18`}),(0,X.jsx)(`line`,{x1:`3`,y1:`6`,x2:`3.01`,y2:`6`}),(0,X.jsx)(`line`,{x1:`3`,y1:`12`,x2:`3.01`,y2:`12`}),(0,X.jsx)(`line`,{x1:`3`,y1:`18`,x2:`3.01`,y2:`18`})]}),activity:(0,X.jsx)(`polyline`,{points:`22 12 18 12 15 21 9 3 6 12 2 12`}),link:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M10 13a5 5 0 0 0 7.54.54l3-3a5 5 0 0 0-7.07-7.07l-1.72 1.71`}),(0,X.jsx)(`path`,{d:`M14 11a5 5 0 0 0-7.54-.54l-3 3a5 5 0 0 0 7.07 7.07l1.71-1.71`})]}),zap:(0,X.jsx)(`polygon`,{points:`13 2 3 14 12 14 11 22 21 10 12 10 13 2`}),box:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`rect`,{x:`4`,y:`4`,width:`16`,height:`16`,rx:`2`}),(0,X.jsx)(`rect`,{x:`9`,y:`9`,width:`6`,height:`6`})]}),calendar:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`rect`,{x:`3`,y:`4`,width:`18`,height:`18`,rx:`2`}),(0,X.jsx)(`line`,{x1:`16`,y1:`2`,x2:`16`,y2:`6`}),(0,X.jsx)(`line`,{x1:`8`,y1:`2`,x2:`8`,y2:`6`}),(0,X.jsx)(`line`,{x1:`3`,y1:`10`,x2:`21`,y2:`10`})]}),pin:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M21 10c0 7-9 13-9 13s-9-6-9-13a9 9 0 0 1 18 0z`}),(0,X.jsx)(`circle`,{cx:`12`,cy:`10`,r:`3`})]}),book:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M2 3h6a4 4 0 0 1 4 4v14a3 3 0 0 0-3-3H2z`}),(0,X.jsx)(`path`,{d:`M22 3h-6a4 4 0 0 0-4 4v14a3 3 0 0 1 3-3h7z`})]}),info:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:`12`,cy:`12`,r:`10`}),(0,X.jsx)(`line`,{x1:`12`,y1:`16`,x2:`12`,y2:`12`}),(0,X.jsx)(`line`,{x1:`12`,y1:`8`,x2:`12.01`,y2:`8`})]}),beaker:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M9 3h6`}),(0,X.jsx)(`path`,{d:`M10 3v6l-5.5 9.5A2 2 0 0 0 6.2 21h11.6a2 2 0 0 0 1.7-3.5L14 9V3`})]}),tag:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M20.59 13.41 13.42 20.58a2 2 0 0 1-2.83 0L2 12V2h10l8.59 8.59a2 2 0 0 1 0 2.82z`}),(0,X.jsx)(`line`,{x1:`7`,y1:`7`,x2:`7.01`,y2:`7`})]}),broadcast:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`path`,{d:`M4 11a9 9 0 0 1 9 9`}),(0,X.jsx)(`path`,{d:`M4 4a16 16 0 0 1 16 16`}),(0,X.jsx)(`circle`,{cx:`5`,cy:`19`,r:`1`})]}),hash:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`line`,{x1:`4`,y1:`9`,x2:`20`,y2:`9`}),(0,X.jsx)(`line`,{x1:`4`,y1:`15`,x2:`20`,y2:`15`}),(0,X.jsx)(`line`,{x1:`10`,y1:`3`,x2:`8`,y2:`21`}),(0,X.jsx)(`line`,{x1:`16`,y1:`3`,x2:`14`,y2:`21`})]}),terminal:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`polyline`,{points:`4 17 10 11 4 5`}),(0,X.jsx)(`line`,{x1:`12`,y1:`19`,x2:`20`,y2:`19`})]}),clock:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:`12`,cy:`12`,r:`10`}),(0,X.jsx)(`polyline`,{points:`12 6 12 12 16 14`})]}),route:(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`circle`,{cx:`6`,cy:`19`,r:`3`}),(0,X.jsx)(`circle`,{cx:`18`,cy:`5`,r:`3`}),(0,X.jsx)(`path`,{d:`M9 19h6a4 4 0 0 0 4-4V9`})]})};function Dn({name:e}){return(0,X.jsx)(`svg`,{className:`tree-group-icon`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`,strokeLinejoin:`round`,"aria-hidden":`true`,children:En[e]})}var On=[[/^(auth|login|register|signin|signup|signout|logout|verify)/i,`lock`],[/^(password|forgot|reset|recover)/i,`key`],[/^(oauth|sso|saml|token|jwt|sanctum|passport)/i,`key`],[/^(admin|backend|manage|mgmt|cp|role|permission|acl|guard|policy|gate|abilit|security|firewall|protect|shield)/i,`shield`],[/^(team|organization|org|company|tenant|workspace)/i,`building`],[/^(user|account|profile|member|people|person)/i,`user`],[/^(group|staff|contributor|follower)/i,`users`],[/^(dashboard|home|overview|index|main|panel)/i,`dashboard`],[/^(setting|config|preference|option|env)/i,`settings`],[/^(billing|payment|invoice|subscription|plan|pricing|wallet|transaction|refund)/i,`card`],[/^(checkout|cart|basket|bag)/i,`cart`],[/^(order|purchase|fulfil|shipping|delivery|product|catalog|catalogue|item|shop|store|inventory|stock)/i,`package`],[/^(blog|post|article|news|content|page|cms)/i,`file`],[/^(message|chat|conversation|inbox|thread|dm|comment|review|rating|feedback|reply)/i,`message`],[/^(notification|notif|alert|push)/i,`bell`],[/^(mail|email|newsletter|campaign)/i,`mail`],[/^(search|explore|discover|find|query|filter)/i,`search`],[/^(upload|file|files|media|image|photo|asset|document|docs?|attachment|storage)/i,`folder`],[/^(download|export|backup|dump)/i,`download`],[/^(import|sync|migrate)/i,`upload`],[/^(report|analytic|stat|statistic|metric|insight|chart|kpi)/i,`chart`],[/^(log|logs|audit|activity|history|track|trace)/i,`list`],[/^(health|status|ping|up|ready|live|heartbeat|probe|monitor)/i,`activity`],[/^(webhook|callback|hook|integration|connect|link)/i,`link`],[/^(cache|redis|optimize)/i,`zap`],[/^(queue|job|jobs|worker|batch|cron)/i,`box`],[/^(calendar|event|booking|appointment|reservation|slot)/i,`calendar`],[/^(map|location|geo|address|place|region|country)/i,`pin`],[/^(project|board|workflow|pipeline)/i,`folder`],[/^(help|support|faq|guide|tutorial|kb|knowledge|wiki)/i,`book`],[/^(contact|enquir|inquir|lead)/i,`user`],[/^(about|info|legal|privacy|terms|policy)/i,`info`],[/^(test|tests|debug|dev|sandbox|playground|demo|example)/i,`beaker`],[/^(tag|tags|category|categories|topic|label)/i,`tag`],[/^(feed|rss|atom|socket|ws|realtime|broadcast|stream)/i,`broadcast`],[/^(api|graphql|ql|rest|rpc)$/i,`hash`],[/^v?\d+(\.\d+)*$/i,`hash`]],kn={"Console Commands":`terminal`,"Broadcast Channels":`broadcast`,Schedules:`clock`,"Model ERD":`box`,"Event Choreography":`zap`,"AI Agents":`zap`,Other:`route`};function An(e,t){if(t)return e.startsWith(`Filament`)?`box`:kn[e]??`route`;for(let[t,n]of On)if(t.test(e))return n;return`route`}function jn(e){if(e.category===`Command`)return`Console Commands`;if(e.category===`Channel`)return`Broadcast Channels`;if(e.category===`Schedule`)return`Schedules`;if(e.category===`ERD`)return`Model ERD`;if(e.category===`Events`)return`Event Choreography`;if(e.category===`AI`)return`AI Agents`;if(e.category===`Filament`){let t=e.panelId??``;return t?`Filament · ${t.charAt(0).toUpperCase()}${t.slice(1)} Panel`:`Filament`}return`Other`}function Mn(e){e.children.sort((e,t)=>e.name.localeCompare(t.name)),e.leaves.sort((e,t)=>e.label.localeCompare(t.label)),e.children.forEach(Mn)}function Nn(e){let t=e.label.split(` `)[0];return t in pn?e.label.slice(t.length).trim().split(`/`).filter(Boolean):null}function Pn(e){let t={name:``,path:``,isCategory:!1,children:[],leaves:[]},n=(e,t,n)=>{let r=e.children.find(e=>e.name===t);return r||(r={name:t,path:e.path?`${e.path}/${t}`:t,isCategory:n,children:[],leaves:[]},e.children.push(r)),r},r=new Set;for(let t of e){let e=Nn(t);if(!e)continue;let n=e.slice(0,-1);for(let e=1;e<=n.length;e++)r.add(n.slice(0,e).join(`/`))}for(let i of e){let e=Nn(i);if(!e){n(t,jn(i),!0).leaves.push(i);continue}let a=e.join(`/`),o=a!==``&&r.has(a)?e:e.slice(0,-1),s=t;for(let e of o)s=n(s,e,!1);s.leaves.push(i)}return Mn(t),t}function Fn(e){return e.leaves.length+e.children.reduce((e,t)=>e+Fn(t),0)}function In({node:e,forceOpen:t,expanded:n,onToggle:r,activeId:i,loadingId:a,onSelect:o,level:s=0}){let c=t||n.has(e.path),l=e.isCategory?e.name:`/${e.name}`;return(0,X.jsxs)(`div`,{className:`tree-group`,children:[(0,X.jsxs)(`button`,{type:`button`,className:`tree-group-header`,onClick:()=>r(e.path),children:[(0,X.jsx)(`span`,{className:`tree-group-chevron`,children:c?`▾`:`▸`}),s===0&&(0,X.jsx)(Dn,{name:An(e.name,e.isCategory)}),(0,X.jsx)(`span`,{className:`tree-group-name`,children:l}),(0,X.jsx)(`span`,{className:`tree-group-count`,children:Fn(e)})]}),c&&(0,X.jsxs)(`div`,{className:`tree-group-body`,children:[e.children.map(e=>(0,X.jsx)(In,{node:e,forceOpen:t,expanded:n,onToggle:r,activeId:i,loadingId:a,onSelect:o,level:s+1},e.path)),e.leaves.map(e=>(0,X.jsx)(Tn,{tab:e,isActive:e.id===i,isLoading:e.id===a,onSelect:o},e.id))]})]})}function Ln({tab:e,isActive:t,onSelect:n,timestamp:r}){let{method:i,uri:a}=hn(e.label),o=_n(e),s=o===`critical`?`critical`:o===`high`?`high`:o===`medium`?`medium`:`low`,c=B[s]??B.medium;return(0,X.jsxs)(`button`,{type:`button`,className:`flag-card ${t?`flag-card--active`:``}`,onClick:()=>n(e),children:[(0,X.jsxs)(`div`,{className:`flag-card-top`,children:[r?(0,X.jsx)(`span`,{className:`flag-card-time`,children:r}):(0,X.jsx)(`span`,{className:`flag-card-sev`,style:{"--sc":c},children:(se[s]??s).toUpperCase()}),i&&(0,X.jsx)(`span`,{className:`flag-card-method`,style:{color:pn[i]},children:i})]}),(0,X.jsx)(`div`,{className:`flag-card-path`,children:a}),(0,X.jsx)(`div`,{className:`flag-card-desc`,children:vn(e)})]})}function Rn({tabs:e,activeId:t,loadingId:n,onSelect:r,mode:i,onModeChange:a,previousAnalyzedAt:o,visibleTypes:s,counts:c,onToggle:l,onShowAll:u,onHideAll:d}){let[f,p]=(0,A.useState)(fn),[m,h]=(0,A.useState)(``),[g,_]=(0,A.useState)(new Set(mn)),[v,y]=(0,A.useState)(new Set),b=(0,A.useCallback)(e=>{_(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n})},[]),x=(0,A.useCallback)(e=>y(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n}),[]),S=(0,A.useRef)(!1),C=(0,A.useRef)(0),w=(0,A.useRef)(fn),T=(0,A.useCallback)(e=>{e.preventDefault(),S.current=!0,C.current=e.clientX,w.current=f;let t=e=>{if(!S.current)return;let t=e.clientX-C.current;p(Math.min(dn,Math.max(un,w.current+t)))},n=()=>{S.current=!1,window.removeEventListener(`mousemove`,t),window.removeEventListener(`mouseup`,n),document.body.style.cursor=``,document.body.style.userSelect=``};document.body.style.cursor=`col-resize`,document.body.style.userSelect=`none`,window.addEventListener(`mousemove`,t),window.addEventListener(`mouseup`,n)},[f]),E=m.trim().toLowerCase(),D=(0,A.useMemo)(()=>{let t=mn.every(e=>g.has(e));return e.filter(e=>{if(E&&!gn(e).includes(E))return!1;if(!t){let t=e.label.split(` `)[0];if(t in pn&&!g.has(t))return!1}return!0})},[e,E,g]),O=E.length>0||!mn.every(e=>g.has(e)),k=(0,A.useMemo)(()=>Pn(D),[D]),j=(0,A.useMemo)(()=>D.filter(e=>_n(e)!==`none`).sort((e,t)=>(ln[_n(t)]??0)-(ln[_n(e)]??0)),[D]),M=(0,A.useMemo)(()=>D.filter(e=>e.changeStatus===`new`||e.changeStatus===`changed`),[D]),N=[{id:`routes`,label:`Routes`,count:D.length},{id:`risks`,label:`Risks`,count:j.length},{id:`recent`,label:`Recent`,count:M.length}];return(0,X.jsxs)(`div`,{className:`left-sidebar-resizable`,style:{width:f,"--left-sidebar-width":`${f}px`},children:[(0,X.jsxs)(`div`,{className:`left-sidebar`,children:[(0,X.jsxs)(`div`,{className:`left-search`,children:[(0,X.jsx)(`input`,{className:`left-search-input`,type:`text`,placeholder:`Search routes…`,value:m,onChange:e=>h(e.target.value)}),m&&(0,X.jsx)(`button`,{type:`button`,className:`left-search-clear`,onClick:()=>h(``),children:`×`})]}),(0,X.jsx)(`div`,{className:`left-method-chips`,children:mn.map(e=>(0,X.jsx)(`button`,{type:`button`,className:`method-chip ${g.has(e)?`method-chip--on`:``}`,style:{"--mc":pn[e]},onClick:()=>b(e),children:e},e))}),(0,X.jsx)(`div`,{className:`mode-tabs`,children:N.map(e=>(0,X.jsxs)(`button`,{type:`button`,className:`mode-tab ${i===e.id?`mode-tab--active`:``}`,onClick:()=>a(e.id),children:[e.label,(0,X.jsx)(`span`,{className:`mode-tab-count ${e.id===`risks`&&i===`risks`&&e.count>0?`mode-tab-count--alert`:``}`,children:e.count})]},e.id))}),(0,X.jsxs)(`div`,{className:`left-content`,children:[i===`routes`&&(0,X.jsxs)(`div`,{className:`route-tree`,children:[k.children.length===0&&k.leaves.length===0&&(0,X.jsx)(`div`,{className:`left-empty`,children:`No routes match.`}),k.children.map(e=>(0,X.jsx)(In,{node:e,forceOpen:E.length>0,expanded:v,onToggle:x,activeId:t,loadingId:n,onSelect:r},e.path)),k.leaves.map(e=>(0,X.jsx)(Tn,{tab:e,isActive:e.id===t,isLoading:e.id===n,onSelect:r},e.id))]}),i===`risks`&&(0,X.jsxs)(`div`,{className:`flag-list`,children:[j.length===0&&(0,X.jsx)(`div`,{className:`left-empty`,children:O?`No flagged routes match the filter.`:`No flagged routes. ✓`}),j.map(e=>(0,X.jsx)(Ln,{tab:e,isActive:e.id===t,onSelect:r},e.id))]}),i===`recent`&&(0,X.jsxs)(`div`,{className:`flag-list`,children:[M.length===0&&(0,X.jsx)(`div`,{className:`left-empty`,children:O?`Nothing matching the filter changed since the previous scan.`:`Nothing changed since the previous scan.`}),M.map(e=>(0,X.jsx)(Ln,{tab:e,isActive:e.id===t,onSelect:r,timestamp:`${e.changeStatus===`new`?`new`:`changed`} · ${yn(o)}`},e.id))]})]}),(0,X.jsx)(`div`,{className:`left-footer`,children:(0,X.jsx)(cn,{visibleTypes:s,counts:c,onToggle:l,onShowAll:u,onHideAll:d})})]}),(0,X.jsx)($,{content:`Drag to resize`,children:(0,X.jsx)(`div`,{className:`left-sidebar-drag-handle`,onMouseDown:T})})]})}var zn=[...`route.middleware.controller.livewire_component.action.action_class.service.validation_request.model.event.listener.job.command.channel.schedule.view.mail.notification.enum.interface.trait.abstract_class.service_provider.ai_agent.ai_tool.filament_panel.filament_resource.filament_page.filament_page_method.filament_widget.filament_relation_manager.macro.macro_group`.split(`.`),`transaction`,`chain`,`batch`];function Bn(){let{theme:e,toggle:t}=ee(),{manifest:n,loading:r,error:i}=M(),{state:a,elements:o,load:s}=P(),c=F(o),[l,u]=(0,A.useState)(null),[d,f]=(0,A.useState)(null),[p,m]=(0,A.useState)(`dagre`),[h,g]=(0,A.useState)(null),[_,v]=(0,A.useState)(`routes`),[y,b]=(0,A.useState)(``),[x,S]=(0,A.useState)(!1),[C,w]=(0,A.useState)(new Set(zn)),[T,E]=(0,A.useState)(`TB`),[D,O]=(0,A.useState)(null),[k,j]=(0,A.useState)(0),N=(0,A.useRef)(null),te=(0,A.useCallback)(e=>{if(l?.id===e.id)return;let t=new URL(window.location.href);t.searchParams.get(`tab`)!==e.id&&(t.searchParams.set(`tab`,e.id),window.history.pushState({tabId:e.id},``,t.toString())),u(e),b(``),S(!0),s(e.file)},[l,s]),[I,ne]=(0,A.useState)(n);if(n!==I&&(ne(n),n&&!l)){let e=new URLSearchParams(window.location.search).get(`tab`),t=n.tabs.find(t=>t.id===e);t&&te(t)}let[L,re]=(0,A.useState)(a.data);if(a.data!==L)if(re(a.data),a.data)if(w(new Set(zn)),x){S(!1);let e=a.data.nodes.find(e=>e.type===`route`);g(e?e.id:null)}else g(null);else g(null);(0,A.useEffect)(()=>{let e=()=>{if(!n)return;let e=new URLSearchParams(window.location.search).get(`tab`),t=n.tabs.find(t=>t.id===e);t&&(u(t),s(t.file))};return window.addEventListener(`popstate`,e),()=>window.removeEventListener(`popstate`,e)},[n,s]);let ie=(0,A.useCallback)(e=>{g(e)},[]),[R,ae]=(0,A.useState)(a.loading);a.loading!==R&&(ae(a.loading),a.loading||f(null));let z=(0,A.useMemo)(()=>n?.tabs??[],[n]),oe=(0,A.useMemo)(()=>z.filter(e=>e.riskLevel===`high`||e.riskLevel===`critical`).length,[z]),B=(0,A.useMemo)(()=>{if(!a.data)return{};let e=a.data.nodes.reduce((e,t)=>(e[t.type]=(e[t.type]??0)+1,e),{}),t={};for(let e of a.data.nodes)for(let n of he(e)){let e=n.kind===`rollback`?`transaction`:n.kind;t[e]=(t[e]??new Set).add(n.id)}for(let[n,r]of Object.entries(t))e[n]=r.size;return e},[a.data]),se=(0,A.useMemo)(()=>a.data?a.data.nodes.filter(e=>C.has(e.type)).length:0,[a.data,C]),ce=(0,A.useCallback)(e=>{w(t=>{let n=new Set(t);return n.has(e)?n.delete(e):n.add(e),n})},[]),le=(0,A.useCallback)(()=>w(new Set(zn)),[]),ue=(0,A.useCallback)(()=>w(new Set),[]),[de,fe]=(0,A.useState)(!1),[pe,V]=(0,A.useState)(!1),[H,U]=(0,A.useState)(`all`),[W,G]=(0,A.useState)(!1),[K,me]=(0,A.useState)(!1);return r?(0,X.jsxs)(`div`,{className:`loading-screen`,children:[(0,X.jsx)(`div`,{className:`loading-spinner`}),(0,X.jsx)(`p`,{children:`Loading project graph...`})]}):i||!n?(0,X.jsx)(`div`,{className:`error-screen welcome-screen`,children:(0,X.jsxs)(`div`,{className:`welcome-card`,children:[(0,X.jsx)(`div`,{className:`welcome-icon`,children:(0,X.jsx)(`img`,{src:`/_laravel-brain/logo.png`,alt:`Laravel Brain`})}),(0,X.jsx)(`h2`,{children:`Welcome to Laravel Brain`}),(0,X.jsx)(`p`,{children:`No project analysis found. To begin exploring your code architecture, please run an initial scan.`}),i&&i!==`HTTP 404`&&(0,X.jsx)(`div`,{className:`error-details`,children:(0,X.jsxs)(`small`,{children:[`Error: `,i]})}),(0,X.jsx)(`button`,{className:`scan-btn ${de?`scan-btn--loading`:``}`,onClick:async()=>{if(window.confirm(`This will scan the entire project. Proceed?`)){fe(!0);try{(await fetch(`/_laravel-brain/api/scan`,{method:`POST`})).ok?window.location.reload():alert(`Scan failed.`)}catch{alert(`Scan failed.`)}finally{fe(!1)}}},disabled:de,children:de?(0,X.jsxs)(X.Fragment,{children:[(0,X.jsx)(`div`,{className:`btn-spinner`}),`Analyzing Project...`]}):`🚀 Start Initial Scan`}),(0,X.jsxs)(`div`,{className:`welcome-hint`,children:[`Alternatively, run `,(0,X.jsx)(`code`,{children:`php artisan brain:scan`}),` in your terminal.`]})]})}):(0,X.jsxs)(`div`,{className:`app`,children:[(0,X.jsx)(rn,{nodeCount:a.data?.meta.nodeCount??l?.nodeCount??0,edgeCount:a.data?.meta.edgeCount??l?.edgeCount??0,visibleCount:se,activeTabLabel:l?.label??`graph`,graphData:a.data??null,analyzedAt:n.analyzedAt,highRiskCount:oe,onOpenRisks:()=>v(`risks`),theme:e,onSearch:b,onToggleTheme:t,graphRef:N}),(0,X.jsxs)(`div`,{className:`main`,children:[(0,X.jsx)(Rn,{tabs:z,activeId:l?.id??null,loadingId:d,onSelect:te,mode:_,onModeChange:v,previousAnalyzedAt:n.previousAnalyzedAt,visibleTypes:C,counts:B,onToggle:ce,onShowAll:le,onHideAll:ue,graphData:a.data??null,complexityFilter:H,onComplexityFilterChange:U,onNodeSelect:ie,selectedId:h}),(0,X.jsxs)(`div`,{className:`graph-container`,children:[a.loading&&(0,X.jsxs)(`div`,{className:`graph-loading-overlay`,children:[(0,X.jsx)(`div`,{className:`loading-spinner`}),(0,X.jsxs)(`p`,{children:[`Loading `,l?.label,`…`]})]}),a.error&&(0,X.jsx)(`div`,{className:`graph-loading-overlay`,children:(0,X.jsxs)(`p`,{style:{color:`#F44336`},children:[`Error: `,a.error]})}),!l&&!a.loading&&(0,X.jsx)($,{content:`Pick a route or command in the left sidebar to load its dependency graph.`,children:(0,X.jsxs)(`div`,{className:`graph-placeholder`,children:[(0,X.jsx)(`div`,{className:`placeholder-icon`,children:(0,X.jsx)(`svg`,{width:`64`,height:`64`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`1.5`,strokeLinecap:`round`,strokeLinejoin:`round`,children:(0,X.jsx)(`polyline`,{points:`22 12 18 12 15 21 9 3 6 12 2 12`})})}),(0,X.jsx)(`h3`,{children:`Select a route to explore`}),(0,X.jsx)(`p`,{children:`Expand the files in the sidebar and choose a route or command to visualize its execution lifecycle and dependencies.`})]})}),!a.loading&&l&&c.length===0&&!a.error&&(0,X.jsx)($,{content:`This endpoint produced no analyzable nodes. It may be a closure, a redirect-only route, or outside the scanner’s rules.`,children:(0,X.jsxs)(`div`,{className:`graph-placeholder`,children:[(0,X.jsx)(`div`,{className:`placeholder-icon`,children:(0,X.jsxs)(`svg`,{width:`64`,height:`64`,viewBox:`0 0 24 24`,fill:`none`,stroke:`currentColor`,strokeWidth:`1.5`,strokeLinecap:`round`,strokeLinejoin:`round`,children:[(0,X.jsx)(`circle`,{cx:`12`,cy:`12`,r:`10`}),(0,X.jsx)(`line`,{x1:`12`,y1:`8`,x2:`12`,y2:`12`}),(0,X.jsx)(`line`,{x1:`12`,y1:`16`,x2:`12.01`,y2:`16`})]})}),(0,X.jsx)(`h3`,{children:`Empty Graph`}),(0,X.jsx)(`p`,{children:`No nodes or edges found for this route.`})]})}),!a.loading&&c.length>0&&(0,X.jsx)(Ve,{elements:c,layout:p,searchQuery:y,rankDir:T,visibleTypes:C,theme:e,onNodeSelect:ie,graphRef:N,stressTestNodeId:D,stressRunKey:k,complexityOverlay:pe,securityOverlay:W,compact:K,onLayoutChange:m,onRankDirChange:E,onToggleComplexityOverlay:()=>V(e=>!e),onToggleSecurityOverlay:()=>G(e=>!e),onToggleCompact:()=>me(e=>!e)},l?.id)]}),h&&(0,X.jsx)(Qt,{selectedId:h,graphData:a.data,theme:e,onClose:()=>g(null),onStressChange:e=>{O(e),e!==null&&j(e=>e+1)}})]})]})}(0,j.createRoot)(document.getElementById(`root`)).render((0,X.jsx)(A.StrictMode,{children:(0,X.jsx)(Bn,{})}));
